Futhark.AD.Derivatives

Description

Partial derivatives of scalar Futhark operations and built-in functions.

Documentation

pdBuiltin :: Name -> [PrimExp VName] -> Maybe [PrimExp VName]

pdBuiltin f args i computes the partial derivative of f applied to args with respect to each of its arguments. Returns Nothing if no such derivative is known.

pdBinOp :: BinOp -> PrimExp VName -> PrimExp VName -> (PrimExp VName, PrimExp VName)

pdUnOp :: UnOp -> PrimExp VName -> PrimExp VName