Futhark.AD.Rev.Monad

Documentation

data ADM a

data RState

Constructors

RState

Fields

stateAdjs :: Map VName Adj

stateLoopTape :: Substitutions

stateSubsts :: Substitutions

stateNameSource :: VNameSource

runADM :: MonadFreshNames m => ADM a -> m a

data Adj

The adjoint of a variable.

Constructors

AdjSparse Sparse

AdjVal SubExp

AdjZero Shape PrimType

data InBounds

Whether Sparse should check bounds or assume they are correct. The latter results in simpler code.

Constructors

CheckBounds (Maybe SubExp)
If a SubExp is provided, it references a boolean that is true when in-bounds.

AssumeBounds

OutOfBounds
Dynamically these will always fail, so don't bother generating code for the update. This is only needed to ensure a consistent representation of sparse Jacobians.

data Sparse

A symbolic representation of an array that is all zeroes, except at one index.

Constructors

Sparse

Fields

sparseShape :: Shape

The shape of the array.

sparseType :: PrimType

Element type of the array.

sparseIdxVals :: [(InBounds, SubExp, SubExp)]

Locations and values of nonzero values. Indexes may be negative, in which case the value is ignored (unless AssumeBounds is used).

adjFromParam :: Param t -> Adj

adjFromVar :: VName -> Adj

lookupAdj :: VName -> ADM Adj

lookupAdjVal :: VName -> ADM VName

adjVal :: Adj -> ADM VName

updateAdj :: VName -> VName -> ADM ()

updateSubExpAdj :: SubExp -> VName -> ADM ()

updateAdjSlice :: Slice SubExp -> VName -> VName -> ADM ()

updateAdjIndex :: VName -> (InBounds, SubExp) -> SubExp -> ADM ()

setAdj :: VName -> Adj -> ADM ()

insAdj :: VName -> VName -> ADM ()

insSubExpAdj :: SubExp -> VName -> ADM ()

adjsReps :: [Adj] -> ([SubExp], [SubExp] -> [Adj])

Conveniently convert a list of Adjs to their representation, as well as produce a function for converting back.

copyConsumedArrsInStm :: Stm -> ADM (Substitutions, Stms SOACS)

Create copies of all arrays consumed in the given statement, and return statements which include copies of the consumed arrays.

See Note [Consumption].

copyConsumedArrsInBody :: [VName] -> Body -> ADM Substitutions

addSubstitution :: VName -> VName -> ADM ()

returnSweepCode :: ADM a -> ADM a

adjVName :: VName -> ADM VName

subAD :: ADM a -> ADM a

noAdjsFor :: Names -> ADM a -> ADM a

subSubsts :: ADM a -> ADM a

isActive :: VName -> ADM Bool

Is this primal variable active in the AD sense? FIXME: this is (obviously) much too conservative.

tabNest :: Int -> [VName] -> ([VName] -> [VName] -> ADM [VName]) -> ADM [VName]

oneExp :: Type -> ExpT rep

zeroExp :: Type -> ExpT rep

unitAdjOfType :: Type -> ADM Adj

addLambda :: Type -> ADM Lambda

Construct a lambda for adding two values of the given type.

data VjpOps

Constructors

VjpOps

Fields

vjpLambda :: [Adj] -> [VName] -> Lambda -> ADM Lambda

vjpStm :: Stm -> ADM () -> ADM ()

setLoopTape :: VName -> VName -> ADM ()

setLoopTape v vs establishes vs as the name of the array where values of loop parameter v from the forward pass are stored.

lookupLoopTape :: VName -> ADM (Maybe VName)

Look-up the name of the array where v is stored.

substLoopTape :: VName -> VName -> ADM ()

substLoopTape v v' substitutes the key v for v'. That is, if v |-> vs then after the substitution v' |-> vs (and v points to nothing).

renameLoopTape :: Substitutions -> ADM ()

Renames the keys of the loop tape. Useful for fixing the the names in the loop tape after a loop rename.