Futhark.Analysis.CallGraph

Description

This module exports functionality for generating a call graph of an Futhark program.

Documentation

data CallGraph

The call graph is a mapping from a function name, i.e., the caller, to a record of the names of functions called *directly* (not transitively!) by the function.

We keep track separately of the functions called by constants.

buildCallGraph :: Prog SOACS -> CallGraph

buildCallGraph prog build the program's call graph.

isFunInCallGraph :: Name -> CallGraph -> Bool

Is the given function known to the call graph?

calls :: Name -> Name -> CallGraph -> Bool

Does the first function call the second?

calledByConsts :: Name -> CallGraph -> Bool

Is the function called in any of the constants?

allCalledBy :: Name -> CallGraph -> Set Name

All functions called by this function.

numOccurences :: CallGraph -> Map Name Int

Produce a mapping of the number of occurences in the call graph of each function. Only counts functions that are called at least once.

findNoninlined :: Prog SOACS -> Set Name

The set of all functions that are called noinline somewhere, or have a noinline attribute on their definition.