Futhark.Analysis.LastUse

Description

Provides last-use analysis for Futhark programs.

Documentation

type LastUseMap = Map VName Names

LastUseMap tells which names were last used in a given statement. Statements are uniquely identified by the VName of the first value parameter in the statement pattern. Names is the set of names last used.

analyseGPUMem :: Prog GPUMem -> (LastUseMap, Used)

analyseSeqMem :: Prog SeqMem -> (LastUseMap, Used)