Futhark.Analysis.Metrics

Description

Abstract Syntax Tree metrics. This is used in the futhark test program, for the structure stanzas.

Documentation

newtype AstMetrics

AST metrics are simply a collection from identifiable node names to the number of times that node appears.

Constructors

AstMetrics (Map Text Int)

progMetrics :: OpMetrics (Op rep) => Prog rep -> AstMetrics

Compute the metrics for a program.

Extensibility

class OpMetrics op where

Compute the metrics for some operation.

Methods

opMetrics :: op -> MetricsM ()

seen :: Text -> MetricsM ()

Add this node to the current tally.

inside :: Text -> MetricsM () -> MetricsM ()

Enclose a metrics counting operation. Most importantly, this prefixes the name of the context to all the metrics computed in the enclosed operation.

data MetricsM a

This monad is used for computing metrics. It internally keeps track of what we've seen so far. Use seen to add more stuff.

stmMetrics :: OpMetrics (Op rep) => Stm rep -> MetricsM ()

Compute metrics for this statement.

lambdaMetrics :: OpMetrics (Op rep) => Lambda rep -> MetricsM ()

Compute metrics for this lambda.