Futhark.Analysis.Metrics.Type

Description

The data type definition for Futhark.Analysis.Metrics, factored out to simplify the module import hierarchies when working on the test modules.

Documentation

newtype AstMetrics

AST metrics are simply a collection from identifiable node names to the number of times that node appears.

Constructors

AstMetrics (Map Text Int)