Futhark.Analysis.PrimExp

Description

A primitive expression is an expression where the non-leaves are primitive operators. Our representation does not guarantee that the expression is type-correct.

Documentation

data PrimExp v

A primitive expression parametrised over the representation of free variables. Note that the Functor, Traversable, and Num instances perform automatic (but simple) constant folding.

Note also that the Num instance assumes OverflowUndef semantics!

Constructors

LeafExp v PrimType

ValueExp PrimValue

BinOpExp BinOp (PrimExp v) (PrimExp v)

CmpOpExp CmpOp (PrimExp v) (PrimExp v)

UnOpExp UnOp (PrimExp v)

ConvOpExp ConvOp (PrimExp v)

FunExp String [PrimExp v] PrimType

newtype TPrimExp t v

A PrimExp tagged with a phantom type used to provide type-safe construction. Does not guarantee that the underlying expression is actually type correct.

Constructors

TPrimExp

Fields

untyped :: PrimExp v

isInt8 :: PrimExp v -> TPrimExp Int8 v

This expression is of type Int8.

isInt16 :: PrimExp v -> TPrimExp Int16 v

This expression is of type Int16.

isInt32 :: PrimExp v -> TPrimExp Int32 v

This expression is of type Int32.

isInt64 :: PrimExp v -> TPrimExp Int64 v

This expression is of type Int64.

isBool :: PrimExp v -> TPrimExp Bool v

This is a boolean expression.

isF16 :: PrimExp v -> TPrimExp Half v

This expression is of type Half.

isF32 :: PrimExp v -> TPrimExp Float v

This expression is of type Float.

isF64 :: PrimExp v -> TPrimExp Double v

This expression is of type Double.

evalPrimExp :: (Pretty v, MonadFail m) => (v -> m PrimValue) -> PrimExp v -> m PrimValue

Evaluate a PrimExp in the given monad. Invokes fail on type errors.

primExpType :: PrimExp v -> PrimType

The type of values returned by a PrimExp. This function returning does not imply that the PrimExp is type-correct.

primExpSizeAtLeast :: Int -> PrimExp v -> Bool

True if the PrimExp has at least this many nodes. This can be much more efficient than comparing with length for large PrimExps, as this function is lazy.

coerceIntPrimExp :: IntType -> PrimExp v -> PrimExp v

If the given PrimExp is a constant of the wrong integer type, coerce it to the given integer type. This is a workaround for an issue in the Num instance.

leafExpTypes :: Ord a => PrimExp a -> Set (a, PrimType)

Produce a mapping from the leaves of the PrimExp to their designated types.

true :: TPrimExp Bool v

Boolean-valued PrimExps.

false :: TPrimExp Bool v

Boolean-valued PrimExps.

constFoldPrimExp :: PrimExp v -> PrimExp v

Perform quick and dirty constant folding on the top level of a PrimExp. This is necessary because we want to consider e.g. equality modulo constant folding.

Construction

module Futhark.IR.Primitive

class NumExp t where

The class of numeric types that can be used for constructing TPrimExps.

Methods

fromInteger' :: Integer -> TPrimExp t v

Construct a typed expression from an integer.

fromBoolExp :: TPrimExp Bool v -> TPrimExp t v

Construct a numeric expression from a boolean expression. This can be used to encode arithmetic control flow.

class NumExp t => IntExp t

The class of integer types that can be used for constructing TPrimExps.

class NumExp t => FloatExp t where

The class of floating-point types that can be used for constructing TPrimExps.

Methods

fromRational' :: Rational -> TPrimExp t v

Construct a typed expression from a rational.

sExt :: IntType -> PrimExp v -> PrimExp v

Untyped smart constructor for sign extension that does a bit of constant folding.

zExt :: IntType -> PrimExp v -> PrimExp v

Untyped smart constructor for zero extension that does a bit of constant folding.

(.&&.) :: TPrimExp Bool v -> TPrimExp Bool v -> TPrimExp Bool v infixr 3

Lifted logical conjunction.

(.||.) :: TPrimExp Bool v -> TPrimExp Bool v -> TPrimExp Bool v infixr 2

Lifted logical conjunction.

(.<.) :: TPrimExp t v -> TPrimExp t v -> TPrimExp Bool v infix 4

Lifted relational operators; assuming signed numbers in case of integers.

(.<=.) :: TPrimExp t v -> TPrimExp t v -> TPrimExp Bool v infix 4

Lifted relational operators; assuming signed numbers in case of integers.

(.>.) :: TPrimExp t v -> TPrimExp t v -> TPrimExp Bool v infix 4

Lifted relational operators; assuming signed numbers in case of integers.

(.>=.) :: TPrimExp t v -> TPrimExp t v -> TPrimExp Bool v infix 4

Lifted relational operators; assuming signed numbers in case of integers.

(.==.) :: TPrimExp t v -> TPrimExp t v -> TPrimExp Bool v infix 4

Lifted relational operators; assuming signed numbers in case of integers.

(.&.) :: TPrimExp t v -> TPrimExp t v -> TPrimExp t v

Lifted bitwise operators. The right-shift is logical, *not* arithmetic.

(.|.) :: TPrimExp t v -> TPrimExp t v -> TPrimExp t v

Lifted bitwise operators. The right-shift is logical, *not* arithmetic.

(.^.) :: TPrimExp t v -> TPrimExp t v -> TPrimExp t v

Lifted bitwise operators. The right-shift is logical, *not* arithmetic.

(.>>.) :: TPrimExp t v -> TPrimExp t v -> TPrimExp t v

Lifted bitwise operators. The right-shift is logical, *not* arithmetic.

(.<<.) :: TPrimExp t v -> TPrimExp t v -> TPrimExp t v

Lifted bitwise operators. The right-shift is logical, *not* arithmetic.

bNot :: TPrimExp Bool v -> TPrimExp Bool v

Boolean negation smart constructor.

sMax32 :: TPrimExp Int32 v -> TPrimExp Int32 v -> TPrimExp Int32 v

SMax on 32-bit integers.

sMin32 :: TPrimExp Int32 v -> TPrimExp Int32 v -> TPrimExp Int32 v

SMin on 32-bit integers.

sMax64 :: TPrimExp Int64 v -> TPrimExp Int64 v -> TPrimExp Int64 v

SMax on 64-bit integers.

sMin64 :: TPrimExp Int64 v -> TPrimExp Int64 v -> TPrimExp Int64 v

SMin on 64-bit integers.

sExt32 :: IntExp t => TPrimExp t v -> TPrimExp Int32 v

Sign-extend to 32 bit integer.

sExt64 :: IntExp t => TPrimExp t v -> TPrimExp Int64 v

Sign-extend to 64 bit integer.

zExt32 :: IntExp t => TPrimExp t v -> TPrimExp Int32 v

Zero-extend to 32 bit integer.

zExt64 :: IntExp t => TPrimExp t v -> TPrimExp Int64 v

Zero-extend to 64 bit integer.

fMin64 :: TPrimExp Double v -> TPrimExp Double v -> TPrimExp Double v

64-bit float minimum.

fMax64 :: TPrimExp Double v -> TPrimExp Double v -> TPrimExp Double v

64-bit float maximum.

Untyped construction

(~*~) :: PrimExp v -> PrimExp v -> PrimExp v infix 7

Multiplication of untyped PrimExps, which must have the same type.

(~/~) :: PrimExp v -> PrimExp v -> PrimExp v infix 7

Division of untyped PrimExps, which must have the same type. For integers, this is unsafe signed division.

(~+~) :: PrimExp v -> PrimExp v -> PrimExp v infix 6

Addition of untyped PrimExps, which must have the same type.

(~-~) :: PrimExp v -> PrimExp v -> PrimExp v infix 6

Subtraction of untyped PrimExps, which must have the same type.