Futhark.Analysis.Rephrase

Description

Facilities for changing the rep of some fragment, with no context. We call this "rephrasing", for no deep reason.

Documentation

rephraseProg :: Monad m => Rephraser m from to -> Prog from -> m (Prog to)

Rephrase an entire program.

rephraseFunDef :: Monad m => Rephraser m from to -> FunDef from -> m (FunDef to)

Rephrase a function definition.

rephraseExp :: Monad m => Rephraser m from to -> Exp from -> m (Exp to)

Rephrase an expression.

rephraseBody :: Monad m => Rephraser m from to -> Body from -> m (Body to)

Rephrase a body.

rephraseStm :: Monad m => Rephraser m from to -> Stm from -> m (Stm to)

Rephrase a statement.

rephraseLambda :: Monad m => Rephraser m from to -> Lambda from -> m (Lambda to)

Rephrase a lambda.

rephrasePat :: Monad m => (from -> m to) -> PatT from -> m (PatT to)

Rephrase a pattern.

rephrasePatElem :: Monad m => (from -> m to) -> PatElemT from -> m (PatElemT to)

Rephrase a pattern element.

data Rephraser m from to

A collection of functions that together allow us to rephrase some IR fragment, in some monad m. If we let m be the Maybe monad, we can conveniently do rephrasing that might fail. This is useful if you want to see if some IR in e.g. the Kernels rep actually uses any Kernels-specific operations.

Constructors

Rephraser

Fields

rephraseExpDec :: ExpDec from -> m (ExpDec to)

rephraseLetBoundDec :: LetDec from -> m (LetDec to)

rephraseFParamDec :: FParamInfo from -> m (FParamInfo to)

rephraseLParamDec :: LParamInfo from -> m (LParamInfo to)

rephraseBodyDec :: BodyDec from -> m (BodyDec to)

rephraseRetType :: RetType from -> m (RetType to)

rephraseBranchType :: BranchType from -> m (BranchType to)

rephraseOp :: Op from -> m (Op to)