Futhark.Analysis.SymbolTable

Documentation

data SymbolTable rep

empty :: SymbolTable rep

fromScope :: ASTRep rep => Scope rep -> SymbolTable rep

toScope :: SymbolTable rep -> Scope rep

Entries

data Entry rep

deepen :: SymbolTable rep -> SymbolTable rep

entryAccInput :: Entry rep -> Maybe (WithAccInput rep)

For names that are tokens of an accumulator, this is the corresponding combining function and neutral element.

entryDepth :: Entry rep -> Int

entryLetBoundDec :: Entry rep -> Maybe (LetDec rep)

entryIsSize :: Entry rep -> Bool

True if this name has been used as an array size, implying that it is non-negative.

entryStm :: Entry rep -> Maybe (Stm rep)

entryFParam :: Entry rep -> Maybe (FParamInfo rep)

entryLParam :: Entry rep -> Maybe (LParamInfo rep)

Lookup

elem :: VName -> SymbolTable rep -> Bool

You almost always want available instead of this one.

lookup :: VName -> SymbolTable rep -> Maybe (Entry rep)

lookupStm :: VName -> SymbolTable rep -> Maybe (Stm rep)

lookupExp :: VName -> SymbolTable rep -> Maybe (Exp rep, Certs)

lookupBasicOp :: VName -> SymbolTable rep -> Maybe (BasicOp, Certs)

lookupType :: ASTRep rep => VName -> SymbolTable rep -> Maybe Type

lookupSubExp :: VName -> SymbolTable rep -> Maybe (SubExp, Certs)

lookupAliases :: VName -> SymbolTable rep -> Names

lookupLoopVar :: VName -> SymbolTable rep -> Maybe SubExp

If the given variable name is the name of a ForLoop parameter, then return the bound of that loop.

lookupLoopParam :: VName -> SymbolTable rep -> Maybe (SubExp, SubExp)

available :: VName -> SymbolTable rep -> Bool

In symbol table and not consumed.

consume :: VName -> SymbolTable rep -> SymbolTable rep

index :: ASTRep rep => VName -> [SubExp] -> SymbolTable rep -> Maybe Indexed

index' :: VName -> [TPrimExp Int64 VName] -> SymbolTable rep -> Maybe Indexed

data Indexed

The result of indexing a delayed array.

Constructors

Indexed Certs (PrimExp VName)
A PrimExp based on the indexes (that is, without accessing any actual array).

IndexedArray Certs VName [TPrimExp Int64 VName]
The indexing corresponds to another (perhaps more advantageous) array.

indexedAddCerts :: Certs -> Indexed -> Indexed

class IndexOp op where

Minimal complete definition

Nothing

Methods

indexOp :: (ASTRep rep, IndexOp (Op rep)) => SymbolTable rep -> Int -> op -> [TPrimExp Int64 VName] -> Maybe Indexed

Insertion

insertStm :: (ASTRep rep, IndexOp (Op rep), Aliased rep) => Stm rep -> SymbolTable rep -> SymbolTable rep

insertStms :: (ASTRep rep, IndexOp (Op rep), Aliased rep) => Stms rep -> SymbolTable rep -> SymbolTable rep

insertFParams :: ASTRep rep => [FParam rep] -> SymbolTable rep -> SymbolTable rep

insertLParam :: ASTRep rep => LParam rep -> SymbolTable rep -> SymbolTable rep

insertLoopVar :: ASTRep rep => VName -> IntType -> SubExp -> SymbolTable rep -> SymbolTable rep

insertLoopMerge :: ASTRep rep => [(FParam rep, SubExp, SubExpRes)] -> SymbolTable rep -> SymbolTable rep

Insert entries corresponding to the parameters of a loop (not distinguishing contect and value part). Apart from the parameter itself, we also insert the initial value and the subexpression providing the final value. Note that the latter is likely not in scope in the symbol at this point. This is OK, and can still be used to help some loop optimisations detect invariant loop parameters.

Misc

hideCertified :: Names -> SymbolTable rep -> SymbolTable rep

Hide these definitions, if they are protected by certificates in the set of names.

noteAccTokens :: [(VName, WithAccInput rep)] -> SymbolTable rep -> SymbolTable rep

Note that these names are tokens for the corresponding accumulators. The names must already be present in the symbol table.