Futhark.Analysis.UsageTable

Description

A usage-table is sort of a bottom-up symbol table, describing how (and if) a variable is used.

Documentation

data UsageTable

A usage table.

without :: UsageTable -> [VName] -> UsageTable

Remove these entries from the usage table.

lookup :: VName -> UsageTable -> Maybe Usages

Look up a variable in the usage table.

used :: VName -> UsageTable -> Bool

Is the variable present in the usage table? That is, has it been used?

expand :: (VName -> Names) -> UsageTable -> UsageTable

Expand the usage table based on aliasing information.

isConsumed :: VName -> UsageTable -> Bool

Has the variable been consumed?

isInResult :: VName -> UsageTable -> Bool

Has the variable been used in the Result of a body?

isUsedDirectly :: VName -> UsageTable -> Bool

Has the given name been used directly (i.e. could we rename it or remove it without anyone noticing?)

isSize :: VName -> UsageTable -> Bool

Is this name used as the size of something (array or memory block)?

usages :: Names -> UsageTable

Construct a usage table reflecting that these variables have been used.

usage :: VName -> Usages -> UsageTable

Construct a usage table where the given variable has been used in this specific way.

consumedUsage :: VName -> UsageTable

Construct a usage table where the given variable has been consumed.

inResultUsage :: VName -> UsageTable

Construct a usage table where the given variable has been used in the Result of a body.

sizeUsage :: VName -> UsageTable

Construct a usage table where the given variable has been used as an array or memory size.

sizeUsages :: Names -> UsageTable

Construct a usage table where the given names have been used as an array or memory size.

withoutU :: Usages -> Usages -> Usages

x - y, but for Usages.

data Usages

A description of how a single variable has been used.

consumedU :: Usages

presentU :: Usages

usageInStm :: (ASTRep rep, Aliased rep) => Stm rep -> UsageTable

Produce a usage table reflecting the use of the free variables in a single statement.