Futhark.Bench

Description

Facilities for handling Futhark benchmark results. A Futhark benchmark program is just like a Futhark test program.

Documentation

newtype RunResult

The runtime of a single succesful run.

Constructors

RunResult

Fields

runMicroseconds :: Int

data DataResult

The results for a single named dataset is either an error message, or runtime measurements, the number of bytes used, and the stderr that was produced.

Constructors

DataResult String (Either Text Result)

data BenchResult

The results for all datasets for some benchmark program.

Constructors

BenchResult FilePath [DataResult]

data Result

The measurements resulting from various successful runs of a benchmark (same dataset).

Constructors

Result

Fields

runResults :: [RunResult]

memoryMap :: Map Text Int

stdErr :: Text

encodeBenchResults :: [BenchResult] -> ByteString

Transform benchmark results to a JSON bytestring.

decodeBenchResults :: ByteString -> Either String [BenchResult]

Decode benchmark results from a JSON bytestring.

binaryName :: FilePath -> FilePath

The name we use for compiled programs.

benchmarkDataset :: Server -> RunOptions -> FutharkExe -> FilePath -> Text -> Values -> Maybe Success -> FilePath -> IO (Either Text ([RunResult], Text))

Run the benchmark program on the indicated dataset.

data RunOptions

How to run a benchmark.

Constructors

RunOptions

Fields

runRuns :: Int

runTimeout :: Int

runVerbose :: Int

runResultAction :: Maybe (Int -> IO ())

Invoked for every runtime measured during the run. Can be used to provide a progress bar.

prepareBenchmarkProgram :: MonadIO m => Maybe Int -> CompileOptions -> FilePath -> [InputOutputs] -> m (Either (String, Maybe ByteString) ())

Compile and produce reference datasets.

data CompileOptions

How to compile a benchmark.

Constructors

CompileOptions

Fields

compFuthark :: String

compBackend :: String

compOptions :: [String]