Futhark.Builder.Class

Description

This module defines a convenience typeclass for creating normalised programs.

See Futhark.Construct for a high-level description.

Documentation

class (ASTRep rep, FParamInfo rep ~ DeclType, LParamInfo rep ~ Type, RetType rep ~ DeclExtType, BranchType rep ~ ExtType, SetType (LetDec rep)) => Buildable rep where

The class of representations that can be constructed solely from an expression, within some monad. Very important: the methods should not have any significant side effects! They may be called more often than you think, and the results thrown away. If used exclusively within a MonadBuilder instance, it is acceptable for them to create new bindings, however.

Methods

mkExpPat :: [Ident] -> Exp rep -> Pat rep

mkExpDec :: Pat rep -> Exp rep -> ExpDec rep

mkBody :: Stms rep -> Result -> Body rep

mkLetNames :: (MonadFreshNames m, HasScope rep m) => [VName] -> Exp rep -> m (Stm rep)

mkLet :: Buildable rep => [Ident] -> Exp rep -> Stm rep

Construct a Stm from identifiers for the context- and value part of the pattern, as well as the expression.

mkLet' :: Buildable rep => [Ident] -> StmAux a -> Exp rep -> Stm rep

Like mkLet, but also take attributes and certificates from the given StmAux.

class (ASTRep (Rep m), MonadFreshNames m, Applicative m, Monad m, LocalScope (Rep m) m) => MonadBuilder m where

A monad that supports the creation of bindings from expressions and bodies from bindings, with a specific rep. This is the main typeclass that a monad must implement in order for it to be useful for generating or modifying Futhark code. Most importantly maintains a current state of Stms (as well as a Scope) that have been added with addStm.

Very important: the methods should not have any significant side effects! They may be called more often than you think, and the results thrown away. It is acceptable for them to create new bindings, however.

Minimal complete definition

mkExpDecM, mkBodyM, mkLetNamesM, addStms, collectStms

Associated Types

type Rep m :: Type

Methods

mkExpDecM :: Pat (Rep m) -> Exp (Rep m) -> m (ExpDec (Rep m))

mkBodyM :: Stms (Rep m) -> Result -> m (Body (Rep m))

mkLetNamesM :: [VName] -> Exp (Rep m) -> m (Stm (Rep m))

addStm :: Stm (Rep m) -> m ()

Add a statement to the Stms under construction.

addStms :: Stms (Rep m) -> m ()

Add multiple statements to the Stms under construction.

collectStms :: m a -> m (a, Stms (Rep m))

Obtain the statements constructed during a monadic action, instead of adding them to the state.

certifying :: Certs -> m a -> m a

Add the provided certificates to any statements added during execution of the action.

insertStms :: Buildable rep => Stms rep -> Body rep -> Body rep

Add several bindings at the outermost level of a Body.

insertStm :: Buildable rep => Stm rep -> Body rep -> Body rep

Add a single binding at the outermost level of a Body.

letBind :: MonadBuilder m => Pat (Rep m) -> Exp (Rep m) -> m ()

Add a statement with the given pattern and expression.

letBindNames :: MonadBuilder m => [VName] -> Exp (Rep m) -> m ()

Add a statement with the given pattern element names and expression.

collectStms_ :: MonadBuilder m => m a -> m (Stms (Rep m))

As collectStms, but throw away the ordinary result.

bodyBind :: MonadBuilder m => Body (Rep m) -> m Result

Add the statements of the body, then return the body result.

attributing :: MonadBuilder m => Attrs -> m a -> m a

Add the given attributes to any statements added by this action.

auxing :: MonadBuilder m => StmAux anyrep -> m a -> m a

Add the certificates and attributes to any statements added by this action.

module Futhark.MonadFreshNames