Futhark.CodeGen.Backends.CCUDA.Boilerplate

Description

Various boilerplate definitions for the CUDA backend.

Documentation

generateBoilerplate :: Text -> Text -> [Name] -> Map KernelName KernelSafety -> Map Name SizeClass -> [FailureMsg] -> CompilerM OpenCL () ()

Called after most code has been generated to generate the bulk of the boilerplate.

profilingEnclosure :: Name -> ([BlockItem], [BlockItem])

Block items to put before and after a thing to be profiled.

failureSwitch :: [FailureMsg] -> Stm

copyDevToDev :: Name

copyDevToHost :: Name

copyHostToDev :: Name

copyScalarToDev :: Name

copyScalarFromDev :: Name

kernelRuntime :: KernelName -> Name

kernelRuns :: KernelName -> Name

costCentreReport :: [Name] -> [BlockItem]