Futhark.CodeGen.Backends.GenericC

Description

C code generator framework.

Documentation

compileProg :: MonadFreshNames m => Text -> Text -> Operations op () -> CompilerM op () () -> Text -> [Space] -> [Option] -> Definitions op -> m CParts

Compile imperative program to a C program. Always uses the function named "main" as entry point, so make sure it is defined.

data CParts

The result of compilation to C is multiple parts, which can be put together in various ways. The obvious way is to concatenate all of them, which yields a CLI program. Another is to compile the library part by itself, and use the header file to call into it.

Constructors

CParts

Fields

cHeader :: Text

cUtils :: Text

Utility definitions that must be visible to both CLI and library parts.

cCLI :: Text

cServer :: Text

cLib :: Text

cJsonManifest :: Text

The manifest, in JSON format.

asLibrary :: CParts -> (Text, Text, Text)

Produce header, implementation, and manifest files.

asExecutable :: CParts -> Text

As executable with command-line interface.

asServer :: CParts -> Text

As server executable.

Pluggable compiler

data Operations op s

Constructors

Operations

Fields

opsWriteScalar :: WriteScalar op s

opsReadScalar :: ReadScalar op s

opsAllocate :: Allocate op s

opsDeallocate :: Deallocate op s

opsCopy :: Copy op s

opsStaticArray :: StaticArray op s

opsMemoryType :: MemoryType op s

opsCompiler :: OpCompiler op s

opsError :: ErrorCompiler op s

opsCall :: CallCompiler op s

opsFatMemory :: Bool

If true, use reference counting. Otherwise, bare pointers.

opsCritical :: ([BlockItem], [BlockItem])

Code to bracket critical sections.

defaultOperations :: Operations op s

A set of operations that fail for every operation involving non-default memory spaces. Uses plain pointers and malloc for memory management.

type OpCompiler op s = op -> CompilerM op s ()

A substitute expression compiler, tried before the main compilation function.

type ErrorCompiler op s = ErrorMsg Exp -> String -> CompilerM op s ()

type CallCompiler op s = [VName] -> Name -> [Exp] -> CompilerM op s ()

Call a function.

type PointerQuals op s = String -> CompilerM op s [TypeQual]

The address space qualifiers for a pointer of the given type with the given annotation.

type MemoryType op s = SpaceId -> CompilerM op s Type

The type of a memory block in the given memory space.

type WriteScalar op s = Exp -> Exp -> Type -> SpaceId -> Volatility -> Exp -> CompilerM op s ()

Write a scalar to the given memory block with the given element index and in the given memory space.

writeScalarPointerWithQuals :: PointerQuals op s -> WriteScalar op s

type ReadScalar op s = Exp -> Exp -> Type -> SpaceId -> Volatility -> CompilerM op s Exp

Read a scalar from the given memory block with the given element index and in the given memory space.

readScalarPointerWithQuals :: PointerQuals op s -> ReadScalar op s

type Allocate op s = Exp -> Exp -> Exp -> SpaceId -> CompilerM op s ()

Allocate a memory block of the given size and with the given tag in the given memory space, saving a reference in the given variable name.

type Deallocate op s = Exp -> Exp -> SpaceId -> CompilerM op s ()

De-allocate the given memory block with the given tag, which is in the given memory space.

type Copy op s = Exp -> Exp -> Space -> Exp -> Exp -> Space -> Exp -> CompilerM op s ()

Copy from one memory block to another.

type StaticArray op s = VName -> SpaceId -> PrimType -> ArrayContents -> CompilerM op s ()

Create a static array of values - initialised at load time.

Monadic compiler interface

data CompilerM op s a

data CompilerState s

getUserState :: CompilerM op s s

modifyUserState :: (s -> s) -> CompilerM op s ()

contextContents :: CompilerM op s ([FieldGroup], [Stm], [Stm])

contextFinalInits :: CompilerM op s [Stm]

runCompilerM :: Operations op s -> VNameSource -> s -> CompilerM op s a -> (a, CompilerState s)

inNewFunction :: Bool -> CompilerM op s a -> CompilerM op s a

Used when we, inside an existing CompilerM action, want to generate code for a new function. Use this so that the compiler understands that previously declared memory doesn't need to be freed inside this action.

cachingMemory :: Map VName Space -> ([BlockItem] -> [Stm] -> CompilerM op s a) -> CompilerM op s a

blockScope :: CompilerM op s () -> CompilerM op s [BlockItem]

compileFun :: [BlockItem] -> [Param] -> (Name, Function op) -> CompilerM op s (Definition, Func)

compileCode :: Code op -> CompilerM op s ()

compileExp :: Exp -> CompilerM op s Exp

compilePrimExp :: Monad m => (v -> m Exp) -> PrimExp v -> m Exp

Tell me how to compile a v, and I'll Compile any PrimExp v for you.

compileExpToName :: String -> PrimType -> Exp -> CompilerM op s VName

rawMem :: VName -> CompilerM op s Exp

item :: BlockItem -> CompilerM op s ()

items :: [BlockItem] -> CompilerM op s ()

stm :: Stm -> CompilerM op s ()

stms :: [Stm] -> CompilerM op s ()

decl :: InitGroup -> CompilerM op s ()

atInit :: Stm -> CompilerM op s ()

headerDecl :: HeaderSection -> Definition -> CompilerM op s ()

publicDef :: String -> HeaderSection -> (String -> (Definition, Definition)) -> CompilerM op s String

Construct a publicly visible definition using the specified name as the template. The first returned definition is put in the header file, and the second is the implementation. Returns the public name.

publicDef_ :: String -> HeaderSection -> (String -> (Definition, Definition)) -> CompilerM op s ()

As publicDef, but ignores the public name.

profileReport :: BlockItem -> CompilerM op s ()

onClear :: BlockItem -> CompilerM op s ()

data HeaderSection

In which part of the header file we put the declaration. This is to ensure that the header file remains structured and readable.

Constructors

ArrayDecl String

OpaqueDecl String

EntryDecl

MiscDecl

InitDecl

libDecl :: Definition -> CompilerM op s ()

earlyDecl :: Definition -> CompilerM op s ()

publicName :: String -> CompilerM op s String

Public names must have a consitent prefix.

contextType :: CompilerM op s Type

The generated code must define a struct with this name.

contextField :: Id -> Type -> Maybe Exp -> CompilerM op s ()

contextFieldDyn :: Id -> Type -> Exp -> Stm -> CompilerM op s ()

memToCType :: VName -> Space -> CompilerM op s Type

cacheMem :: ToExp a => a -> CompilerM op s (Maybe VName)

fatMemory :: Space -> CompilerM op s Bool

rawMemCType :: Space -> CompilerM op s Type

cproduct :: [Exp] -> Exp

Return an expression multiplying together the given expressions. If an empty list is given, the expression 1 is returned.

fatMemType :: Space -> Type

freeAllocatedMem :: CompilerM op s [BlockItem]

Building Blocks

primTypeToCType :: PrimType -> Type

The C type corresponding to a primitive type. Integers are assumed to be unsigned.

intTypeToCType :: IntType -> Type

The C type corresponding to a signed integer type.

copyMemoryDefaultSpace :: Exp -> Exp -> Exp -> Exp -> Exp -> CompilerM op s ()