Futhark.CodeGen.Backends.GenericPython

Description

A generic Python code generator which is polymorphic in the type of the operations. Concretely, we use this to handle both sequential and PyOpenCL Python code.

Documentation

compileProg :: MonadFreshNames m => CompilerMode -> String -> Constructor -> [PyStmt] -> [PyStmt] -> Operations op s -> s -> [PyStmt] -> [Option] -> Definitions op -> m Text

data CompilerMode

Are we compiling a library or an executable?

data Constructor

The class generated by the code generator must have a constructor, although it can be vacuous.

Constructors

Constructor [String] [PyStmt]

emptyConstructor :: Constructor

A constructor that takes no arguments and does nothing.

compileName :: VName -> String

compileVar :: VName -> CompilerM op s PyExp

compileDim :: DimSize -> CompilerM op s PyExp

compileExp :: Exp -> CompilerM op s PyExp

compilePrimExp :: Monad m => (v -> m PyExp) -> PrimExp v -> m PyExp

Tell me how to compile a v, and I'll Compile any PrimExp v for you.

compileCode :: Code op -> CompilerM op s ()

compilePrimValue :: PrimValue -> PyExp

compilePrimType :: PrimType -> String

The ctypes type corresponding to a PrimType.

compilePrimTypeExt :: PrimType -> Signedness -> String

The ctypes type corresponding to a PrimType, taking sign into account.

compilePrimToNp :: PrimType -> String

The Numpy type corresponding to a PrimType.

compilePrimToExtNp :: PrimType -> Signedness -> String

The Numpy type corresponding to a PrimType, taking sign into account.

fromStorage :: PrimType -> PyExp -> PyExp

Convert from storage to scalar representation for the given type.

toStorage :: PrimType -> PyExp -> PyExp

Convert from scalar to storage representation for the given type.

data Operations op s

Constructors

Operations

Fields

opsWriteScalar :: WriteScalar op s

opsReadScalar :: ReadScalar op s

opsAllocate :: Allocate op s

opsCopy :: Copy op s

opsStaticArray :: StaticArray op s

opsCompiler :: OpCompiler op s

opsEntryOutput :: EntryOutput op s

opsEntryInput :: EntryInput op s

defaultOperations :: Operations op s

A set of operations that fail for every operation involving non-default memory spaces. Uses plain pointers and malloc for memory management.

unpackDim :: PyExp -> DimSize -> Int32 -> CompilerM op s ()

newtype CompilerM op s a

Constructors

CompilerM (RWS (CompilerEnv op s) [PyStmt] (CompilerState s) a)

type OpCompiler op s = op -> CompilerM op s ()

A substitute expression compiler, tried before the main compilation function.

type WriteScalar op s = PyExp -> PyExp -> PrimType -> SpaceId -> PyExp -> CompilerM op s ()

Write a scalar to the given memory block with the given index and in the given memory space.

type ReadScalar op s = PyExp -> PyExp -> PrimType -> SpaceId -> CompilerM op s PyExp

Read a scalar from the given memory block with the given index and in the given memory space.

type Allocate op s = PyExp -> PyExp -> SpaceId -> CompilerM op s ()

Allocate a memory block of the given size in the given memory space, saving a reference in the given variable name.

type Copy op s = PyExp -> PyExp -> Space -> PyExp -> PyExp -> Space -> PyExp -> PrimType -> CompilerM op s ()

Copy from one memory block to another.

type StaticArray op s = VName -> SpaceId -> PrimType -> ArrayContents -> CompilerM op s ()

Create a static array of values - initialised at load time.

type EntryOutput op s = VName -> SpaceId -> PrimType -> Signedness -> [DimSize] -> CompilerM op s PyExp

Construct the Python array being returned from an entry point.

type EntryInput op s = PyExp -> SpaceId -> PrimType -> Signedness -> [DimSize] -> PyExp -> CompilerM op s ()

Unpack the array being passed to an entry point.

data CompilerEnv op s

Constructors

CompilerEnv

Fields

envOperations :: Operations op s

envVarExp :: Map VName PyExp

data CompilerState s

Constructors

CompilerState

Fields

compNameSrc :: VNameSource

compInit :: [PyStmt]

compUserState :: s

stm :: PyStmt -> CompilerM op s ()

atInit :: PyStmt -> CompilerM op s ()

collect' :: CompilerM op s a -> CompilerM op s (a, [PyStmt])

collect :: CompilerM op s () -> CompilerM op s [PyStmt]

simpleCall :: String -> [PyExp] -> PyExp

A Call where the function is a variable and every argument is a simple Arg.

copyMemoryDefaultSpace :: PyExp -> PyExp -> PyExp -> PyExp -> PyExp -> CompilerM op s ()