Futhark.CodeGen.Backends.SequentialC

Description

C code generator. This module can convert a correct ImpCode program to an equivalent C program. The C code is strictly sequential, but can handle the full Futhark language.

Documentation

compileProg :: MonadFreshNames m => Text -> Prog SeqMem -> m (Warnings, CParts)

Compile the program to sequential C.

data CParts

The result of compilation to C is multiple parts, which can be put together in various ways. The obvious way is to concatenate all of them, which yields a CLI program. Another is to compile the library part by itself, and use the header file to call into it.

Constructors

CParts

Fields

cHeader :: Text

cUtils :: Text

Utility definitions that must be visible to both CLI and library parts.

cCLI :: Text

cServer :: Text

cLib :: Text

cJsonManifest :: Text

The manifest, in JSON format.

asLibrary :: CParts -> (Text, Text, Text)

Produce header, implementation, and manifest files.

asExecutable :: CParts -> Text

As executable with command-line interface.

asServer :: CParts -> Text

As server executable.