Futhark.CodeGen.Backends.SimpleRep

Description

Simple C runtime representation.

Most types use the same memory and scalar variable representation. For those that do not (as of this writing, only Float16), we use primStorageType for the array element representation, and primTypeToCType for their scalar representation. Use toStorage and fromStorage to convert back and forth.

Documentation

tupleField :: Int -> String

tupleField i is the name of field number i in a tuple.

funName :: Name -> String

funName f is the name of the C function corresponding to the Futhark function f.

defaultMemBlockType :: Type

The type of memory blocks in the default memory space.

intTypeToCType :: IntType -> Type

The C type corresponding to a signed integer type.

primTypeToCType :: PrimType -> Type

The C type corresponding to a primitive type. Integers are assumed to be unsigned.

primStorageType :: PrimType -> Type

The C storage type for arrays of this primitive type.

primAPIType :: Signedness -> PrimType -> Type

The C API corresponding to a primitive type. Integers are assumed to have the specified sign.

arrayName :: PrimType -> Signedness -> Int -> String

The name of exposed array type structs.

opaqueName :: String -> [ValueDesc] -> String

The name of exposed opaque types.

toStorage :: PrimType -> Exp -> Exp

Convert from scalar to storage representation for the given type.

fromStorage :: PrimType -> Exp -> Exp

Convert from storage to scalar representation for the given type.

cproduct :: [Exp] -> Exp

Return an expression multiplying together the given expressions. If an empty list is given, the expression 1 is returned.

csum :: [Exp] -> Exp

Return an expression summing the given expressions. If an empty list is given, the expression 0 is returned.

scalarToPrim :: Text -> (Signedness, PrimType)

The PrimType (and sign) correspond to a human-readable scalar type name (e.g. f64). Beware: partial!

Primitive value operations

cScalarDefs :: Text

Implementations of scalar operations.

Storing/restoring values in byte sequences

storageSize :: PrimType -> Int -> Exp -> Exp

storeValueHeader :: Signedness -> PrimType -> Int -> Exp -> Exp -> [Stm]

loadValueHeader :: Signedness -> PrimType -> Int -> Exp -> Exp -> [Stm]

Orphan instances

ToIdent Text

Instance details

Methods

toIdent :: Text -> SrcLoc -> Id

ToIdent VName

Instance details

Methods

toIdent :: VName -> SrcLoc -> Id

ToIdent Name

Instance details

Methods

toIdent :: Name -> SrcLoc -> Id

ToExp PrimValue

Instance details

Methods

toExp :: PrimValue -> SrcLoc -> Exp

ToExp FloatValue

Instance details

Methods

toExp :: FloatValue -> SrcLoc -> Exp

ToExp IntValue

Instance details

Methods

toExp :: IntValue -> SrcLoc -> Exp

ToExp VName

Instance details

Methods

toExp :: VName -> SrcLoc -> Exp

ToExp SubExp

Instance details

Methods

toExp :: SubExp -> SrcLoc -> Exp