Futhark.CodeGen.ImpCode

Description

Imperative intermediate language used as a stepping stone in code generation.

This is a generic representation parametrised on an extensible arbitrary operation.

Originally inspired by the paper "Defunctionalizing Push Arrays" (FHPC '14).

Documentation

data Definitions a

A collection of imperative functions and constants.

Constructors

Definitions

Fields

defConsts :: Constants a

defFuns :: Functions a

newtype Functions a

A collection of imperative functions.

Constructors

Functions [(Name, Function a)]

type Function = FunctionT

Type alias for namespace control.

data FunctionT a

A imperative function, containing the body as well as its low-level inputs and outputs, as well as its high-level arguments and results. The latter are only used if the function is an entry point.

Constructors

Function

Fields

functionEntry :: Maybe Name

functionOutput :: [Param]

functionInput :: [Param]

functionBody :: Code a

functionResult :: [ExternalValue]

functionArgs :: [(Name, ExternalValue)]

data Constants a

A collection of imperative constants.

Constructors

Constants

Fields

constsDecl :: [Param]

The constants that are made available to the functions.

constsInit :: Code a

Setting the value of the constants. Note that this must not contain declarations of the names defined in constsDecl.

data ValueDesc

A description of an externally meaningful value.

Constructors

ArrayValue VName Space PrimType Signedness [DimSize]
An array with memory block memory space, element type, signedness of element type (if applicable), and shape.

ScalarValue PrimType Signedness VName
A scalar value with signedness if applicable.

data Signedness

Since the core language does not care for signedness, but the source language does, entry point input/output information has metadata for integer types (and arrays containing these) that indicate whether they are really unsigned integers.

Constructors

TypeUnsigned

TypeDirect

data ExternalValue

^ An externally visible value. This can be an opaque value (covering several physical internal values), or a single value that can be used externally. We record the uniqueness because it is important to the external interface as well.

Constructors

OpaqueValue Uniqueness String [ValueDesc]
The string is a human-readable description with no other semantics. not matter.

TransparentValue Uniqueness ValueDesc

data Param

An ImpCode function parameter.

Constructors

MemParam VName Space

ScalarParam VName PrimType

paramName :: Param -> VName

The name of a parameter.

data SubExp

A subexpression is either a scalar constant or a variable. One important property is that evaluation of a subexpression is guaranteed to complete in constant time.

Constructors

Constant PrimValue

Var VName

type MemSize = SubExp

The size of a memory block.

type DimSize = SubExp

The size of an array.

data Space

The memory space of a block. If DefaultSpace, this is the "default" space, whatever that is. The exact meaning of the SpaceId depends on the backend used. In GPU kernels, for example, this is used to distinguish between constant, global and shared memory spaces. In GPU-enabled host code, it is used to distinguish between host memory (DefaultSpace) and GPU space.

Constructors

DefaultSpace

Space SpaceId

ScalarSpace [SubExp] PrimType
A special kind of memory that is a statically sized array of some primitive type. Used for private memory on GPUs.

type SpaceId = String

A string representing a specific non-default memory space.

data Code a

A block of imperative code. Parameterised by an Op, which allows extensibility. Concrete uses of this type will instantiate the type parameter with e.g. a construct for launching GPU kernels.

Constructors

Skip
No-op. Crucial for the Monoid instance.

(:>>:)

Fields

(Code a)

Statement composition. Crucial for the Semigroup instance.

(Code a)

For VName Exp (Code a)
A for-loop iterating the given number of times. The loop parameter starts counting from zero and will have the same (integer) type as the bound. The bound is evaluated just once, before the loop is entered.

While (TExp Bool) (Code a)
While loop. The conditional is (of course) re-evaluated before every iteration of the loop.

DeclareMem VName Space
Declare a memory block variable that will point to memory in the given memory space. Note that this is distinct from allocation. The memory block must be the target of either an Allocate or a SetMem before it can be used for reading or writing.

DeclareScalar VName Volatility PrimType
Declare a scalar variable with an initially undefined value.

DeclareArray VName Space PrimType ArrayContents
Create an array containing the given values. The lifetime of the array will be the entire application. This is mostly used for constant arrays, but also for some bookkeeping data, like the synchronisation counts used to implement reduction.

Allocate VName (Count Bytes (TExp Int64)) Space
Memory space must match the corresponding DeclareMem.

Free VName Space
Indicate that some memory block will never again be referenced via the indicated variable. However, it may still be accessed through aliases. It is only safe to actually deallocate the memory block if this is the last reference. There is no guarantee that all memory blocks will be freed with this statement. Backends are free to ignore it entirely.

Copy VName (Count Bytes (TExp Int64)) Space VName (Count Bytes (TExp Int64)) Space (Count Bytes (TExp Int64))
Destination, offset in destination, destination space, source, offset in source, offset space, number of bytes.

Write VName (Count Elements (TExp Int64)) PrimType Space Volatility Exp
Write mem i t space vol v writes the value v to mem offset by i elements of type t. The Space argument is the memory space of mem (technically redundant, but convenient). Note that reading is done with an Exp (Read).

SetScalar VName Exp
Set a scalar variable.

Read VName VName (Count Elements (TExp Int64)) PrimType Space Volatility
Read a scalar from memory from memory. The first VName is the target scalar variable, and the remaining arguments have the same meaning as with Write.

SetMem VName VName Space
Must be in same space.

Call [VName] Name [Arg]
Function call. The results are written to the provided VName variables.

If (TExp Bool) (Code a) (Code a)
Conditional execution.

Assert Exp (ErrorMsg Exp) (SrcLoc, [SrcLoc])
Assert that something must be true. Should it turn out not to be true, then report a failure along with the given error message.

Comment String (Code a)
Has the same semantics as the contained code, but the comment should show up in generated code for ease of inspection.

DebugPrint String (Maybe Exp)
Print the given value to the screen, somehow annotated with the given string as a description. If no type/value pair, just print the string. This has no semantic meaning, but is used entirely for debugging. Code generators are free to ignore this statement.

TracePrint (ErrorMsg Exp)
Log the given message, *without* a trailing linebreak (unless part of the mssage).

Op a
Perform an extensible operation.

data PrimValue

Non-array values.

Constructors

IntValue !IntValue

FloatValue !FloatValue

BoolValue !Bool

UnitValue
The only value of type Unit.

type Exp = PrimExp VName

A side-effect free expression whose execution will produce a single primitive value.

type TExp t = TPrimExp t VName

Like Exp, but with a required/known type.

data Volatility

The volatility of a memory access or variable. Feel free to ignore this for backends where it makes no sense (anything but C and similar low-level things)

Constructors

Volatile

Nonvolatile

data Arg

A function call argument.

Constructors

ExpArg Exp

MemArg VName

var :: VName -> PrimType -> Exp

Turn a VName into a Exp.

newtype ErrorMsg a

An error message is a list of error parts, which are concatenated to form the final message.

Constructors

ErrorMsg [ErrorMsgPart a]

data ErrorMsgPart a

A part of an error message.

Constructors

ErrorString String
A literal string.

ErrorVal PrimType a
A run-time value.

errorMsgArgTypes :: ErrorMsg a -> [PrimType]

How many non-constant parts does the error message have, and what is their type?

data ArrayContents

The contents of a statically declared constant array. Such arrays are always unidimensional, and reshaped if necessary in the code that uses them.

Constructors

ArrayValues [PrimValue]
Precisely these values.

ArrayZeros Int
This many zeroes.

declaredIn :: Code a -> Names

The names declared with DeclareMem, DeclareScalar, and DeclareArray in the given code.

lexicalMemoryUsage :: Function a -> Map VName Space

Find those memory blocks that are used only lexically. That is, are not used as the source or target of a SetMem, or are the result of the function, nor passed as arguments to other functions. This is interesting because such memory blocks do not need reference counting, but can be managed in a purely stack-like fashion.

We do not look inside any Ops. We assume that no Op is going to SetMem a memory block declared outside it.

calledFuncs :: Code a -> Set Name

The set of functions that are called by this code. Assumes there are no function calls in Ops.

Typed enumerations

data Bytes

Phantom type for a count of bytes.

data Elements

Phantom type for a count of elements.

elements :: a -> Count Elements a

This expression counts elements.

bytes :: a -> Count Bytes a

This expression counts bytes.

withElemType :: Count Elements (TExp Int64) -> PrimType -> Count Bytes (TExp Int64)

Convert a count of elements into a count of bytes, given the per-element size.

Re-exports from other modules.

pretty :: Pretty a => a -> String

Prettyprint a value, wrapped to 80 characters.

module Language.Futhark.Core

module Futhark.IR.Primitive

module Futhark.Analysis.PrimExp

module Futhark.Analysis.PrimExp.Convert

newtype Count u e

A wrapper supporting a phantom type for indicating what we are counting.

Constructors

Count

Fields

unCount :: e

module Futhark.IR.Prop.Names