Futhark.CodeGen.ImpCode.OpenCL

Description

Imperative code with an OpenCL component.

Apart from ordinary imperative code, this also carries around an OpenCL program as a string, as well as a list of kernels defined by the OpenCL program.

The imperative code has been augmented with a LaunchKernel operation that allows one to execute an OpenCL kernel.

Documentation

data Program

An program calling OpenCL kernels.

Constructors

Program

Fields

openClProgram :: Text

openClPrelude :: Text

Must be prepended to the program.

openClKernelNames :: Map KernelName KernelSafety

openClUsedTypes :: [PrimType]

So we can detect whether the device is capable.

openClSizes :: Map Name SizeClass

Runtime-configurable constants.

openClFailures :: [FailureMsg]

Assertion failure error messages.

hostDefinitions :: Definitions OpenCL

type Function = Function OpenCL

A function calling OpenCL kernels.

data FunctionT a

A imperative function, containing the body as well as its low-level inputs and outputs, as well as its high-level arguments and results. The latter are only used if the function is an entry point.

Constructors

Function (Maybe Name) [Param] [Param] (Code a) [ExternalValue] [(Name, ExternalValue)]

type Code = Code OpenCL

A piece of code calling OpenCL.

type KernelName = Name

The name of a kernel.

data KernelArg

An argument to be passed to a kernel.

Constructors

ValueKArg Exp PrimType
Pass the value of this scalar expression as argument.

MemKArg VName
Pass this pointer as argument.

SharedMemoryKArg (Count Bytes Exp)
Create this much local memory per workgroup.

data OpenCL

Host-level OpenCL operation.

Constructors

LaunchKernel KernelSafety KernelName [KernelArg] [Exp] [Exp]

GetSize VName Name

CmpSizeLe VName Name Exp

GetSizeMax VName SizeClass

data KernelSafety

Information about bounds checks and how sensitive it is to errors. Ordered by least demanding to most.

Constructors

SafetyNone
Does not need to know if we are in a failing state, and also cannot fail.

SafetyCheap
Needs to be told if there's a global failure, and that's it, and cannot fail.

SafetyFull
Needs all parameters, may fail itself.

numFailureParams :: KernelSafety -> Int

How many leading failure arguments we must pass when launching a kernel with these safety characteristics.

data KernelTarget

The target platform when compiling imperative code to a Program

Constructors

TargetOpenCL

TargetCUDA

data FailureMsg

Something that can go wrong in a kernel. Part of the machinery for reporting error messages from within kernels.

Constructors

FailureMsg

Fields

failureError :: ErrorMsg Exp

failureBacktrace :: String

data Int8

data Int16

data Int32

data Int64

data Word8

data Word16

data Word32

data Word64

data Half

srclocOf :: Located a => a -> SrcLoc

The SrcLoc of a Located value.

data Loc

Location type, consisting of a beginning position and an end position.

data SrcLoc

Source location type. Source location are all equal, which allows AST nodes to be compared modulo location information.

class Located a where

Located values have a location.

Minimal complete definition

locOf

Methods

locOf :: a -> Loc

locOfList :: [a] -> Loc

pretty :: Pretty a => a -> String

Prettyprint a value, wrapped to 80 characters.

data ConvOp

Conversion operators try to generalise the from t0 x to t1 instructions from LLVM.

Constructors

ZExt IntType IntType
Zero-extend the former integer type to the latter. If the new type is smaller, the result is a truncation.

SExt IntType IntType
Sign-extend the former integer type to the latter. If the new type is smaller, the result is a truncation.

FPConv FloatType FloatType
Convert value of the former floating-point type to the latter. If the new type is smaller, the result is a truncation.

FPToUI FloatType IntType
Convert a floating-point value to the nearest unsigned integer (rounding towards zero).

FPToSI FloatType IntType
Convert a floating-point value to the nearest signed integer (rounding towards zero).

UIToFP IntType FloatType
Convert an unsigned integer to a floating-point value.

SIToFP IntType FloatType
Convert a signed integer to a floating-point value.

IToB IntType
Convert an integer to a boolean value. Zero becomes false; anything else is true.

BToI IntType
Convert a boolean to an integer. True is converted to 1 and False to 0.

data CmpOp

Comparison operators are like BinOps, but they always return a boolean value. The somewhat ugly constructor names are straight out of LLVM.

Constructors

CmpEq PrimType
All types equality.

CmpUlt IntType
Unsigned less than.

CmpUle IntType
Unsigned less than or equal.

CmpSlt IntType
Signed less than.

CmpSle IntType
Signed less than or equal.

FCmpLt FloatType
Floating-point less than.

FCmpLe FloatType
Floating-point less than or equal.

CmpLlt
Boolean less than.

CmpLle
Boolean less than or equal.

data BinOp

Binary operators. These correspond closely to the binary operators in LLVM. Most are parametrised by their expected input and output types.

Constructors

Add IntType Overflow
Integer addition.

FAdd FloatType
Floating-point addition.

Sub IntType Overflow
Integer subtraction.

FSub FloatType
Floating-point subtraction.

Mul IntType Overflow
Integer multiplication.

FMul FloatType
Floating-point multiplication.

UDiv IntType Safety
Unsigned integer division. Rounds towards negativity infinity. Note: this is different from LLVM.

UDivUp IntType Safety
Unsigned integer division. Rounds towards positive infinity.

SDiv IntType Safety
Signed integer division. Rounds towards negativity infinity. Note: this is different from LLVM.

SDivUp IntType Safety
Signed integer division. Rounds towards positive infinity.

FDiv FloatType
Floating-point division.

FMod FloatType
Floating-point modulus.

UMod IntType Safety
Unsigned integer modulus; the countepart to UDiv.

SMod IntType Safety
Signed integer modulus; the countepart to SDiv.

SQuot IntType Safety
Signed integer division. Rounds towards zero. This corresponds to the sdiv instruction in LLVM and integer division in C.

SRem IntType Safety
Signed integer division. Rounds towards zero. This corresponds to the srem instruction in LLVM and integer modulo in C.

SMin IntType
Returns the smallest of two signed integers.

UMin IntType
Returns the smallest of two unsigned integers.

FMin FloatType
Returns the smallest of two floating-point numbers.

SMax IntType
Returns the greatest of two signed integers.

UMax IntType
Returns the greatest of two unsigned integers.

FMax FloatType
Returns the greatest of two floating-point numbers.

Shl IntType
Left-shift.

LShr IntType
Logical right-shift, zero-extended.

AShr IntType
Arithmetic right-shift, sign-extended.

And IntType
Bitwise and.

Or IntType
Bitwise or.

Xor IntType
Bitwise exclusive-or.

Pow IntType
Integer exponentiation.

FPow FloatType
Floating-point exponentiation.

LogAnd
Boolean and - not short-circuiting.

LogOr
Boolean or - not short-circuiting.

data Safety

Whether something is safe or unsafe (mostly function calls, and in the context of whether operations are dynamically checked). When we inline an Unsafe function, we remove all safety checks in its body. The Ord instance picks Unsafe as being less than Safe.

For operations like integer division, a safe division will not explode the computer in case of division by zero, but instead return some unspecified value. This always involves a run-time check, so generally the unsafe variant is what the compiler will insert, but guarded by an explicit assertion elsewhere. Safe operations are useful when the optimiser wants to move e.g. a division to a location where the divisor may be zero, but where the result will only be used when it is non-zero (so it doesn't matter what result is provided with a zero divisor, as long as the program keeps running).

Constructors

Unsafe

Safe

data Overflow

What to do in case of arithmetic overflow. Futhark's semantics are that overflow does wraparound, but for generated code (like address arithmetic), it can be beneficial for overflow to be undefined behaviour, as it allows better optimisation of things such as GPU kernels.

Note that all values of this type are considered equal for Eq and Ord.

Constructors

OverflowWrap

OverflowUndef

data UnOp

Various unary operators. It is a bit ad-hoc what is a unary operator and what is a built-in function. Perhaps these should all go away eventually.

Constructors

Not
E.g., ! True == False.

Complement IntType
E.g., ~(~1) = 1.

Abs IntType
abs(-2) = 2.

FAbs FloatType
fabs(-2.0) = 2.0.

SSignum IntType
Signed sign function: ssignum(-2) = -1.

USignum IntType
Unsigned sign function: usignum(2) = 1.

FSignum FloatType
Floating-point sign function.

data PrimValue

Non-array values.

Constructors

IntValue !IntValue

FloatValue !FloatValue

BoolValue !Bool

UnitValue
The only value of type Unit.

data FloatValue

A floating-point value.

Constructors

Float16Value !Half

Float32Value !Float

Float64Value !Double

data IntValue

An integer value.

Constructors

Int8Value !Int8

Int16Value !Int16

Int32Value !Int32

Int64Value !Int64

data PrimType

Low-level primitive types.

Constructors

IntType IntType

FloatType FloatType

Bool

Unit
An informationless type - An array of this type takes up no space.

data FloatType

A floating point type.

Constructors

Float16

Float32

Float64

data IntType

An integer type, ordered by size. Note that signedness is not a property of the type, but a property of the operations performed on values of these types.

Constructors

Int8

Int16

Int32

Int64

allIntTypes :: [IntType]

A list of all integer types.

allFloatTypes :: [FloatType]

A list of all floating-point types.

allPrimTypes :: [PrimType]

A list of all primitive types.

intValue :: Integral int => IntType -> int -> IntValue

Create an IntValue from a type and an Integer.

intValueType :: IntValue -> IntType

The type of an integer value.

valueIntegral :: Integral int => IntValue -> int

Convert an IntValue to any Integral type.

floatValue :: Real num => FloatType -> num -> FloatValue

Create a FloatValue from a type and a Rational.

floatValueType :: FloatValue -> FloatType

The type of a floating-point value.

primValueType :: PrimValue -> PrimType

The type of a basic value.

blankPrimValue :: PrimType -> PrimValue

A "blank" value of the given primitive type - this is zero, or whatever is close to it. Don't depend on this value, but use it for e.g. creating arrays to be populated by do-loops.

allUnOps :: [UnOp]

A list of all unary operators for all types.

allBinOps :: [BinOp]

A list of all binary operators for all types.

allCmpOps :: [CmpOp]

A list of all comparison operators for all types.

allConvOps :: [ConvOp]

A list of all conversion operators for all types.

doUnOp :: UnOp -> PrimValue -> Maybe PrimValue

Apply an UnOp to an operand. Returns Nothing if the application is mistyped.

doComplement :: IntValue -> IntValue

E.g., ~(~1) = 1.

doAbs :: IntValue -> IntValue

abs(-2) = 2.

doFAbs :: FloatValue -> FloatValue

abs(-2.0) = 2.0.

doSSignum :: IntValue -> IntValue

ssignum(-2) = -1.

doUSignum :: IntValue -> IntValue

usignum(-2) = -1.

doBinOp :: BinOp -> PrimValue -> PrimValue -> Maybe PrimValue

Apply a BinOp to an operand. Returns Nothing if the application is mistyped, or outside the domain (e.g. division by zero).

doAdd :: IntValue -> IntValue -> IntValue

Integer addition.

doMul :: IntValue -> IntValue -> IntValue

Integer multiplication.

doSDiv :: IntValue -> IntValue -> Maybe IntValue

Signed integer division. Rounds towards negativity infinity. Note: this is different from LLVM.

doSMod :: IntValue -> IntValue -> Maybe IntValue

Signed integer modulus; the countepart to SDiv.

doPow :: IntValue -> IntValue -> Maybe IntValue

Signed integer exponentatation.

doConvOp :: ConvOp -> PrimValue -> Maybe PrimValue

Apply a ConvOp to an operand. Returns Nothing if the application is mistyped.

flipConvOp :: ConvOp -> ConvOp

Turn the conversion the other way around. Note that most conversions are lossy, so there is no guarantee the value will round-trip.

doZExt :: IntValue -> IntType -> IntValue

Zero-extend the given integer value to the size of the given type. If the type is smaller than the given value, the result is a truncation.

doSExt :: IntValue -> IntType -> IntValue

Sign-extend the given integer value to the size of the given type. If the type is smaller than the given value, the result is a truncation.

doFPConv :: FloatValue -> FloatType -> FloatValue

Convert the former floating-point type to the latter.

doFPToUI :: FloatValue -> IntType -> IntValue

Convert a floating-point value to the nearest unsigned integer (rounding towards zero).

doFPToSI :: FloatValue -> IntType -> IntValue

Convert a floating-point value to the nearest signed integer (rounding towards zero).

doUIToFP :: IntValue -> FloatType -> FloatValue

Convert an unsigned integer to a floating-point value.

doSIToFP :: IntValue -> FloatType -> FloatValue

Convert a signed integer to a floating-point value.

doCmpOp :: CmpOp -> PrimValue -> PrimValue -> Maybe Bool

Apply a CmpOp to an operand. Returns Nothing if the application is mistyped.

doCmpEq :: PrimValue -> PrimValue -> Bool

Compare any two primtive values for exact equality.

doCmpUlt :: IntValue -> IntValue -> Bool

Unsigned less than.

doCmpUle :: IntValue -> IntValue -> Bool

Unsigned less than or equal.

doCmpSlt :: IntValue -> IntValue -> Bool

Signed less than.

doCmpSle :: IntValue -> IntValue -> Bool

Signed less than or equal.

doFCmpLt :: FloatValue -> FloatValue -> Bool

Floating-point less than.

doFCmpLe :: FloatValue -> FloatValue -> Bool

Floating-point less than or equal.

intToWord64 :: IntValue -> Word64

Translate an IntValue to Word64. This is guaranteed to fit.

intToInt64 :: IntValue -> Int64

Translate an IntValue to Int64. This is guaranteed to fit.

binOpType :: BinOp -> PrimType

The result type of a binary operator.

cmpOpType :: CmpOp -> PrimType

The operand types of a comparison operator.

unOpType :: UnOp -> PrimType

The operand and result type of a unary operator.

convOpType :: ConvOp -> (PrimType, PrimType)

The input and output types of a conversion operator.

primFuns :: Map String ([PrimType], PrimType, [PrimValue] -> Maybe PrimValue)

A mapping from names of primitive functions to their parameter types, their result type, and a function for evaluating them.

zeroIsh :: PrimValue -> Bool

Is the given value kind of zero?

oneIsh :: PrimValue -> Bool

Is the given value kind of one?

negativeIsh :: PrimValue -> Bool

Is the given value kind of negative?

zeroIshInt :: IntValue -> Bool

Is the given integer value kind of zero?

oneIshInt :: IntValue -> Bool

Is the given integer value kind of one?

primBitSize :: PrimType -> Int

The size of a value of a given primitive type in bites.

primByteSize :: Num a => PrimType -> a

The size of a value of a given primitive type in eight-bit bytes.

intByteSize :: Num a => IntType -> a

The size of a value of a given integer type in eight-bit bytes.

floatByteSize :: Num a => FloatType -> a

The size of a value of a given floating-point type in eight-bit bytes.

commutativeBinOp :: BinOp -> Bool

True if the given binary operator is commutative.

convOpFun :: ConvOp -> String

The human-readable name for a ConvOp. This is used to expose the ConvOp in the intrinsics module of a Futhark program.

prettySigned :: Bool -> PrimType -> String

True if signed. Only makes a difference for integer types.

data VName

A name tagged with some integer. Only the integer is used in comparisons, no matter the type of vn.

Constructors

VName !Name !Int

data Name

The abstract (not really) type representing names in the Futhark compiler. Strings, being lists of characters, are very slow, while Texts are based on byte-arrays.

data Uniqueness

The uniqueness attribute of a type. This essentially indicates whether or not in-place modifications are acceptable. With respect to ordering, Unique is greater than Nonunique.

Constructors

Nonunique
May have references outside current function.

Unique
No references outside current function.

defaultEntryPoint :: Name

The name of the default program entry point (main).

nameToString :: Name -> String

Convert a name to the corresponding list of characters.

nameFromString :: String -> Name

Convert a list of characters to the corresponding name.

nameToText :: Name -> Text

Convert a name to the corresponding Text.

nameFromText :: Text -> Name

Convert a Text to the corresponding name.

locStr :: Located a => a -> String

A human-readable location string, of the form filename:lineno:columnno. This follows the GNU coding standards for error messages: https://www.gnu.org/prep/standards/html_node/Errors.html

This function assumes that both start and end position is in the same file (it is not clear what the alternative would even mean).

locStrRel :: (Located a, Located b) => a -> b -> String

Like locStr, but locStrRel prev now prints the location now with the file name left out if the same as prev. This is useful when printing messages that are all in the context of some initially printed location (e.g. the first mention contains the file name; the rest just line and column name).

prettyStacktrace :: Int -> [String] -> String

Given a list of strings representing entries in the stack trace and the index of the frame to highlight, produce a final newline-terminated string for showing to the user. This string should also be preceded by a newline. The most recent stack frame must come first in the list.

baseTag :: VName -> Int

Return the tag contained in the VName.

baseName :: VName -> Name

Return the name contained in the VName.

baseString :: VName -> String

Return the base Name converted to a string.

quote :: String -> String

Enclose a string in the prefered quotes used in error messages. These are picked to not collide with characters permitted in identifiers.

pquote :: Doc -> Doc

As quote, but works on prettyprinted representation.

data ErrorMsgPart a

A part of an error message.

Constructors

ErrorString String
A literal string.

ErrorVal PrimType a
A run-time value.

newtype ErrorMsg a

An error message is a list of error parts, which are concatenated to form the final message.

Constructors

ErrorMsg [ErrorMsgPart a]

data SubExp

A subexpression is either a scalar constant or a variable. One important property is that evaluation of a subexpression is guaranteed to complete in constant time.

Constructors

Constant PrimValue

Var VName

type SpaceId = String

A string representing a specific non-default memory space.

data Space

The memory space of a block. If DefaultSpace, this is the "default" space, whatever that is. The exact meaning of the SpaceId depends on the backend used. In GPU kernels, for example, this is used to distinguish between constant, global and shared memory spaces. In GPU-enabled host code, it is used to distinguish between host memory (DefaultSpace) and GPU space.

Constructors

DefaultSpace

Space SpaceId

ScalarSpace [SubExp] PrimType
A special kind of memory that is a statically sized array of some primitive type. Used for private memory on GPUs.

errorMsgArgTypes :: ErrorMsg a -> [PrimType]

How many non-constant parts does the error message have, and what is their type?

class FreeIn dec => FreeDec dec where

Either return precomputed free names stored in the attribute, or the freshly computed names. Relies on lazy evaluation to avoid the work.

Minimal complete definition

Nothing

Methods

precomputed :: dec -> FV -> FV

class FreeIn a where

A class indicating that we can obtain free variable information from values of this type.

Minimal complete definition

Nothing

Methods

freeIn' :: a -> FV

data FV

A computation to build a free variable set.

data Names

A set of names. Note that the Ord instance is a dummy that treats everything as EQ if ==, and otherwise LT.

namesIntMap :: Names -> IntMap VName

Retrieve the data structure underlying the names representation.

nameIn :: VName -> Names -> Bool

Does the set of names contain this name?

namesFromList :: [VName] -> Names

Construct a name set from a list. Slow.

namesToList :: Names -> [VName]

Turn a name set into a list of names. Slow.

oneName :: VName -> Names

Construct a name set from a single name.

namesIntersection :: Names -> Names -> Names

The intersection of two name sets.

namesIntersect :: Names -> Names -> Bool

Do the two name sets intersect?

namesSubtract :: Names -> Names -> Names

Subtract the latter name set from the former.

mapNames :: (VName -> VName) -> Names -> Names

Map over the names in a set.

fvBind :: Names -> FV -> FV

Consider a variable to be bound in the given FV computation.

fvName :: VName -> FV

Take note of a variable reference.

fvNames :: Names -> FV

Take note of a set of variable references.

freeInStmsAndRes :: (FreeIn (Op rep), FreeIn (LetDec rep), FreeIn (LParamInfo rep), FreeIn (FParamInfo rep), FreeDec (BodyDec rep), FreeIn (RetType rep), FreeIn (BranchType rep), FreeDec (ExpDec rep)) => Stms rep -> Result -> FV

Return the set of variable names that are free in the given statements and result. Filters away the names that are bound by the statements.

freeIn :: FreeIn a => a -> Names

The free variables of some syntactic construct.

boundInBody :: Body rep -> Names

The names bound by the bindings immediately in a Body.

boundByStm :: Stm rep -> Names

The names bound by a binding.

boundByStms :: Stms rep -> Names

The names bound by the bindings.

boundByLambda :: Lambda rep -> [VName]

The names of the lambda parameters plus the index parameter.

class NumExp t => FloatExp t where

The class of floating-point types that can be used for constructing TPrimExps.

Methods

fromRational' :: Rational -> TPrimExp t v

Construct a typed expression from a rational.

class NumExp t => IntExp t

The class of integer types that can be used for constructing TPrimExps.

class NumExp t where

The class of numeric types that can be used for constructing TPrimExps.

Methods

fromInteger' :: Integer -> TPrimExp t v

Construct a typed expression from an integer.

fromBoolExp :: TPrimExp Bool v -> TPrimExp t v

Construct a numeric expression from a boolean expression. This can be used to encode arithmetic control flow.

newtype TPrimExp t v

A PrimExp tagged with a phantom type used to provide type-safe construction. Does not guarantee that the underlying expression is actually type correct.

Constructors

TPrimExp

Fields

untyped :: PrimExp v

data PrimExp v

A primitive expression parametrised over the representation of free variables. Note that the Functor, Traversable, and Num instances perform automatic (but simple) constant folding.

Note also that the Num instance assumes OverflowUndef semantics!

Constructors

LeafExp v PrimType

ValueExp PrimValue

BinOpExp BinOp (PrimExp v) (PrimExp v)

CmpOpExp CmpOp (PrimExp v) (PrimExp v)

UnOpExp UnOp (PrimExp v)

ConvOpExp ConvOp (PrimExp v)

FunExp String [PrimExp v] PrimType

isInt8 :: PrimExp v -> TPrimExp Int8 v

This expression is of type Int8.

isInt16 :: PrimExp v -> TPrimExp Int16 v

This expression is of type Int16.

isInt32 :: PrimExp v -> TPrimExp Int32 v

This expression is of type Int32.

isInt64 :: PrimExp v -> TPrimExp Int64 v

This expression is of type Int64.

isBool :: PrimExp v -> TPrimExp Bool v

This is a boolean expression.

isF16 :: PrimExp v -> TPrimExp Half v

This expression is of type Half.

isF32 :: PrimExp v -> TPrimExp Float v

This expression is of type Float.

isF64 :: PrimExp v -> TPrimExp Double v

This expression is of type Double.

primExpSizeAtLeast :: Int -> PrimExp v -> Bool

True if the PrimExp has at least this many nodes. This can be much more efficient than comparing with length for large PrimExps, as this function is lazy.

constFoldPrimExp :: PrimExp v -> PrimExp v

Perform quick and dirty constant folding on the top level of a PrimExp. This is necessary because we want to consider e.g. equality modulo constant folding.

(.&&.) :: TPrimExp Bool v -> TPrimExp Bool v -> TPrimExp Bool v infixr 3

Lifted logical conjunction.

(.||.) :: TPrimExp Bool v -> TPrimExp Bool v -> TPrimExp Bool v infixr 2

Lifted logical conjunction.

(.<.) :: TPrimExp t v -> TPrimExp t v -> TPrimExp Bool v infix 4

Lifted relational operators; assuming signed numbers in case of integers.

(.<=.) :: TPrimExp t v -> TPrimExp t v -> TPrimExp Bool v infix 4

Lifted relational operators; assuming signed numbers in case of integers.

(.==.) :: TPrimExp t v -> TPrimExp t v -> TPrimExp Bool v infix 4

Lifted relational operators; assuming signed numbers in case of integers.

(.>.) :: TPrimExp t v -> TPrimExp t v -> TPrimExp Bool v infix 4

Lifted relational operators; assuming signed numbers in case of integers.

(.>=.) :: TPrimExp t v -> TPrimExp t v -> TPrimExp Bool v infix 4

Lifted relational operators; assuming signed numbers in case of integers.

(.&.) :: TPrimExp t v -> TPrimExp t v -> TPrimExp t v

Lifted bitwise operators. The right-shift is logical, *not* arithmetic.

(.|.) :: TPrimExp t v -> TPrimExp t v -> TPrimExp t v

Lifted bitwise operators. The right-shift is logical, *not* arithmetic.

(.^.) :: TPrimExp t v -> TPrimExp t v -> TPrimExp t v

Lifted bitwise operators. The right-shift is logical, *not* arithmetic.

(.>>.) :: TPrimExp t v -> TPrimExp t v -> TPrimExp t v

Lifted bitwise operators. The right-shift is logical, *not* arithmetic.

(.<<.) :: TPrimExp t v -> TPrimExp t v -> TPrimExp t v

Lifted bitwise operators. The right-shift is logical, *not* arithmetic.

sExt :: IntType -> PrimExp v -> PrimExp v

Untyped smart constructor for sign extension that does a bit of constant folding.

zExt :: IntType -> PrimExp v -> PrimExp v

Untyped smart constructor for zero extension that does a bit of constant folding.

evalPrimExp :: (Pretty v, MonadFail m) => (v -> m PrimValue) -> PrimExp v -> m PrimValue

Evaluate a PrimExp in the given monad. Invokes fail on type errors.

primExpType :: PrimExp v -> PrimType

The type of values returned by a PrimExp. This function returning does not imply that the PrimExp is type-correct.

coerceIntPrimExp :: IntType -> PrimExp v -> PrimExp v

If the given PrimExp is a constant of the wrong integer type, coerce it to the given integer type. This is a workaround for an issue in the Num instance.

true :: TPrimExp Bool v

Boolean-valued PrimExps.

false :: TPrimExp Bool v

Boolean-valued PrimExps.

bNot :: TPrimExp Bool v -> TPrimExp Bool v

Boolean negation smart constructor.

sMax32 :: TPrimExp Int32 v -> TPrimExp Int32 v -> TPrimExp Int32 v

SMax on 32-bit integers.

sMin32 :: TPrimExp Int32 v -> TPrimExp Int32 v -> TPrimExp Int32 v

SMin on 32-bit integers.

sMax64 :: TPrimExp Int64 v -> TPrimExp Int64 v -> TPrimExp Int64 v

SMax on 64-bit integers.

sMin64 :: TPrimExp Int64 v -> TPrimExp Int64 v -> TPrimExp Int64 v

SMin on 64-bit integers.

sExt32 :: IntExp t => TPrimExp t v -> TPrimExp Int32 v

Sign-extend to 32 bit integer.

sExt64 :: IntExp t => TPrimExp t v -> TPrimExp Int64 v

Sign-extend to 64 bit integer.

zExt32 :: IntExp t => TPrimExp t v -> TPrimExp Int32 v

Zero-extend to 32 bit integer.

zExt64 :: IntExp t => TPrimExp t v -> TPrimExp Int64 v

Zero-extend to 64 bit integer.

fMin64 :: TPrimExp Double v -> TPrimExp Double v -> TPrimExp Double v

64-bit float minimum.

fMax64 :: TPrimExp Double v -> TPrimExp Double v -> TPrimExp Double v

64-bit float maximum.

leafExpTypes :: Ord a => PrimExp a -> Set (a, PrimType)

Produce a mapping from the leaves of the PrimExp to their designated types.

(~*~) :: PrimExp v -> PrimExp v -> PrimExp v infix 7

Multiplication of untyped PrimExps, which must have the same type.

(~/~) :: PrimExp v -> PrimExp v -> PrimExp v infix 7

Division of untyped PrimExps, which must have the same type. For integers, this is unsafe signed division.

(~+~) :: PrimExp v -> PrimExp v -> PrimExp v infix 6

Addition of untyped PrimExps, which must have the same type.

(~-~) :: PrimExp v -> PrimExp v -> PrimExp v infix 6

Subtraction of untyped PrimExps, which must have the same type.

newtype Count u e

A wrapper supporting a phantom type for indicating what we are counting.

Constructors

Count

Fields

unCount :: e

primExpFromExp :: (MonadFail m, RepTypes rep) => (VName -> m (PrimExp v)) -> Exp rep -> m (PrimExp v)

Convert an expression to a PrimExp. The provided function is used to convert expressions that are not trivially PrimExps. This includes constants and variable names, which are passed as SubExps.

primExpFromSubExpM :: Applicative m => (VName -> m (PrimExp v)) -> SubExp -> m (PrimExp v)

Like primExpFromExp, but for a SubExp.

primExpFromSubExp :: PrimType -> SubExp -> PrimExp VName

Convert SubExps of a given type.

pe32 :: SubExp -> TPrimExp Int32 VName

Shorthand for constructing a TPrimExp of type Int32.

le32 :: a -> TPrimExp Int32 a

Shorthand for constructing a TPrimExp of type Int32, from a leaf.

pe64 :: SubExp -> TPrimExp Int64 VName

Shorthand for constructing a TPrimExp of type Int64.

le64 :: a -> TPrimExp Int64 a

Shorthand for constructing a TPrimExp of type Int64, from a leaf.

f32pe :: SubExp -> TPrimExp Float VName

Shorthand for constructing a TPrimExp of type Float32.

f32le :: a -> TPrimExp Float a

Shorthand for constructing a TPrimExp of type Float32, from a leaf.

f64pe :: SubExp -> TPrimExp Double VName

Shorthand for constructing a TPrimExp of type Float64.

f64le :: a -> TPrimExp Double a

Shorthand for constructing a TPrimExp of type Float64, from a leaf.

replaceInPrimExpM :: Monad m => (a -> PrimType -> m (PrimExp b)) -> PrimExp a -> m (PrimExp b)

Applying a monadic transformation to the leaves in a PrimExp.

replaceInPrimExp :: (a -> PrimType -> PrimExp b) -> PrimExp a -> PrimExp b

As replaceInPrimExpM, but in the identity monad.

substituteInPrimExp :: Ord v => Map v (PrimExp v) -> PrimExp v -> PrimExp v

Substituting names in a PrimExp with other PrimExps

primExpSlice :: Slice SubExp -> Slice (TPrimExp Int64 VName)

Convert a SubExp slice to a PrimExp slice.

subExpSlice :: MonadBuilder m => Slice (TPrimExp Int64 VName) -> m (Slice SubExp)

Convert a PrimExp slice to a SubExp slice.

data Bytes

Phantom type for a count of bytes.

data Elements

Phantom type for a count of elements.

data Arg

A function call argument.

Constructors

ExpArg Exp

MemArg VName

type TExp t = TPrimExp t VName

Like Exp, but with a required/known type.

type Exp = PrimExp VName

A side-effect free expression whose execution will produce a single primitive value.

data Volatility

The volatility of a memory access or variable. Feel free to ignore this for backends where it makes no sense (anything but C and similar low-level things)

Constructors

Volatile

Nonvolatile

pattern TracePrint :: ErrorMsg Exp -> Code a

Log the given message, *without* a trailing linebreak (unless part of the mssage).

pattern DebugPrint :: String -> Maybe Exp -> Code a

Print the given value to the screen, somehow annotated with the given string as a description. If no type/value pair, just print the string. This has no semantic meaning, but is used entirely for debugging. Code generators are free to ignore this statement.

pattern SetMem :: VName -> VName -> Space -> Code a

Must be in same space.

pattern SetScalar :: VName -> Exp -> Code a

Set a scalar variable.

pattern DeclareArray :: VName -> Space -> PrimType -> ArrayContents -> Code a

Create an array containing the given values. The lifetime of the array will be the entire application. This is mostly used for constant arrays, but also for some bookkeeping data, like the synchronisation counts used to implement reduction.

pattern DeclareScalar :: VName -> Volatility -> PrimType -> Code a

Declare a scalar variable with an initially undefined value.

pattern DeclareMem :: VName -> Space -> Code a

Declare a memory block variable that will point to memory in the given memory space. Note that this is distinct from allocation. The memory block must be the target of either an Allocate or a SetMem before it can be used for reading or writing.

pattern (:>>:)

Arguments

:: Code a
Statement composition. Crucial for the Semigroup instance.

-> Code a

-> Code a

pattern Assert :: Exp -> ErrorMsg Exp -> (SrcLoc, [SrcLoc]) -> Code a

Assert that something must be true. Should it turn out not to be true, then report a failure along with the given error message.

pattern Call :: [VName] -> Name -> [Arg] -> Code a

Function call. The results are written to the provided VName variables.

pattern Copy :: VName -> Count Bytes (TExp Int64) -> Space -> VName -> Count Bytes (TExp Int64) -> Space -> Count Bytes (TExp Int64) -> Code a

Destination, offset in destination, destination space, source, offset in source, offset space, number of bytes.

pattern Allocate :: VName -> Count Bytes (TExp Int64) -> Space -> Code a

Memory space must match the corresponding DeclareMem.

pattern Skip :: Code a

No-op. Crucial for the Monoid instance.

pattern For :: VName -> Exp -> Code a -> Code a

A for-loop iterating the given number of times. The loop parameter starts counting from zero and will have the same (integer) type as the bound. The bound is evaluated just once, before the loop is entered.

pattern While :: TExp Bool -> Code a -> Code a

While loop. The conditional is (of course) re-evaluated before every iteration of the loop.

pattern If :: TExp Bool -> Code a -> Code a -> Code a

Conditional execution.

pattern Free :: VName -> Space -> Code a

Indicate that some memory block will never again be referenced via the indicated variable. However, it may still be accessed through aliases. It is only safe to actually deallocate the memory block if this is the last reference. There is no guarantee that all memory blocks will be freed with this statement. Backends are free to ignore it entirely.

pattern Comment :: String -> Code a -> Code a

Has the same semantics as the contained code, but the comment should show up in generated code for ease of inspection.

pattern Write :: VName -> Count Elements (TExp Int64) -> PrimType -> Space -> Volatility -> Exp -> Code a

Write mem i t space vol v writes the value v to mem offset by i elements of type t. The Space argument is the memory space of mem (technically redundant, but convenient). Note that reading is done with an Exp (Read).

pattern Read :: VName -> VName -> Count Elements (TExp Int64) -> PrimType -> Space -> Volatility -> Code a

Read a scalar from memory from memory. The first VName is the target scalar variable, and the remaining arguments have the same meaning as with Write.

pattern Op :: a -> Code a

Perform an extensible operation.

data ArrayContents

The contents of a statically declared constant array. Such arrays are always unidimensional, and reshaped if necessary in the code that uses them.

Constructors

ArrayValues [PrimValue]
Precisely these values.

ArrayZeros Int
This many zeroes.

data FunctionT a

A imperative function, containing the body as well as its low-level inputs and outputs, as well as its high-level arguments and results. The latter are only used if the function is an entry point.

data ExternalValue

^ An externally visible value. This can be an opaque value (covering several physical internal values), or a single value that can be used externally. We record the uniqueness because it is important to the external interface as well.

Constructors

OpaqueValue Uniqueness String [ValueDesc]
The string is a human-readable description with no other semantics. not matter.

TransparentValue Uniqueness ValueDesc

data ValueDesc

A description of an externally meaningful value.

Constructors

ArrayValue VName Space PrimType Signedness [DimSize]
An array with memory block memory space, element type, signedness of element type (if applicable), and shape.

ScalarValue PrimType Signedness VName
A scalar value with signedness if applicable.

data Signedness

Since the core language does not care for signedness, but the source language does, entry point input/output information has metadata for integer types (and arrays containing these) that indicate whether they are really unsigned integers.

Constructors

TypeUnsigned

TypeDirect

data Constants a

A collection of imperative constants.

Constructors

Constants

Fields

constsDecl :: [Param]

The constants that are made available to the functions.

constsInit :: Code a

Setting the value of the constants. Note that this must not contain declarations of the names defined in constsDecl.

newtype Functions a

A collection of imperative functions.

Constructors

Functions [(Name, Function a)]

data Definitions a

A collection of imperative functions and constants.

Constructors

Definitions

Fields

defConsts :: Constants a

defFuns :: Functions a

data Param

An ImpCode function parameter.

Constructors

MemParam VName Space

ScalarParam VName PrimType

type DimSize = SubExp

The size of an array.

type MemSize = SubExp

The size of a memory block.

paramName :: Param -> VName

The name of a parameter.

lexicalMemoryUsage :: Function a -> Map VName Space

Find those memory blocks that are used only lexically. That is, are not used as the source or target of a SetMem, or are the result of the function, nor passed as arguments to other functions. This is interesting because such memory blocks do not need reference counting, but can be managed in a purely stack-like fashion.

We do not look inside any Ops. We assume that no Op is going to SetMem a memory block declared outside it.

calledFuncs :: Code a -> Set Name

The set of functions that are called by this code. Assumes there are no function calls in Ops.

elements :: a -> Count Elements a

This expression counts elements.

bytes :: a -> Count Bytes a

This expression counts bytes.

withElemType :: Count Elements (TExp Int64) -> PrimType -> Count Bytes (TExp Int64)

Convert a count of elements into a count of bytes, given the per-element size.

var :: VName -> PrimType -> Exp

Turn a VName into a Exp.

declaredIn :: Code a -> Names

The names declared with DeclareMem, DeclareScalar, and DeclareArray in the given code.

module Futhark.IR.GPU.Sizes