Futhark.CodeGen.ImpGen

Entry Points

compileProg :: (Mem rep inner, FreeIn op, MonadFreshNames m) => r -> Operations rep r op -> Space -> Prog rep -> m (Warnings, Definitions op)

Pluggable Compiler

type OpCompiler rep r op = Pat rep -> Op rep -> ImpM rep r op ()

How to compile an Op.

type ExpCompiler rep r op = Pat rep -> Exp rep -> ImpM rep r op ()

How to compile an Exp.

type CopyCompiler rep r op = PrimType -> MemLoc -> MemLoc -> ImpM rep r op ()

type StmsCompiler rep r op = Names -> Stms rep -> ImpM rep r op () -> ImpM rep r op ()

How to compile some Stms.

type AllocCompiler rep r op = VName -> Count Bytes (TExp Int64) -> ImpM rep r op ()

An alternate way of compiling an allocation.

data Operations rep r op

Constructors

Operations

Fields

opsExpCompiler :: ExpCompiler rep r op

opsOpCompiler :: OpCompiler rep r op

opsStmsCompiler :: StmsCompiler rep r op

opsCopyCompiler :: CopyCompiler rep r op

opsAllocCompilers :: Map Space (AllocCompiler rep r op)

defaultOperations :: (Mem rep inner, FreeIn op) => OpCompiler rep r op -> Operations rep r op

An operations set for which the expression compiler always returns defCompileExp.

data MemLoc

When an array is declared, this is where it is stored.

Constructors

MemLoc

Fields

memLocName :: VName

memLocShape :: [DimSize]

memLocIxFun :: IxFun (TExp Int64)

sliceMemLoc :: MemLoc -> Slice (TExp Int64) -> MemLoc

newtype MemEntry

Constructors

MemEntry

Fields

entryMemSpace :: Space

newtype ScalarEntry

Constructors

ScalarEntry

Fields

entryScalarType :: PrimType

Monadic Compiler Interface

data ImpM rep r op a

localDefaultSpace :: Space -> ImpM rep r op a -> ImpM rep r op a

askFunction :: ImpM rep r op (Maybe Name)

newVNameForFun :: String -> ImpM rep r op VName

Generate a VName, prefixed with askFunction if it exists.

nameForFun :: String -> ImpM rep r op Name

Generate a Name, prefixed with askFunction if it exists.

askEnv :: ImpM rep r op r

localEnv :: (r -> r) -> ImpM rep r op a -> ImpM rep r op a

localOps :: Operations rep r op -> ImpM rep r op a -> ImpM rep r op a

type VTable rep = Map VName (VarEntry rep)

The symbol table used during compilation.

getVTable :: ImpM rep r op (VTable rep)

Get the current symbol table.

localVTable :: (VTable rep -> VTable rep) -> ImpM rep r op a -> ImpM rep r op a

Run an action with a modified symbol table. All changes to the symbol table will be reverted once the action is done!

subImpM :: r' -> Operations rep r' op' -> ImpM rep r' op' a -> ImpM rep r op (a, Code op')

subImpM_ :: r' -> Operations rep r' op' -> ImpM rep r' op' a -> ImpM rep r op (Code op')

emit :: Code op -> ImpM rep r op ()

Emit some generated imperative code.

emitFunction :: Name -> Function op -> ImpM rep r op ()

Emit a function in the generated code.

hasFunction :: Name -> ImpM rep r op Bool

Check if a function of a given name exists.

collect :: ImpM rep r op () -> ImpM rep r op (Code op)

Execute a code generation action, returning the code that was emitted.

collect' :: ImpM rep r op a -> ImpM rep r op (a, Code op)

comment :: String -> ImpM rep r op () -> ImpM rep r op ()

Execute a code generation action, wrapping the generated code within a Comment with the given description.

data VarEntry rep

Every non-scalar variable must be associated with an entry.

Constructors

ArrayVar (Maybe (Exp rep)) ArrayEntry

ScalarVar (Maybe (Exp rep)) ScalarEntry

MemVar (Maybe (Exp rep)) MemEntry

AccVar (Maybe (Exp rep)) (VName, Shape, [Type])

data ArrayEntry

Constructors

ArrayEntry

Fields

entryArrayLoc :: MemLoc

entryArrayElemType :: PrimType

Lookups

lookupVar :: VName -> ImpM rep r op (VarEntry rep)

lookupArray :: VName -> ImpM rep r op ArrayEntry

lookupMemory :: VName -> ImpM rep r op MemEntry

lookupAcc :: VName -> [TExp Int64] -> ImpM rep r op (VName, Space, [VName], [TExp Int64], Maybe (Lambda rep))

In the case of a histogram-like accumulator, also sets the index parameters.

Building Blocks

data TV t

A typed variable, which we can turn into a typed expression, or use as the target for an assignment. This is used to aid in type safety when doing code generation, by keeping the types straight. It is still easy to cheat when you need to.

mkTV :: VName -> PrimType -> TV t

Create a typed variable from a name and a dynamic type. Note that there is no guarantee that the dynamic type corresponds to the inferred static type, but the latter will at least have to be used consistently.

tvSize :: TV t -> DimSize

Convert a typed variable to a size (a SubExp).

tvExp :: TV t -> TExp t

Convert a typed variable to a similarly typed expression.

tvVar :: TV t -> VName

Extract the underlying variable name from a typed variable.

class ToExp a where

Compile things to Exp.

Minimal complete definition

toExp, toExp'

Methods

toExp :: a -> ImpM rep r op Exp

Compile to an Exp, where the type (must must still be a primitive) is deduced monadically.

toExp' :: PrimType -> a -> Exp

Compile where we know the type in advance.

toInt64Exp :: a -> TExp Int64

toBoolExp :: a -> TExp Bool

compileAlloc :: Mem rep inner => Pat rep -> SubExp -> Space -> ImpM rep r op ()

compileAlloc pat size space allocates n bytes of memory in space, writing the result to dest, which must be a single MemoryDestination,

everythingVolatile :: ImpM rep r op a -> ImpM rep r op a

compileBody :: Pat rep -> Body rep -> ImpM rep r op ()

compileBody' :: [Param dec] -> Body rep -> ImpM rep r op ()

compileLoopBody :: Typed dec => [Param dec] -> Body rep -> ImpM rep r op ()

defCompileStms :: (Mem rep inner, FreeIn op) => Names -> Stms rep -> ImpM rep r op () -> ImpM rep r op ()

compileStms :: Names -> Stms rep -> ImpM rep r op () -> ImpM rep r op ()

compileExp :: Pat rep -> Exp rep -> ImpM rep r op ()

defCompileExp :: Mem rep inner => Pat rep -> Exp rep -> ImpM rep r op ()

fullyIndexArray :: VName -> [TExp Int64] -> ImpM rep r op (VName, Space, Count Elements (TExp Int64))

fullyIndexArray' :: MemLoc -> [TExp Int64] -> ImpM rep r op (VName, Space, Count Elements (TExp Int64))

copy :: CopyCompiler rep r op

copyDWIM :: VName -> [DimIndex (TExp Int64)] -> SubExp -> [DimIndex (TExp Int64)] -> ImpM rep r op ()

Copy from here to there; both destination and source be indexeded. If so, they better be arrays of enough dimensions. This function will generally just Do What I Mean, and Do The Right Thing. Both destination and source must be in scope.

copyDWIMFix :: VName -> [TExp Int64] -> SubExp -> [TExp Int64] -> ImpM rep r op ()

As copyDWIM, but implicitly DimFixes the indexes.

copyElementWise :: CopyCompiler rep r op

typeSize :: Type -> Count Bytes (TExp Int64)

The number of bytes needed to represent the array in a straightforward contiguous format, as an Int64 expression.

inBounds :: Slice (TExp Int64) -> [TExp Int64] -> TExp Bool

Is this indexing in-bounds for an array of the given shape? This is useful for things like scatter, which ignores out-of-bounds writes.

isMapTransposeCopy :: PrimType -> MemLoc -> MemLoc -> Maybe (TExp Int64, TExp Int64, TExp Int64, TExp Int64, TExp Int64)

Is this copy really a mapping with transpose?

Constructing code.

dLParams :: Mem rep inner => [LParam rep] -> ImpM rep r op ()

dFParams :: Mem rep inner => [FParam rep] -> ImpM rep r op ()

dScope :: Mem rep inner => Maybe (Exp rep) -> Scope rep -> ImpM rep r op ()

dArray :: VName -> PrimType -> ShapeBase SubExp -> VName -> IxFun -> ImpM rep r op ()

dPrim :: String -> PrimType -> ImpM rep r op (TV t)

The return type is polymorphic, so there is no guarantee it actually matches the PrimType, but at least we have to use it consistently.

dPrimVol :: String -> PrimType -> TExp t -> ImpM rep r op (TV t)

dPrim_ :: VName -> PrimType -> ImpM rep r op ()

dPrimV_ :: VName -> TExp t -> ImpM rep r op ()

dPrimV :: String -> TExp t -> ImpM rep r op (TV t)

dPrimVE :: String -> TExp t -> ImpM rep r op (TExp t)

dIndexSpace :: [(VName, TExp Int64)] -> TExp Int64 -> ImpM rep r op ()

dIndexSpace f dims i computes a list of indices into an array with dimension dims given the flat index i. The resulting list will have the same size as dims. Intermediate results are passed to f.

dIndexSpace' :: String -> [TExp Int64] -> TExp Int64 -> ImpM rep r op [TExp Int64]

Like dIndexSpace, but invent some new names for the indexes based on the given template.

sFor :: String -> TExp t -> (TExp t -> ImpM rep r op ()) -> ImpM rep r op ()

sWhile :: TExp Bool -> ImpM rep r op () -> ImpM rep r op ()

sComment :: String -> ImpM rep r op () -> ImpM rep r op ()

sIf :: TExp Bool -> ImpM rep r op () -> ImpM rep r op () -> ImpM rep r op ()

sWhen :: TExp Bool -> ImpM rep r op () -> ImpM rep r op ()

sUnless :: TExp Bool -> ImpM rep r op () -> ImpM rep r op ()

sOp :: op -> ImpM rep r op ()

sDeclareMem :: String -> Space -> ImpM rep r op VName

sAlloc :: String -> Count Bytes (TExp Int64) -> Space -> ImpM rep r op VName

sAlloc_ :: VName -> Count Bytes (TExp Int64) -> Space -> ImpM rep r op ()

sArray :: String -> PrimType -> ShapeBase SubExp -> VName -> IxFun -> ImpM rep r op VName

sArrayInMem :: String -> PrimType -> ShapeBase SubExp -> VName -> ImpM rep r op VName

Declare an array in row-major order in the given memory block.

sAllocArray :: String -> PrimType -> ShapeBase SubExp -> Space -> ImpM rep r op VName

Uses linear/iota index function.

sAllocArrayPerm :: String -> PrimType -> ShapeBase SubExp -> Space -> [Int] -> ImpM rep r op VName

Like sAllocArray, but permute the in-memory representation of the indices as specified.

sStaticArray :: String -> Space -> PrimType -> ArrayContents -> ImpM rep r op VName

Uses linear/iota index function.

sWrite :: VName -> [TExp Int64] -> Exp -> ImpM rep r op ()

sUpdate :: VName -> Slice (TExp Int64) -> SubExp -> ImpM rep r op ()

sLoopNest :: Shape -> ([TExp Int64] -> ImpM rep r op ()) -> ImpM rep r op ()

(<--) :: TV t -> TExp t -> ImpM rep r op () infixl 3

Typed assignment.

(<~~) :: VName -> Exp -> ImpM rep r op () infixl 3

Untyped assignment.

function :: Name -> [Param] -> [Param] -> ImpM rep r op () -> ImpM rep r op ()

Constructing an ad-hoc function that does not correspond to any of the IR functions in the input program.

warn :: Located loc => loc -> [loc] -> String -> ImpM rep r op ()

Emit a warning about something the user should be aware of.

module Language.Futhark.Warnings