Futhark.CodeGen.ImpGen.Multicore.Base

Documentation

extractAllocations :: Code -> (Code, Code)

Try to extract invariant allocations. If we assume that the given Code is the body of a SegOp, then it is always safe to move the immediate allocations to the prebody.

compileThreadResult :: SegSpace -> PatElem MCMem -> KernelResult -> MulticoreGen ()

data Locks

Information about the locks available for accumulators.

Constructors

Locks

Fields

locksArray :: VName

locksCount :: Int

data HostEnv

Constructors

HostEnv

Fields

hostAtomics :: AtomicBinOp

hostLocks :: Map VName Locks

type AtomicBinOp = BinOp -> Maybe (VName -> VName -> Count Elements (TExp Int32) -> Exp -> AtomicOp)

Is there an atomic BinOp corresponding to this BinOp?

type MulticoreGen = ImpM MCMem HostEnv Multicore

decideScheduling :: Code -> Scheduling

decideScheduling' :: SegOp () rep -> Code -> Scheduling

groupResultArrays :: String -> SubExp -> [SegBinOp MCMem] -> MulticoreGen [[VName]]

Arrays for storing group results shared between threads

renameSegBinOp :: [SegBinOp MCMem] -> MulticoreGen [SegBinOp MCMem]

freeParams :: Code -> [VName] -> MulticoreGen [Param]

renameHistOpLambda :: [HistOp MCMem] -> MulticoreGen [HistOp MCMem]

atomicUpdateLocking :: AtomicBinOp -> Lambda MCMem -> AtomicUpdate MCMem ()

data AtomicUpdate rep r

The mechanism that will be used for performing the atomic update. Approximates how efficient it will be. Ordered from most to least efficient.

Constructors

AtomicPrim (DoAtomicUpdate rep r)

AtomicCAS (DoAtomicUpdate rep r)
Can be done by efficient swaps.

AtomicLocking (Locking -> DoAtomicUpdate rep r)
Requires explicit locking.

data Locking

Locking strategy used for an atomic update.

Constructors

Locking

Fields

lockingArray :: VName

Array containing the lock.

lockingIsUnlocked :: TExp Int32

Value for us to consider the lock free.

lockingToLock :: TExp Int32

What to write when we lock it.

lockingToUnlock :: TExp Int32

What to write when we unlock it.

lockingMapping :: [TExp Int64] -> [TExp Int64]

A transformation from the logical lock index to the physical position in the array. This can also be used to make the lock array smaller.

getSpace :: SegOp () MCMem -> SegSpace

getIterationDomain :: SegOp () MCMem -> SegSpace -> MulticoreGen (TExp Int64)

getReturnParams :: Pat MCMem -> SegOp () MCMem -> MulticoreGen [Param]

segOpString :: SegOp () MCMem -> MulticoreGen String