Futhark.CodeGen.ImpGen.Transpose

Description

A cache-oblivious sequential transposition for CPU execution. Generates a recursive function.

Documentation

mapTransposeFunction :: Name -> PrimType -> Function op

We need to know the name of the function we are generating, as this function is recursive.

transposeArgs :: PrimType -> VName -> Count Bytes (TExp Int64) -> VName -> Count Bytes (TExp Int64) -> TExp Int64 -> TExp Int64 -> TExp Int64 -> [Arg]

Take well-typed arguments to the transpose function and produce the actual argument list.