Futhark.CodeGen.OpenCL.Heuristics

Description

Some OpenCL platforms have a SIMDwarpwavefront-based execution model that execute groups of threads in lockstep, permitting us to perform cross-thread synchronisation within each such group without the use of barriers. Unfortunately, there seems to be no reliable way to query these sizes at runtime. Instead, we use builtin tables to figure out which size we should use for a specific platform and device. If nothing matches here, the wave size should be set to one.

We also use this to select reasonable default group sizes and group counts.

Documentation

data SizeHeuristic

A heuristic for setting the default value for something.

Constructors

SizeHeuristic

Fields

platformName :: String

deviceType :: DeviceType

heuristicSize :: WhichSize

heuristicValue :: TPrimExp Int32 DeviceInfo

data DeviceType

The type of OpenCL device that this heuristic applies to.

Constructors

DeviceCPU

DeviceGPU

data WhichSize

A size that can be assigned a default.

Constructors

LockstepWidth

NumGroups

GroupSize

TileSize

RegTileSize

Threshold

newtype DeviceInfo

The value supplies by a heuristic can depend on some device information. This will be translated into a call to clGetDeviceInfo(). Make sure to only request info that can be casted to a scalar type.

Constructors

DeviceInfo String

sizeHeuristicsTable :: [SizeHeuristic]

All of our heuristics.