Futhark.CodeGen.SetDefaultSpace

Description

Change DefaultSpace in a program to some other memory space. This is needed because the GPU backends use DefaultSpace to refer to GPU memory for most of the pipeline, but final code generation assumes that DefaultSpace is CPU memory.

Documentation

setDefaultSpace :: Space -> Definitions op -> Definitions op

Set all uses of DefaultSpace in the given definitions to another memory space.

setDefaultCodeSpace :: Space -> Code op -> Code op

Like setDefaultSpace, but for Code.