Futhark.Compiler

Description

High-level API for invoking the Futhark compiler.

Documentation

runPipelineOnProgram :: FutharkConfig -> Pipeline SOACS torep -> FilePath -> FutharkM (Prog torep)

Read a program from the given FilePath, run the given Pipeline, and return it.

runCompilerOnProgram :: FutharkConfig -> Pipeline SOACS rep -> Action rep -> FilePath -> IO ()

Read a program from the given FilePath, run the given Pipeline, and finish up with the given Action.

dumpError :: FutharkConfig -> CompilerError -> IO ()

Print a compiler error to stdout. The FutharkConfig controls to which degree auxiliary information (e.g. the failing program) is also printed.

handleWarnings :: FutharkConfig -> FutharkM (Warnings, a) -> FutharkM a

Run an operation that produces warnings, and handle them appropriately, yielding the non-warning return value. "Proper handling" means e.g. to print them to the screen, as directed by the compiler configuration.

module Futhark.Compiler.Program

module Futhark.Compiler.Config

readProgram :: (MonadError CompilerError m, MonadIO m) => [Name] -> FilePath -> m (Warnings, Imports, VNameSource)

Read and type-check a Futhark program, including all imports.

readProgramOrDie :: MonadIO m => FilePath -> m (Warnings, Imports, VNameSource)

Not verbose, and terminates process on error.

readUntypedProgram :: (MonadError CompilerError m, MonadIO m) => FilePath -> m [(String, UncheckedProg)]

Read and parse (but do not type-check) a Futhark program, including all imports.

readUntypedProgramOrDie :: MonadIO m => FilePath -> m [(String, UncheckedProg)]

Not verbose, and terminates process on error.