Futhark.Compiler.Config

Documentation

data FutharkConfig

The compiler configuration. This only contains options related to core compiler functionality, such as reading the initial program and running passes. Options related to code generation are handled elsewhere.

Constructors

FutharkConfig

Fields

futharkVerbose :: (Verbosity, Maybe FilePath)

futharkWarn :: Bool

Warn if True.

futharkWerror :: Bool

If true, error on any warnings.

futharkSafe :: Bool

If True, ignore unsafe.

futharkEntryPoints :: [Name]

Additional functions that should be exposed as entry points.

futharkTypeCheck :: Bool

If false, disable type-checking

newFutharkConfig :: FutharkConfig

The default compiler configuration.

data Verbosity

How much information to print to stderr while the compiler is running.

Constructors

NotVerbose
Silence is golden.

Verbose
Print messages about which pass is running.

VeryVerbose
Also print logs from individual passes.