Futhark.Construct

Description

Constructing Futhark ASTs

This module re-exports and defines a bunch of building blocks for constructing fragments of Futhark ASTs. More importantly, it also contains a basic introduction on how to use them.

The Futhark.IR.Syntax module contains the core AST definition. One important invariant is that all bound names in a Futhark program must be globally unique. In principle, you could use the facilities from Futhark.MonadFreshNames (or your own bespoke source of unique names) to manually construct expressions, statements, and entire ASTs. In practice, this would be very tedious. Instead, we have defined a collection of building blocks (centered around the MonadBuilder type class) that permits a more abstract way of generating code.

Constructing ASTs with these building blocks requires you to ensure that all free variables are in scope. See Futhark.IR.Prop.Scope.

MonadBuilder

A monad that implements MonadBuilder tracks the statements added so far, the current names in scope, and allows you to add additional statements with addStm. Any monad that implements MonadBuilder also implements the Rep type family, which indicates which rep it works with. Inside a MonadBuilder we can use collectStms to gather up the Stms added with addStm in some nested computation.

The BuilderT monad (and its convenient Builder version) provides the simplest implementation of MonadBuilder.

Higher-level building blocks

On top of the raw facilities provided by MonadBuilder, we have more convenient facilities. For example, letSubExp lets us conveniently create a Stm for an Exp that produces a single value, and returns the (fresh) name for the resulting variable:

z <- letExp "z" $ BasicOp $ BinOp (Add Int32) (Var x) (Var y)

Examples

The Futhark.Transform.FirstOrderTransform module is a (relatively) simple example of how to use these components. As are some of the high-level building blocks in this very module.

Documentation

letSubExp :: MonadBuilder m => String -> Exp (Rep m) -> m SubExp

letSubExps :: MonadBuilder m => String -> [Exp (Rep m)] -> m [SubExp]

letExp :: MonadBuilder m => String -> Exp (Rep m) -> m VName

letTupExp :: MonadBuilder m => String -> Exp (Rep m) -> m [VName]

letTupExp' :: MonadBuilder m => String -> Exp (Rep m) -> m [SubExp]

letInPlace :: MonadBuilder m => String -> VName -> Slice SubExp -> Exp (Rep m) -> m VName

eSubExp :: MonadBuilder m => SubExp -> m (Exp (Rep m))

eParam :: MonadBuilder m => Param t -> m (Exp (Rep m))

eIf :: (MonadBuilder m, BranchType (Rep m) ~ ExtType) => m (Exp (Rep m)) -> m (Body (Rep m)) -> m (Body (Rep m)) -> m (Exp (Rep m))

eIf' :: (MonadBuilder m, BranchType (Rep m) ~ ExtType) => m (Exp (Rep m)) -> m (Body (Rep m)) -> m (Body (Rep m)) -> IfSort -> m (Exp (Rep m))

As eIf, but an IfSort can be given.

eBinOp :: MonadBuilder m => BinOp -> m (Exp (Rep m)) -> m (Exp (Rep m)) -> m (Exp (Rep m))

eCmpOp :: MonadBuilder m => CmpOp -> m (Exp (Rep m)) -> m (Exp (Rep m)) -> m (Exp (Rep m))

eConvOp :: MonadBuilder m => ConvOp -> m (Exp (Rep m)) -> m (Exp (Rep m))

eSignum :: MonadBuilder m => m (Exp (Rep m)) -> m (Exp (Rep m))

eCopy :: MonadBuilder m => m (Exp (Rep m)) -> m (Exp (Rep m))

eBody :: MonadBuilder m => [m (Exp (Rep m))] -> m (Body (Rep m))

eLambda :: MonadBuilder m => Lambda (Rep m) -> [m (Exp (Rep m))] -> m [SubExpRes]

eRoundToMultipleOf :: MonadBuilder m => IntType -> m (Exp (Rep m)) -> m (Exp (Rep m)) -> m (Exp (Rep m))

eSliceArray :: MonadBuilder m => Int -> VName -> m (Exp (Rep m)) -> m (Exp (Rep m)) -> m (Exp (Rep m))

Construct an Index expressions that slices an array with unit stride.

eBlank :: MonadBuilder m => Type -> m (Exp (Rep m))

Construct an unspecified value of the given type.

eAll :: MonadBuilder m => [SubExp] -> m (Exp (Rep m))

True if all operands are true.

eDimInBounds :: MonadBuilder m => m (Exp (Rep m)) -> m (Exp (Rep m)) -> m (Exp (Rep m))

eInBoundsForDim w i produces 0 <= i < w.

eOutOfBounds :: MonadBuilder m => VName -> [m (Exp (Rep m))] -> m (Exp (Rep m))

Are these indexes out-of-bounds for the array?

asIntZ :: MonadBuilder m => IntType -> SubExp -> m SubExp

Zero-extend to the given integer type.

asIntS :: MonadBuilder m => IntType -> SubExp -> m SubExp

Sign-extend to the given integer type.

resultBody :: Buildable rep => [SubExp] -> Body rep

Conveniently construct a body that contains no bindings.

resultBodyM :: MonadBuilder m => [SubExp] -> m (Body (Rep m))

Conveniently construct a body that contains no bindings - but this time, monadically!

insertStmsM :: MonadBuilder m => m (Body (Rep m)) -> m (Body (Rep m))

Evaluate the action, producing a body, then wrap it in all the bindings it created using addStm.

buildBody :: MonadBuilder m => m (Result, a) -> m (Body (Rep m), a)

Evaluate an action that produces a Result and an auxiliary value, then return the body constructed from the Result and any statements added during the action, along the auxiliary value.

buildBody_ :: MonadBuilder m => m Result -> m (Body (Rep m))

As buildBody, but there is no auxiliary value.

mapResult :: Buildable rep => (Result -> Body rep) -> Body rep -> Body rep

Change that result where evaluation of the body would stop. Also change type annotations at branches.

foldBinOp :: MonadBuilder m => BinOp -> SubExp -> [SubExp] -> m (Exp (Rep m))

Apply a binary operator to several subexpressions. A left-fold.

binOpLambda :: (MonadBuilder m, Buildable (Rep m)) => BinOp -> PrimType -> m (Lambda (Rep m))

Create a two-parameter lambda whose body applies the given binary operation to its arguments. It is assumed that both argument and result types are the same. (This assumption should be fixed at some point.)

cmpOpLambda :: (MonadBuilder m, Buildable (Rep m)) => CmpOp -> m (Lambda (Rep m))

As binOpLambda, but for CmpOps.

mkLambda :: MonadBuilder m => [LParam (Rep m)] -> m Result -> m (Lambda (Rep m))

Easily construct a Lambda within a MonadBuilder.

sliceDim :: SubExp -> DimIndex SubExp

Slice a full dimension of the given size.

fullSlice :: Type -> [DimIndex SubExp] -> Slice SubExp

fullSlice t slice returns slice, but with DimSlices of entire dimensions appended to the full dimensionality of t. This function is used to turn incomplete indexing complete, as required by Index.

fullSliceNum :: Num d => [d] -> [DimIndex d] -> Slice d

Like fullSlice, but the dimensions are simply numeric.

isFullSlice :: Shape -> Slice SubExp -> Bool

Does the slice describe the full size of the array? The most obvious such slice is one that DimSlices the full span of every dimension, but also one that fixes all unit dimensions.

sliceAt :: Type -> Int -> [DimIndex SubExp] -> Slice SubExp

sliceAt t n slice returns slice but with DimSlices of the outer n dimensions prepended, and as many appended as to make it a full slice. This is a generalisation of fullSlice.

ifCommon :: [Type] -> IfDec ExtType

module Futhark.Builder

Result types

instantiateShapes :: Monad m => (Int -> m SubExp) -> [TypeBase ExtShape u] -> m [TypeBase Shape u]

Instantiate all existential parts dimensions of the given type, using a monadic action to create the necessary SubExps. You should call this function within some monad that allows you to collect the actions performed (say, Writer).

instantiateShapes' :: [VName] -> [TypeBase ExtShape u] -> [TypeBase Shape u]

removeExistentials :: ExtType -> Type -> Type

Convenience

simpleMkLetNames :: (ExpDec rep ~ (), LetDec rep ~ Type, MonadFreshNames m, TypedOp (Op rep), HasScope rep m) => [VName] -> Exp rep -> m (Stm rep)

Can be used as the definition of mkLetNames for a Buildable instance for simple representations.

class ToExp a where

Instances of this class can be converted to Futhark expressions within a MonadBuilder.

Methods

toExp :: MonadBuilder m => a -> m (Exp (Rep m))

toSubExp :: (MonadBuilder m, ToExp a) => String -> a -> m SubExp

A convenient composition of letSubExp and toExp.