Futhark.Error

Description

Futhark error definitions.

Documentation

data CompilerError

A compiler error.

Constructors

ExternalError Doc
An error that happened due to something the user did, such as provide incorrect code or options.

InternalError Text Text ErrorClass
An internal compiler error. The second text is extra data for debugging, which can be written to a file.

data ErrorClass

There are two classes of internal errors: actual bugs, and implementation limitations. The latter are already known and need not be reported.

Constructors

CompilerBug

CompilerLimitation

externalError :: MonadError CompilerError m => Doc -> m a

Raise an ExternalError based on a prettyprinting result.

externalErrorS :: MonadError CompilerError m => String -> m a

Raise an ExternalError based on a string.

data InternalError

An error that is not the users fault, but a bug (or limitation) in the compiler. Compiler passes should only ever report this error - any problems after the type checker are *our* fault, not the users. These are generally thrown as IO exceptions, and caught at the top level.

Constructors

Error ErrorClass Text

compilerBug :: Text -> a

Throw an InternalError that is a CompilerBug.

compilerBugS :: String -> a

Like compilerBug, but with a String.

compilerLimitation :: Text -> a

Throw an InternalError that is a CompilerLimitation.

compilerLimitationS :: String -> a

Like compilerLimitation, but with a String.

internalErrorS :: MonadError CompilerError m => String -> Doc -> m a

Raise an InternalError based on a prettyprinting result.