Futhark.IR.Aliases

Description

A representation where all bindings are annotated with aliasing information.

The representation definition

data Aliases rep

The rep for the basic representation.

newtype AliasDec

A wrapper around AliasDec to get around the fact that we need an Ord instance, which 'AliasDec does not have.

Constructors

AliasDec

Fields

unAliases :: Names

type VarAliases = AliasDec

The aliases of the let-bound variable.

type ConsumedInExp = AliasDec

Everything consumed in the expression.

type BodyAliasing = ([VarAliases], ConsumedInExp)

The aliases of what is returned by the Body, and what is consumed inside of it.

module Futhark.IR.Prop.Aliases

Module re-exports

module Futhark.IR.Prop

module Futhark.IR.Traversals

module Futhark.IR.Pretty

module Futhark.IR.Syntax

Adding aliases

addAliasesToPat :: (ASTRep rep, CanBeAliased (Op rep), Typed dec) => PatT dec -> Exp (Aliases rep) -> PatT (VarAliases, dec)

mkAliasedLetStm :: (ASTRep rep, CanBeAliased (Op rep)) => Pat rep -> StmAux (ExpDec rep) -> Exp (Aliases rep) -> Stm (Aliases rep)

mkAliasedBody :: (ASTRep rep, CanBeAliased (Op rep)) => BodyDec rep -> Stms (Aliases rep) -> Result -> Body (Aliases rep)

mkPatAliases :: (Aliased rep, Typed dec) => PatT dec -> Exp rep -> [PatElemT (VarAliases, dec)]

mkBodyAliases :: Aliased rep => Stms rep -> Result -> BodyAliasing

Removing aliases

removeProgAliases :: CanBeAliased (Op rep) => Prog (Aliases rep) -> Prog rep

removeFunDefAliases :: CanBeAliased (Op rep) => FunDef (Aliases rep) -> FunDef rep

removeExpAliases :: CanBeAliased (Op rep) => Exp (Aliases rep) -> Exp rep

removeStmAliases :: CanBeAliased (Op rep) => Stm (Aliases rep) -> Stm rep

removeLambdaAliases :: CanBeAliased (Op rep) => Lambda (Aliases rep) -> Lambda rep

removePatAliases :: PatT (AliasDec, a) -> PatT a

removeScopeAliases :: Scope (Aliases rep) -> Scope rep

Tracking aliases

type AliasesAndConsumed = (Map VName Names, Names)

trackAliases :: Aliased rep => AliasesAndConsumed -> Stm rep -> AliasesAndConsumed

mkStmsAliases :: Aliased rep => Stms rep -> Result -> ([Names], Names)

The aliases of the result and everything consumed in the given statements.

consumedInStms :: Aliased rep => Stms rep -> Names

The variables consumed in these statements.