Futhark.IR.GPU.Op

Size operations

data SizeOp

A simple size-level query or computation.

Constructors

SplitSpace SplitOrdering SubExp SubExp SubExp
SplitSpace o w i elems_per_thread.

Computes how to divide array elements to threads in a kernel. Returns the number of elements in the chunk that the current thread should take.

w is the length of the outer dimension in the array. i is the current thread index. Each thread takes at most elems_per_thread elements.

If the order o is SplitContiguous, thread with index i should receive elements i*elems_per_tread, i*elems_per_thread + 1, ..., i*elems_per_thread + (elems_per_thread-1).

If the order o is SplitStrided stride, the thread will receive elements i, i+stride, i+2*stride, ..., i+(elems_per_thread-1)*stride.

GetSize Name SizeClass
Produce some runtime-configurable size.

GetSizeMax SizeClass
The maximum size of some class.

CmpSizeLe Name SizeClass SubExp
Compare size (likely a threshold) with some integer value.

CalcNumGroups SubExp Name SubExp
CalcNumGroups w max_num_groups group_size calculates the number of GPU workgroups to use for an input of the given size. The Name is a size name. Note that w is an i64 to avoid overflow issues.

Host operations

data HostOp rep op

A host-level operation; parameterised by what else it can do.

Constructors

SegOp (SegOp SegLevel rep)
A segmented operation.

SizeOp SizeOp

OtherOp op

traverseHostOpStms :: Monad m => OpStmsTraverser m op rep -> OpStmsTraverser m (HostOp rep op) rep

A helper for defining TraverseOpStms.

typeCheckHostOp :: Checkable rep => (SegLevel -> OpWithAliases (Op rep) -> TypeM rep ()) -> Maybe SegLevel -> (op -> TypeM rep ()) -> HostOp (Aliases rep) op -> TypeM rep ()

SegOp refinements

data SegLevel

At which level the *body* of a SegOp executes.

Constructors

SegThread

Fields

segNumGroups :: Count NumGroups SubExp

segGroupSize :: Count GroupSize SubExp

segVirt :: SegVirt

SegGroup

Fields

segNumGroups :: Count NumGroups SubExp

segGroupSize :: Count GroupSize SubExp

segVirt :: SegVirt

Reexports

module Futhark.IR.GPU.Sizes

module Futhark.IR.SegOp