Futhark.IR.GPU.Sizes

Description

In the context of this module, a "size" is any kind of tunable (run-time) constant.

Documentation

data SizeClass

The class of some kind of configurable size. Each class may impose constraints on the valid values.

Constructors

SizeThreshold KernelPath (Maybe Int64)
A threshold with an optional default.

SizeGroup

SizeNumGroups

SizeTile

SizeRegTile

SizeLocalMemory
Likely not useful on its own, but querying the maximum can be handy.

SizeBespoke Name Int64
A bespoke size with a default.

sizeDefault :: SizeClass -> Maybe Int64

The default value for the size. If Nothing, that means the backend gets to decide.

type KernelPath = [(Name, Bool)]

An indication of which comparisons have been performed to get to this point, as well as the result of each comparison.

newtype Count u e

A wrapper supporting a phantom type for indicating what we are counting.

Constructors

Count

Fields

unCount :: e

data NumGroups

Phantom type for the number of groups of some kernel.

data GroupSize

Phantom type for the group size of some kernel.

data NumThreads

Phantom type for number of threads.