Futhark.IR.Pretty

Description

Futhark prettyprinter. This module defines Pretty instances for the AST defined in Futhark.IR.Syntax, but also a number of convenience functions if you don't want to use the interface from Pretty.

Documentation

prettyTuple :: Pretty a => [a] -> String

Prettyprint a list enclosed in curly braces.

prettyTupleLines :: Pretty a => [a] -> String

Like prettyTuple, but put a linebreak after every element.

pretty :: Pretty a => a -> String

Prettyprint a value, wrapped to 80 characters.

class (RepTypes rep, Pretty (RetType rep), Pretty (BranchType rep), Pretty (FParamInfo rep), Pretty (LParamInfo rep), Pretty (LetDec rep), Pretty (Op rep)) => PrettyRep rep where

The class of representations whose annotations can be prettyprinted.

Minimal complete definition

Nothing

Methods

ppExpDec :: ExpDec rep -> Exp rep -> Maybe Doc

ppTuple' :: Pretty a => [a] -> Doc

Like prettyTuple, but produces a Doc.

Orphan instances

Pretty VName

Instance details

Methods

ppr :: VName -> Doc

pprPrec :: Int -> VName -> Doc

pprList :: [VName] -> Doc

Pretty Attrs

Instance details

Methods

ppr :: Attrs -> Doc

pprPrec :: Int -> Attrs -> Doc

pprList :: [Attrs] -> Doc

Pretty Attr

Instance details

Methods

ppr :: Attr -> Doc

pprPrec :: Int -> Attr -> Doc

pprList :: [Attr] -> Doc

Pretty SubExp

Instance details

Methods

ppr :: SubExp -> Doc

pprPrec :: Int -> SubExp -> Doc

pprList :: [SubExp] -> Doc

Pretty Certs

Instance details

Methods

ppr :: Certs -> Doc

pprPrec :: Int -> Certs -> Doc

pprList :: [Certs] -> Doc

Pretty Ident

Instance details

Methods

ppr :: Ident -> Doc

pprPrec :: Int -> Ident -> Doc

pprList :: [Ident] -> Doc

Pretty NoUniqueness

Instance details

Methods

ppr :: NoUniqueness -> Doc

pprPrec :: Int -> NoUniqueness -> Doc

pprList :: [NoUniqueness] -> Doc

Pretty Space

Instance details

Methods

ppr :: Space -> Doc

pprPrec :: Int -> Space -> Doc

pprList :: [Space] -> Doc

Pretty ExtShape

Instance details

Methods

ppr :: ExtShape -> Doc

pprPrec :: Int -> ExtShape -> Doc

pprList :: [ExtShape] -> Doc

Pretty Shape

Instance details

Methods

ppr :: Shape -> Doc

pprPrec :: Int -> Shape -> Doc

pprList :: [Shape] -> Doc

Pretty Commutativity

Instance details

Methods

ppr :: Commutativity -> Doc

pprPrec :: Int -> Commutativity -> Doc

pprList :: [Commutativity] -> Doc

Pretty EntryParam

Instance details

Methods

ppr :: EntryParam -> Doc

pprPrec :: Int -> EntryParam -> Doc

pprList :: [EntryParam] -> Doc

Pretty EntryPointType

Instance details

Methods

ppr :: EntryPointType -> Doc

pprPrec :: Int -> EntryPointType -> Doc

pprList :: [EntryPointType] -> Doc

Pretty BasicOp

Instance details

Methods

ppr :: BasicOp -> Doc

pprPrec :: Int -> BasicOp -> Doc

pprList :: [BasicOp] -> Doc

Pretty SubExpRes

Instance details

Methods

ppr :: SubExpRes -> Doc

pprPrec :: Int -> SubExpRes -> Doc

pprList :: [SubExpRes] -> Doc

Pretty a => Pretty (ErrorMsg a)

Instance details

Methods

ppr :: ErrorMsg a -> Doc

pprPrec :: Int -> ErrorMsg a -> Doc

pprList :: [ErrorMsg a] -> Doc

Pretty t => Pretty (PatElemT t)

Instance details

Methods

ppr :: PatElemT t -> Doc

pprPrec :: Int -> PatElemT t -> Doc

pprList :: [PatElemT t] -> Doc

Pretty a => Pretty (FlatSlice a)

Instance details

Methods

ppr :: FlatSlice a -> Doc

pprPrec :: Int -> FlatSlice a -> Doc

pprList :: [FlatSlice a] -> Doc

Pretty d => Pretty (FlatDimIndex d)

Instance details

Methods

ppr :: FlatDimIndex d -> Doc

pprPrec :: Int -> FlatDimIndex d -> Doc

pprList :: [FlatDimIndex d] -> Doc

Pretty a => Pretty (Slice a)

Instance details

Methods

ppr :: Slice a -> Doc

pprPrec :: Int -> Slice a -> Doc

pprList :: [Slice a] -> Doc

Pretty d => Pretty (DimIndex d)

Instance details

Methods

ppr :: DimIndex d -> Doc

pprPrec :: Int -> DimIndex d -> Doc

pprList :: [DimIndex d] -> Doc

Pretty t => Pretty (Param t)

Instance details

Methods

ppr :: Param t -> Doc

pprPrec :: Int -> Param t -> Doc

pprList :: [Param t] -> Doc

Pretty a => Pretty (Ext a)

Instance details

Methods

ppr :: Ext a -> Doc

pprPrec :: Int -> Ext a -> Doc

pprList :: [Ext a] -> Doc

PrettyRep rep => Pretty (Prog rep)

Instance details

Methods

ppr :: Prog rep -> Doc

pprPrec :: Int -> Prog rep -> Doc

pprList :: [Prog rep] -> Doc

PrettyRep rep => Pretty (FunDef rep)

Instance details

Methods

ppr :: FunDef rep -> Doc

pprPrec :: Int -> FunDef rep -> Doc

pprList :: [FunDef rep] -> Doc

PrettyRep rep => Pretty (Lambda rep)

Instance details

Methods

ppr :: Lambda rep -> Doc

pprPrec :: Int -> Lambda rep -> Doc

pprList :: [Lambda rep] -> Doc

PrettyRep rep => Pretty (Exp rep)

Instance details

Methods

ppr :: Exp rep -> Doc

pprPrec :: Int -> Exp rep -> Doc

pprList :: [Exp rep] -> Doc

Pretty d => Pretty (DimChange d)

Instance details

Methods

ppr :: DimChange d -> Doc

pprPrec :: Int -> DimChange d -> Doc

pprList :: [DimChange d] -> Doc

PrettyRep rep => Pretty (Body rep)

Instance details

Methods

ppr :: Body rep -> Doc

pprPrec :: Int -> Body rep -> Doc

pprList :: [Body rep] -> Doc

PrettyRep rep => Pretty (Stms rep)

Instance details

Methods

ppr :: Stms rep -> Doc

pprPrec :: Int -> Stms rep -> Doc

pprList :: [Stms rep] -> Doc

PrettyRep rep => Pretty (Stm rep)

Instance details

Methods

ppr :: Stm rep -> Doc

pprPrec :: Int -> Stm rep -> Doc

pprList :: [Stm rep] -> Doc

Pretty (PatElemT dec) => Pretty (PatT dec)

Instance details

Methods

ppr :: PatT dec -> Doc

pprPrec :: Int -> PatT dec -> Doc

pprList :: [PatT dec] -> Doc

Pretty u => Pretty (TypeBase Rank u)

Instance details

Methods

ppr :: TypeBase Rank u -> Doc

pprPrec :: Int -> TypeBase Rank u -> Doc

pprList :: [TypeBase Rank u] -> Doc

Pretty u => Pretty (TypeBase ExtShape u)

Instance details

Methods

ppr :: TypeBase ExtShape u -> Doc

pprPrec :: Int -> TypeBase ExtShape u -> Doc

pprList :: [TypeBase ExtShape u] -> Doc

Pretty u => Pretty (TypeBase Shape u)

Instance details

Methods

ppr :: TypeBase Shape u -> Doc

pprPrec :: Int -> TypeBase Shape u -> Doc

pprList :: [TypeBase Shape u] -> Doc