Futhark.IR.Primitive

Description

Definitions of primitive types, the values that inhabit these types, and operations on these values. A primitive value can also be called a scalar.

Essentially, this module describes the subset of the (internal) Futhark language that operates on primitive types.

Types

data IntType

An integer type, ordered by size. Note that signedness is not a property of the type, but a property of the operations performed on values of these types.

Constructors

Int8

Int16

Int32

Int64

allIntTypes :: [IntType]

A list of all integer types.

data FloatType

A floating point type.

Constructors

Float16

Float32

Float64

allFloatTypes :: [FloatType]

A list of all floating-point types.

data PrimType

Low-level primitive types.

Constructors

IntType IntType

FloatType FloatType

Bool

Unit
An informationless type - An array of this type takes up no space.

allPrimTypes :: [PrimType]

A list of all primitive types.

data Int8

data Int16

data Int32

data Int64

data Half

Values

data IntValue

An integer value.

Constructors

Int8Value !Int8

Int16Value !Int16

Int32Value !Int32

Int64Value !Int64

intValue :: Integral int => IntType -> int -> IntValue

Create an IntValue from a type and an Integer.

intValueType :: IntValue -> IntType

The type of an integer value.

valueIntegral :: Integral int => IntValue -> int

Convert an IntValue to any Integral type.

data FloatValue

A floating-point value.

Constructors

Float16Value !Half

Float32Value !Float

Float64Value !Double

floatValue :: Real num => FloatType -> num -> FloatValue

Create a FloatValue from a type and a Rational.

floatValueType :: FloatValue -> FloatType

The type of a floating-point value.

data PrimValue

Non-array values.

Constructors

IntValue !IntValue

FloatValue !FloatValue

BoolValue !Bool

UnitValue
The only value of type Unit.

primValueType :: PrimValue -> PrimType

The type of a basic value.

blankPrimValue :: PrimType -> PrimValue

A "blank" value of the given primitive type - this is zero, or whatever is close to it. Don't depend on this value, but use it for e.g. creating arrays to be populated by do-loops.

Operations

data Overflow

What to do in case of arithmetic overflow. Futhark's semantics are that overflow does wraparound, but for generated code (like address arithmetic), it can be beneficial for overflow to be undefined behaviour, as it allows better optimisation of things such as GPU kernels.

Note that all values of this type are considered equal for Eq and Ord.

Constructors

OverflowWrap

OverflowUndef

data Safety

Whether something is safe or unsafe (mostly function calls, and in the context of whether operations are dynamically checked). When we inline an Unsafe function, we remove all safety checks in its body. The Ord instance picks Unsafe as being less than Safe.

For operations like integer division, a safe division will not explode the computer in case of division by zero, but instead return some unspecified value. This always involves a run-time check, so generally the unsafe variant is what the compiler will insert, but guarded by an explicit assertion elsewhere. Safe operations are useful when the optimiser wants to move e.g. a division to a location where the divisor may be zero, but where the result will only be used when it is non-zero (so it doesn't matter what result is provided with a zero divisor, as long as the program keeps running).

Constructors

Unsafe

Safe

data UnOp

Various unary operators. It is a bit ad-hoc what is a unary operator and what is a built-in function. Perhaps these should all go away eventually.

Constructors

Not
E.g., ! True == False.

Complement IntType
E.g., ~(~1) = 1.

Abs IntType
abs(-2) = 2.

FAbs FloatType
fabs(-2.0) = 2.0.

SSignum IntType
Signed sign function: ssignum(-2) = -1.

USignum IntType
Unsigned sign function: usignum(2) = 1.

FSignum FloatType
Floating-point sign function.

allUnOps :: [UnOp]

A list of all unary operators for all types.

data BinOp

Binary operators. These correspond closely to the binary operators in LLVM. Most are parametrised by their expected input and output types.

Constructors

Add IntType Overflow
Integer addition.

FAdd FloatType
Floating-point addition.

Sub IntType Overflow
Integer subtraction.

FSub FloatType
Floating-point subtraction.

Mul IntType Overflow
Integer multiplication.

FMul FloatType
Floating-point multiplication.

UDiv IntType Safety
Unsigned integer division. Rounds towards negativity infinity. Note: this is different from LLVM.

UDivUp IntType Safety
Unsigned integer division. Rounds towards positive infinity.

SDiv IntType Safety
Signed integer division. Rounds towards negativity infinity. Note: this is different from LLVM.

SDivUp IntType Safety
Signed integer division. Rounds towards positive infinity.

FDiv FloatType
Floating-point division.

FMod FloatType
Floating-point modulus.

UMod IntType Safety
Unsigned integer modulus; the countepart to UDiv.

SMod IntType Safety
Signed integer modulus; the countepart to SDiv.

SQuot IntType Safety
Signed integer division. Rounds towards zero. This corresponds to the sdiv instruction in LLVM and integer division in C.

SRem IntType Safety
Signed integer division. Rounds towards zero. This corresponds to the srem instruction in LLVM and integer modulo in C.

SMin IntType
Returns the smallest of two signed integers.

UMin IntType
Returns the smallest of two unsigned integers.

FMin FloatType
Returns the smallest of two floating-point numbers.

SMax IntType
Returns the greatest of two signed integers.

UMax IntType
Returns the greatest of two unsigned integers.

FMax FloatType
Returns the greatest of two floating-point numbers.

Shl IntType
Left-shift.

LShr IntType
Logical right-shift, zero-extended.

AShr IntType
Arithmetic right-shift, sign-extended.

And IntType
Bitwise and.

Or IntType
Bitwise or.

Xor IntType
Bitwise exclusive-or.

Pow IntType
Integer exponentiation.

FPow FloatType
Floating-point exponentiation.

LogAnd
Boolean and - not short-circuiting.

LogOr
Boolean or - not short-circuiting.

allBinOps :: [BinOp]

A list of all binary operators for all types.

data ConvOp

Conversion operators try to generalise the from t0 x to t1 instructions from LLVM.

Constructors

ZExt IntType IntType
Zero-extend the former integer type to the latter. If the new type is smaller, the result is a truncation.

SExt IntType IntType
Sign-extend the former integer type to the latter. If the new type is smaller, the result is a truncation.

FPConv FloatType FloatType
Convert value of the former floating-point type to the latter. If the new type is smaller, the result is a truncation.

FPToUI FloatType IntType
Convert a floating-point value to the nearest unsigned integer (rounding towards zero).

FPToSI FloatType IntType
Convert a floating-point value to the nearest signed integer (rounding towards zero).

UIToFP IntType FloatType
Convert an unsigned integer to a floating-point value.

SIToFP IntType FloatType
Convert a signed integer to a floating-point value.

IToB IntType
Convert an integer to a boolean value. Zero becomes false; anything else is true.

BToI IntType
Convert a boolean to an integer. True is converted to 1 and False to 0.

allConvOps :: [ConvOp]

A list of all conversion operators for all types.

data CmpOp

Comparison operators are like BinOps, but they always return a boolean value. The somewhat ugly constructor names are straight out of LLVM.

Constructors

CmpEq PrimType
All types equality.

CmpUlt IntType
Unsigned less than.

CmpUle IntType
Unsigned less than or equal.

CmpSlt IntType
Signed less than.

CmpSle IntType
Signed less than or equal.

FCmpLt FloatType
Floating-point less than.

FCmpLe FloatType
Floating-point less than or equal.

CmpLlt
Boolean less than.

CmpLle
Boolean less than or equal.

allCmpOps :: [CmpOp]

A list of all comparison operators for all types.

Unary Operations

doUnOp :: UnOp -> PrimValue -> Maybe PrimValue

Apply an UnOp to an operand. Returns Nothing if the application is mistyped.

doComplement :: IntValue -> IntValue

E.g., ~(~1) = 1.

doAbs :: IntValue -> IntValue

abs(-2) = 2.

doFAbs :: FloatValue -> FloatValue

abs(-2.0) = 2.0.

doSSignum :: IntValue -> IntValue

ssignum(-2) = -1.

doUSignum :: IntValue -> IntValue

usignum(-2) = -1.

Binary Operations

doBinOp :: BinOp -> PrimValue -> PrimValue -> Maybe PrimValue

Apply a BinOp to an operand. Returns Nothing if the application is mistyped, or outside the domain (e.g. division by zero).

doAdd :: IntValue -> IntValue -> IntValue

Integer addition.

doMul :: IntValue -> IntValue -> IntValue

Integer multiplication.

doSDiv :: IntValue -> IntValue -> Maybe IntValue

Signed integer division. Rounds towards negativity infinity. Note: this is different from LLVM.

doSMod :: IntValue -> IntValue -> Maybe IntValue

Signed integer modulus; the countepart to SDiv.

doPow :: IntValue -> IntValue -> Maybe IntValue

Signed integer exponentatation.

Conversion Operations

doConvOp :: ConvOp -> PrimValue -> Maybe PrimValue

Apply a ConvOp to an operand. Returns Nothing if the application is mistyped.

doZExt :: IntValue -> IntType -> IntValue

Zero-extend the given integer value to the size of the given type. If the type is smaller than the given value, the result is a truncation.

doSExt :: IntValue -> IntType -> IntValue

Sign-extend the given integer value to the size of the given type. If the type is smaller than the given value, the result is a truncation.

doFPConv :: FloatValue -> FloatType -> FloatValue

Convert the former floating-point type to the latter.

doFPToUI :: FloatValue -> IntType -> IntValue

Convert a floating-point value to the nearest unsigned integer (rounding towards zero).

doFPToSI :: FloatValue -> IntType -> IntValue

Convert a floating-point value to the nearest signed integer (rounding towards zero).

doUIToFP :: IntValue -> FloatType -> FloatValue

Convert an unsigned integer to a floating-point value.

doSIToFP :: IntValue -> FloatType -> FloatValue

Convert a signed integer to a floating-point value.

intToInt64 :: IntValue -> Int64

Translate an IntValue to Int64. This is guaranteed to fit.

intToWord64 :: IntValue -> Word64

Translate an IntValue to Word64. This is guaranteed to fit.

flipConvOp :: ConvOp -> ConvOp

Turn the conversion the other way around. Note that most conversions are lossy, so there is no guarantee the value will round-trip.

Comparison Operations

doCmpOp :: CmpOp -> PrimValue -> PrimValue -> Maybe Bool

Apply a CmpOp to an operand. Returns Nothing if the application is mistyped.

doCmpEq :: PrimValue -> PrimValue -> Bool

Compare any two primtive values for exact equality.

doCmpUlt :: IntValue -> IntValue -> Bool

Unsigned less than.

doCmpUle :: IntValue -> IntValue -> Bool

Unsigned less than or equal.

doCmpSlt :: IntValue -> IntValue -> Bool

Signed less than.

doCmpSle :: IntValue -> IntValue -> Bool

Signed less than or equal.

doFCmpLt :: FloatValue -> FloatValue -> Bool

Floating-point less than.

doFCmpLe :: FloatValue -> FloatValue -> Bool

Floating-point less than or equal.

Type Of

binOpType :: BinOp -> PrimType

The result type of a binary operator.

unOpType :: UnOp -> PrimType

The operand and result type of a unary operator.

cmpOpType :: CmpOp -> PrimType

The operand types of a comparison operator.

convOpType :: ConvOp -> (PrimType, PrimType)

The input and output types of a conversion operator.

Primitive functions

primFuns :: Map String ([PrimType], PrimType, [PrimValue] -> Maybe PrimValue)

A mapping from names of primitive functions to their parameter types, their result type, and a function for evaluating them.

Utility

zeroIsh :: PrimValue -> Bool

Is the given value kind of zero?

zeroIshInt :: IntValue -> Bool

Is the given integer value kind of zero?

oneIsh :: PrimValue -> Bool

Is the given value kind of one?

oneIshInt :: IntValue -> Bool

Is the given integer value kind of one?

negativeIsh :: PrimValue -> Bool

Is the given value kind of negative?

primBitSize :: PrimType -> Int

The size of a value of a given primitive type in bites.

primByteSize :: Num a => PrimType -> a

The size of a value of a given primitive type in eight-bit bytes.

intByteSize :: Num a => IntType -> a

The size of a value of a given integer type in eight-bit bytes.

floatByteSize :: Num a => FloatType -> a

The size of a value of a given floating-point type in eight-bit bytes.

commutativeBinOp :: BinOp -> Bool

True if the given binary operator is commutative.

Prettyprinting

convOpFun :: ConvOp -> String

The human-readable name for a ConvOp. This is used to expose the ConvOp in the intrinsics module of a Futhark program.

prettySigned :: Bool -> PrimType -> String

True if signed. Only makes a difference for integer types.