Futhark.IR.Prop

Description

This module provides various simple ways to query and manipulate fundamental Futhark terms, such as types and values. The intent is to keep Futhark.IRrsentation.AST.Syntax simple, and put whatever embellishments we need here. This is an internal, desugared representation.

Documentation

module Futhark.IR.Prop.Reshape

module Futhark.IR.Prop.Rearrange

module Futhark.IR.Prop.Types

module Futhark.IR.Prop.Constants

module Futhark.IR.Prop.TypeOf

module Futhark.IR.Prop.Patterns

module Futhark.IR.Prop.Names

module Futhark.IR.RetType

Built-in functions

isBuiltInFunction :: Name -> Bool

isBuiltInFunction k is True if k is an element of builtInFunctions.

builtInFunctions :: Map Name (PrimType, [PrimType])

A map of all built-in functions and their types.

Extra tools

asBasicOp :: Exp rep -> Maybe BasicOp

If the expression is a BasicOp, return it, otherwise Nothing.

safeExp :: IsOp (Op rep) => Exp rep -> Bool

An expression is safe if it is always well-defined (assuming that any required certificates have been checked) in any context. For example, array indexing is not safe, as the index may be out of bounds. On the other hand, adding two numbers cannot fail.

subExpVars :: [SubExp] -> [VName]

Return the variable names used in Var subexpressions. May contain duplicates.

subExpVar :: SubExp -> Maybe VName

If the SubExp is a Var return the variable name.

commutativeLambda :: Lambda rep -> Bool

Does the given lambda represent a known commutative function? Based on pattern matching and checking whether the lambda represents a known arithmetic operator; don't expect anything clever here.

entryPointSize :: EntryPointType -> Int

How many value parameters are accepted by this entry point? This is used to determine which of the function parameters correspond to the parameters of the original function (they must all come at the end).

defAux :: dec -> StmAux dec

A StmAux with empty Certs.

stmCerts :: Stm rep -> Certs

The certificates associated with a statement.

certify :: Certs -> Stm rep -> Stm rep

Add certificates to a statement.

expExtTypesFromPat :: Typed dec => PatT dec -> [ExtType]

Construct the type of an expression that would match the pattern.

attrsForAssert :: Attrs -> Attrs

Keep only those attributes that are relevant for Assert expressions.

lamIsBinOp :: ASTRep rep => Lambda rep -> Maybe [(BinOp, PrimType, VName, VName)]

Horizontally fission a lambda that models a binary operator.

type ASTConstraints a = (Eq a, Ord a, Show a, Rename a, Substitute a, FreeIn a, Pretty a)

A handy shorthand for properties that we usually want to things we stuff into ASTs.

class (ASTConstraints op, TypedOp op) => IsOp op where

A type class for operations.

Methods

safeOp :: op -> Bool

Like safeExp, but for arbitrary ops.

cheapOp :: op -> Bool

Should we try to hoist this out of branches?

class (RepTypes rep, PrettyRep rep, Renameable rep, Substitutable rep, FreeDec (ExpDec rep), FreeIn (LetDec rep), FreeDec (BodyDec rep), FreeIn (FParamInfo rep), FreeIn (LParamInfo rep), FreeIn (RetType rep), FreeIn (BranchType rep), IsOp (Op rep)) => ASTRep rep where

Representation-specific attributes; also means the rep supports some basic facilities.

Methods

expTypesFromPat :: (HasScope rep m, Monad m) => Pat rep -> m [BranchType rep]

Given a pattern, construct the type of a body that would match it. An implementation for many representations would be expExtTypesFromPat.