Futhark.IR.Prop.Aliases

Description

The IR tracks aliases, mostly to ensure the soundness of in-place updates, but it can also be used for other things (such as memory optimisations). This module contains the raw building blocks for determining the aliases of the values produced by expressions. It also contains some building blocks for inspecting consumption.

One important caveat is that all aliases computed here are local. Thus, they do not take aliases-of-aliases into account. See Futhark.Analysis.Alias if this is not what you want.

Documentation

subExpAliases :: SubExp -> Names

The alises of a subexpression.

expAliases :: Aliased rep => Exp rep -> [Names]

The aliases of an expression, one per non-context value returned.

patAliases :: AliasesOf dec => PatT dec -> [Names]

The aliases of each pattern element (including the context).

lookupAliases :: AliasesOf (LetDec rep) => VName -> Scope rep -> Names

Also includes the name itself.

class (RepTypes rep, AliasedOp (Op rep), AliasesOf (LetDec rep)) => Aliased rep where

The class of representations that contain aliasing information.

Methods

bodyAliases :: Body rep -> [Names]

The aliases of the body results.

consumedInBody :: Body rep -> Names

The variables consumed in the body.

class AliasesOf a where

Something that contains alias information.

Methods

aliasesOf :: a -> Names

The alias of the argument element.

Consumption

consumedInStm :: Aliased rep => Stm rep -> Names

The variables consumed in this statement.

consumedInExp :: Aliased rep => Exp rep -> Names

The variables consumed in this expression.

consumedByLambda :: Aliased rep => Lambda rep -> Names

The variables consumed by this lambda.

Extensibility

type AliasTable = Map VName Names

Pre-existing aliases for variables. Used to add transitive aliases.

class IsOp op => AliasedOp op where

The class of operations that can produce aliasing and consumption information.

Methods

opAliases :: op -> [Names]

consumedInOp :: op -> Names

class AliasedOp (OpWithAliases op) => CanBeAliased op where

The class of operations that can be given aliasing information. This is a somewhat subtle concept that is only used in the simplifier and when using "rep adapters".

Associated Types

type OpWithAliases op :: Type

The op that results when we add aliases to this op.

Methods

removeOpAliases :: OpWithAliases op -> op

Remove aliases from this op.

addOpAliases :: AliasTable -> op -> OpWithAliases op

Add aliases to this op.