Futhark.IR.Prop.Constants

Description

Possibly convenient facilities for constructing constants.

Documentation

class IsValue a where

If a Haskell type is an instance of IsValue, it means that a value of that type can be converted to a Futhark PrimValue. This is intended to cut down on boilerplate when writing compiler code - for example, you'll quickly grow tired of writing Constant (LogVal True) loc.

Methods

value :: a -> PrimValue

constant :: IsValue v => v -> SubExp

Create a Constant SubExp containing the given value.

intConst :: IntType -> Integer -> SubExp

Utility definition for reasons of type ambiguity.

floatConst :: FloatType -> Double -> SubExp

Utility definition for reasons of type ambiguity.