Futhark.IR.Prop.Names

Description

Facilities for determining which names are used in some syntactic construct. The most important interface is the FreeIn class and its instances, but for reasons related to the Haskell type system, some constructs have specialised functions.

Free names

data Names

A set of names. Note that the Ord instance is a dummy that treats everything as EQ if ==, and otherwise LT.

namesIntMap :: Names -> IntMap VName

Retrieve the data structure underlying the names representation.

nameIn :: VName -> Names -> Bool

Does the set of names contain this name?

oneName :: VName -> Names

Construct a name set from a single name.

namesFromList :: [VName] -> Names

Construct a name set from a list. Slow.

namesToList :: Names -> [VName]

Turn a name set into a list of names. Slow.

namesIntersection :: Names -> Names -> Names

The intersection of two name sets.

namesIntersect :: Names -> Names -> Bool

Do the two name sets intersect?

namesSubtract :: Names -> Names -> Names

Subtract the latter name set from the former.

mapNames :: (VName -> VName) -> Names -> Names

Map over the names in a set.

Class

class FreeIn a where

A class indicating that we can obtain free variable information from values of this type.

Minimal complete definition

Nothing

Methods

freeIn' :: a -> FV

freeIn :: FreeIn a => a -> Names

The free variables of some syntactic construct.

Specialised Functions

freeInStmsAndRes :: (FreeIn (Op rep), FreeIn (LetDec rep), FreeIn (LParamInfo rep), FreeIn (FParamInfo rep), FreeDec (BodyDec rep), FreeIn (RetType rep), FreeIn (BranchType rep), FreeDec (ExpDec rep)) => Stms rep -> Result -> FV

Return the set of variable names that are free in the given statements and result. Filters away the names that are bound by the statements.

Bound Names

boundInBody :: Body rep -> Names

The names bound by the bindings immediately in a Body.

boundByStm :: Stm rep -> Names

The names bound by a binding.

boundByStms :: Stms rep -> Names

The names bound by the bindings.

boundByLambda :: Lambda rep -> [VName]

The names of the lambda parameters plus the index parameter.

Efficient computation

class FreeIn dec => FreeDec dec where

Either return precomputed free names stored in the attribute, or the freshly computed names. Relies on lazy evaluation to avoid the work.

Minimal complete definition

Nothing

Methods

precomputed :: dec -> FV -> FV

data FV

A computation to build a free variable set.

fvBind :: Names -> FV -> FV

Consider a variable to be bound in the given FV computation.

fvName :: VName -> FV

Take note of a variable reference.

fvNames :: Names -> FV

Take note of a set of variable references.