Futhark.IR.Prop.Patterns

Description

Inspecing and modifying Pats, function parameters and pattern elements.

Function parameters

paramIdent :: Typed dec => Param dec -> Ident

An Ident corresponding to a parameter.

paramType :: Typed dec => Param dec -> Type

The Type of a parameter.

paramDeclType :: DeclTyped dec => Param dec -> DeclType

The DeclType of a parameter.

Pat elements

patElemIdent :: Typed dec => PatElemT dec -> Ident

An Ident corresponding to a pattern element.

patElemType :: Typed dec => PatElemT dec -> Type

The type of a name bound by a PatElem.

setPatElemDec :: PatElemT oldattr -> newattr -> PatElemT newattr

Set the rep of a PatElem.

patIdents :: Typed dec => PatT dec -> [Ident]

Return a list of the Idents bound by the Pat.

patNames :: PatT dec -> [VName]

Return a list of the Names bound by the Pat.

patTypes :: Typed dec => PatT dec -> [Type]

Return a list of the typess bound by the pattern.

patSize :: PatT dec -> Int

Return the number of names bound by the pattern.

Pat construction

basicPat :: [Ident] -> PatT Type

Create a pattern using Type as the attribute.