Futhark.IR.Prop.Types

Description

Functions for inspecting and constructing various types.

Documentation

rankShaped :: ArrayShape shape => TypeBase shape u -> TypeBase Rank u

Remove shape information from a type.

arrayRank :: ArrayShape shape => TypeBase shape u -> Int

Return the dimensionality of a type. For non-arrays, this is zero. For a one-dimensional array it is one, for a two-dimensional it is two, and so forth.

arrayShape :: ArrayShape shape => TypeBase shape u -> shape

Return the shape of a type - for non-arrays, this is the mempty.

setArrayShape :: ArrayShape newshape => TypeBase oldshape u -> newshape -> TypeBase newshape u

Set the shape of an array. If the given type is not an array, return the type unchanged.

existential :: ExtType -> Bool

True if the given type has a dimension that is existentially sized.

uniqueness :: TypeBase shape Uniqueness -> Uniqueness

Return the uniqueness of a type.

unique :: TypeBase shape Uniqueness -> Bool

unique t is True if the type of the argument is unique.

staticShapes :: [TypeBase Shape u] -> [TypeBase ExtShape u]

Convert types with non-existential shapes to types with non-existential shapes. Only the representation is changed, so all the shapes will be Free.

staticShapes1 :: TypeBase Shape u -> TypeBase ExtShape u

As staticShapes, but on a single type.

primType :: TypeBase shape u -> Bool

A type is a primitive type if it is not an array or memory block.

arrayOf :: ArrayShape shape => TypeBase shape u_unused -> shape -> u -> TypeBase shape u

arrayOf t s u constructs an array type. The convenience compared to using the Array constructor directly is that t can itself be an array. If t is an n-dimensional array, and s is a list of length n, the resulting type is of an n+m dimensions. The uniqueness of the new array will be u, no matter the uniqueness of t. If the shape s has rank 0, then the t will be returned, although if it is an array, with the uniqueness changed to u.

arrayOfRow :: ArrayShape (ShapeBase d) => TypeBase (ShapeBase d) NoUniqueness -> d -> TypeBase (ShapeBase d) NoUniqueness

Construct an array whose rows are the given type, and the outer size is the given dimension. This is just a convenient wrapper around arrayOf.

arrayOfShape :: Type -> Shape -> Type

Construct an array whose rows are the given type, and the outer size is the given Shape. This is just a convenient wrapper around arrayOf.

setOuterSize :: ArrayShape (ShapeBase d) => TypeBase (ShapeBase d) u -> d -> TypeBase (ShapeBase d) u

Replace the size of the outermost dimension of an array. If the given type is not an array, it is returned unchanged.

setDimSize :: ArrayShape (ShapeBase d) => Int -> TypeBase (ShapeBase d) u -> d -> TypeBase (ShapeBase d) u

Replace the size of the given dimension of an array. If the given type is not an array, it is returned unchanged.

setOuterDim :: ShapeBase d -> d -> ShapeBase d

Replace the outermost dimension of an array shape.

setOuterDims :: ShapeBase d -> Int -> ShapeBase d -> ShapeBase d

Replace some outermost dimensions of an array shape.

setDim :: Int -> ShapeBase d -> d -> ShapeBase d

Replace the specified dimension of an array shape.

setArrayDims :: TypeBase oldshape u -> [SubExp] -> TypeBase Shape u

Set the dimensions of an array. If the given type is not an array, return the type unchanged.

peelArray :: Int -> TypeBase Shape u -> Maybe (TypeBase Shape u)

peelArray n t returns the type resulting from peeling the first n array dimensions from t. Returns Nothing if t has less than n dimensions.

stripArray :: Int -> TypeBase Shape u -> TypeBase Shape u

stripArray n t removes the n outermost layers of the array. Essentially, it is the type of indexing an array of type t with n indexes.

arrayDims :: TypeBase Shape u -> [SubExp]

Return the dimensions of a type - for non-arrays, this is the empty list.

arrayExtDims :: TypeBase ExtShape u -> [ExtSize]

Return the existential dimensions of a type - for non-arrays, this is the empty list.

shapeSize :: Int -> Shape -> SubExp

Return the size of the given dimension. If the dimension does not exist, the zero constant is returned.

arraySize :: Int -> TypeBase Shape u -> SubExp

Return the size of the given dimension. If the dimension does not exist, the zero constant is returned.

arraysSize :: Int -> [TypeBase Shape u] -> SubExp

Return the size of the given dimension in the first element of the given type list. If the dimension does not exist, or no types are given, the zero constant is returned.

elemType :: TypeBase shape u -> PrimType

Returns the bottommost type of an array. For [][]i32, this would be i32. If the given type is not an array, it is returned.

rowType :: TypeBase Shape u -> TypeBase Shape u

Return the immediate row-type of an array. For [[int]], this would be [int].

transposeType :: Type -> Type

Swap the two outer dimensions of the type.

rearrangeType :: [Int] -> Type -> Type

Rearrange the dimensions of the type. If the length of the permutation does not match the rank of the type, the permutation will be extended with identity.

mapOnExtType :: Monad m => (SubExp -> m SubExp) -> TypeBase ExtShape u -> m (TypeBase ExtShape u)

Transform any SubExps in the type.

mapOnType :: Monad m => (SubExp -> m SubExp) -> TypeBase Shape u -> m (TypeBase Shape u)

Transform any SubExps in the type.

diet :: TypeBase shape Uniqueness -> Diet

diet t returns a description of how a function parameter of type t might consume its argument.

subtypeOf :: (Ord u, ArrayShape shape) => TypeBase shape u -> TypeBase shape u -> Bool

x `subtypeOf` y is true if x is a subtype of y (or equal to y), meaning x is valid whenever y is.

subtypesOf :: (Ord u, ArrayShape shape) => [TypeBase shape u] -> [TypeBase shape u] -> Bool

xs `subtypesOf` ys is true if xs is the same size as ys, and each element in xs is a subtype of the corresponding element in ys..

toDecl :: TypeBase shape NoUniqueness -> Uniqueness -> TypeBase shape Uniqueness

Add the given uniqueness information to the types.

fromDecl :: TypeBase shape Uniqueness -> TypeBase shape NoUniqueness

Remove uniqueness information from the type.

isExt :: Ext a -> Maybe Int

If an existential, then return its existential index.

isFree :: Ext a -> Maybe a

If a known size, then return that size.

extractShapeContext :: [TypeBase ExtShape u] -> [[a]] -> [a]

Given the existential return type of a function, and the shapes of the values returned by the function, return the existential shape context. That is, those sizes that are existential in the return type.

shapeContext :: [TypeBase ExtShape u] -> Set Int

The Ext integers used for existential sizes in the given types.

hasStaticShape :: TypeBase ExtShape u -> Maybe (TypeBase Shape u)

If all dimensions of the given ExtShape are statically known, change to the corresponding Shape.

generaliseExtTypes :: [TypeBase ExtShape u] -> [TypeBase ExtShape u] -> [TypeBase ExtShape u]

Given two lists of ExtTypes of the same length, return a list of ExtTypes that is a subtype of the two operands.

existentialiseExtTypes :: [VName] -> [ExtType] -> [ExtType]

Given a list of ExtTypes and a list of "forbidden" names, modify the dimensions of the ExtTypes such that they are Ext where they were previously Free with a variable in the set of forbidden names.

shapeExtMapping :: [TypeBase ExtShape u] -> [TypeBase Shape u1] -> Map Int SubExp

Produce a mapping for the dimensions context.

Abbreviations

int8 :: PrimType

IntType Int8

int16 :: PrimType

IntType Int16

int32 :: PrimType

IntType Int32

int64 :: PrimType

IntType Int64

float32 :: PrimType

FloatType Float32

float64 :: PrimType

FloatType Float64

The Typed typeclass

class Typed t where

Typeclass for things that contain Types.

Methods

typeOf :: t -> Type

class DeclTyped t where

Typeclass for things that contain DeclTypes.

Methods

declTypeOf :: t -> DeclType

class FixExt t => ExtTyped t where

Typeclass for things that contain ExtTypes.

Methods

extTypeOf :: t -> ExtType

class FixExt t => DeclExtTyped t where

Typeclass for things that contain DeclExtTypes.

Methods

declExtTypeOf :: t -> DeclExtType

class Typed a => SetType a where

Typeclass for things whose type can be changed.

Methods

setType :: a -> Type -> a

class FixExt t where

Something with an existential context that can be (partially) fixed.

Methods

fixExt :: Int -> SubExp -> t -> t

Fix the given existentional variable to the indicated free value.