Futhark.IR.Rep

Description

The core Futhark AST is parameterised by a rep type parameter, which is then used to invoke the type families defined here.

Documentation

class (Show (LetDec l), Show (ExpDec l), Show (BodyDec l), Show (FParamInfo l), Show (LParamInfo l), Show (RetType l), Show (BranchType l), Show (Op l), Eq (LetDec l), Eq (ExpDec l), Eq (BodyDec l), Eq (FParamInfo l), Eq (LParamInfo l), Eq (RetType l), Eq (BranchType l), Eq (Op l), Ord (LetDec l), Ord (ExpDec l), Ord (BodyDec l), Ord (FParamInfo l), Ord (LParamInfo l), Ord (RetType l), Ord (BranchType l), Ord (Op l), IsRetType (RetType l), IsBodyType (BranchType l), Typed (FParamInfo l), Typed (LParamInfo l), Typed (LetDec l), DeclTyped (FParamInfo l)) => RepTypes l

A collection of type families giving various common types for a representation, along with constraints specifying that the types they map to should satisfy some minimal requirements.

Associated Types

type LetDec l :: Type

Decoration for every let-pattern element.

type LetDec l = Type

type ExpDec l :: Type

Decoration for every expression.

type ExpDec l = ()

type BodyDec l :: Type

Decoration for every body.

type BodyDec l = ()

type FParamInfo l :: Type

Decoration for every (non-lambda) function parameter.

type FParamInfo l = DeclType

type LParamInfo l :: Type

Decoration for every lambda function parameter.

type LParamInfo l = Type

type RetType l :: Type

The return type decoration of function calls.

type RetType l = DeclExtType

type BranchType l :: Type

The return type decoration of branches.

type BranchType l = ExtType

type Op l :: Type

Extensible operation.

type Op l = ()

module Futhark.IR.RetType