Futhark.IR.RetType

Description

This module exports a type class covering representations of function return types.

Documentation

class (Show rt, Eq rt, Ord rt, ExtTyped rt) => IsBodyType rt where

A type representing the return type of a body. It should contain at least the information contained in a list of ExtTypes, but may have more, notably an existential context.

Methods

primBodyType :: PrimType -> rt

Construct a body type from a primitive type.

class (Show rt, Eq rt, Ord rt, DeclExtTyped rt) => IsRetType rt where

A type representing the return type of a function. In practice, a list of these will be used. It should contain at least the information contained in an ExtType, but may have more, notably an existential context.

Methods

primRetType :: PrimType -> rt

Contruct a return type from a primitive type.

applyRetType :: Typed dec => [rt] -> [Param dec] -> [(SubExp, Type)] -> Maybe [rt]

Given a function return type, the parameters of the function, and the arguments for a concrete call, return the instantiated return type for the concrete call, if valid.

expectedTypes :: Typed t => [VName] -> [t] -> [SubExp] -> [Type]

Given shape parameter names and types, produce the types of arguments accepted.