Futhark.IR.SOACS

Description

A simple representation with SOACs and nested parallelism.

Documentation

data SOACS

The rep for the basic representation.

Syntax types

type Body = Body SOACS

type Stm = Stm SOACS

type Pat = Pat SOACS

type Exp = Exp SOACS

type Lambda = Lambda SOACS

type FParam = FParam SOACS

type LParam = LParam SOACS

type RetType = RetType SOACS

type PatElem = PatElem SOACS

Module re-exports

module Futhark.IR.Prop

module Futhark.IR.Traversals

module Futhark.IR.Pretty

data Int8

data Int16

data Int32

data Int64

data Word8

data Word16

data Word32

data Word64

data Half

srclocOf :: Located a => a -> SrcLoc

The SrcLoc of a Located value.

data Loc

Location type, consisting of a beginning position and an end position.

data SrcLoc

Source location type. Source location are all equal, which allows AST nodes to be compared modulo location information.

class Located a where

Located values have a location.

Minimal complete definition

locOf

Methods

locOf :: a -> Loc

locOfList :: [a] -> Loc

pretty :: Pretty a => a -> String

Prettyprint a value, wrapped to 80 characters.

data ConvOp

Conversion operators try to generalise the from t0 x to t1 instructions from LLVM.

Constructors

ZExt IntType IntType
Zero-extend the former integer type to the latter. If the new type is smaller, the result is a truncation.

SExt IntType IntType
Sign-extend the former integer type to the latter. If the new type is smaller, the result is a truncation.

FPConv FloatType FloatType
Convert value of the former floating-point type to the latter. If the new type is smaller, the result is a truncation.

FPToUI FloatType IntType
Convert a floating-point value to the nearest unsigned integer (rounding towards zero).

FPToSI FloatType IntType
Convert a floating-point value to the nearest signed integer (rounding towards zero).

UIToFP IntType FloatType
Convert an unsigned integer to a floating-point value.

SIToFP IntType FloatType
Convert a signed integer to a floating-point value.

IToB IntType
Convert an integer to a boolean value. Zero becomes false; anything else is true.

BToI IntType
Convert a boolean to an integer. True is converted to 1 and False to 0.

data CmpOp

Comparison operators are like BinOps, but they always return a boolean value. The somewhat ugly constructor names are straight out of LLVM.

Constructors

CmpEq PrimType
All types equality.

CmpUlt IntType
Unsigned less than.

CmpUle IntType
Unsigned less than or equal.

CmpSlt IntType
Signed less than.

CmpSle IntType
Signed less than or equal.

FCmpLt FloatType
Floating-point less than.

FCmpLe FloatType
Floating-point less than or equal.

CmpLlt
Boolean less than.

CmpLle
Boolean less than or equal.

data BinOp

Binary operators. These correspond closely to the binary operators in LLVM. Most are parametrised by their expected input and output types.

Constructors

Add IntType Overflow
Integer addition.

FAdd FloatType
Floating-point addition.

Sub IntType Overflow
Integer subtraction.

FSub FloatType
Floating-point subtraction.

Mul IntType Overflow
Integer multiplication.

FMul FloatType
Floating-point multiplication.

UDiv IntType Safety
Unsigned integer division. Rounds towards negativity infinity. Note: this is different from LLVM.

UDivUp IntType Safety
Unsigned integer division. Rounds towards positive infinity.

SDiv IntType Safety
Signed integer division. Rounds towards negativity infinity. Note: this is different from LLVM.

SDivUp IntType Safety
Signed integer division. Rounds towards positive infinity.

FDiv FloatType
Floating-point division.

FMod FloatType
Floating-point modulus.

UMod IntType Safety
Unsigned integer modulus; the countepart to UDiv.

SMod IntType Safety
Signed integer modulus; the countepart to SDiv.

SQuot IntType Safety
Signed integer division. Rounds towards zero. This corresponds to the sdiv instruction in LLVM and integer division in C.

SRem IntType Safety
Signed integer division. Rounds towards zero. This corresponds to the srem instruction in LLVM and integer modulo in C.

SMin IntType
Returns the smallest of two signed integers.

UMin IntType
Returns the smallest of two unsigned integers.

FMin FloatType
Returns the smallest of two floating-point numbers.

SMax IntType
Returns the greatest of two signed integers.

UMax IntType
Returns the greatest of two unsigned integers.

FMax FloatType
Returns the greatest of two floating-point numbers.

Shl IntType
Left-shift.

LShr IntType
Logical right-shift, zero-extended.

AShr IntType
Arithmetic right-shift, sign-extended.

And IntType
Bitwise and.

Or IntType
Bitwise or.

Xor IntType
Bitwise exclusive-or.

Pow IntType
Integer exponentiation.

FPow FloatType
Floating-point exponentiation.

LogAnd
Boolean and - not short-circuiting.

LogOr
Boolean or - not short-circuiting.

data Safety

Whether something is safe or unsafe (mostly function calls, and in the context of whether operations are dynamically checked). When we inline an Unsafe function, we remove all safety checks in its body. The Ord instance picks Unsafe as being less than Safe.

For operations like integer division, a safe division will not explode the computer in case of division by zero, but instead return some unspecified value. This always involves a run-time check, so generally the unsafe variant is what the compiler will insert, but guarded by an explicit assertion elsewhere. Safe operations are useful when the optimiser wants to move e.g. a division to a location where the divisor may be zero, but where the result will only be used when it is non-zero (so it doesn't matter what result is provided with a zero divisor, as long as the program keeps running).

Constructors

Unsafe

Safe

data Overflow

What to do in case of arithmetic overflow. Futhark's semantics are that overflow does wraparound, but for generated code (like address arithmetic), it can be beneficial for overflow to be undefined behaviour, as it allows better optimisation of things such as GPU kernels.

Note that all values of this type are considered equal for Eq and Ord.

Constructors

OverflowWrap

OverflowUndef

data UnOp

Various unary operators. It is a bit ad-hoc what is a unary operator and what is a built-in function. Perhaps these should all go away eventually.

Constructors

Not
E.g., ! True == False.

Complement IntType
E.g., ~(~1) = 1.

Abs IntType
abs(-2) = 2.

FAbs FloatType
fabs(-2.0) = 2.0.

SSignum IntType
Signed sign function: ssignum(-2) = -1.

USignum IntType
Unsigned sign function: usignum(2) = 1.

FSignum FloatType
Floating-point sign function.

data PrimValue

Non-array values.

Constructors

IntValue !IntValue

FloatValue !FloatValue

BoolValue !Bool

UnitValue
The only value of type Unit.

data FloatValue

A floating-point value.

Constructors

Float16Value !Half

Float32Value !Float

Float64Value !Double

data IntValue

An integer value.

Constructors

Int8Value !Int8

Int16Value !Int16

Int32Value !Int32

Int64Value !Int64

data PrimType

Low-level primitive types.

Constructors

IntType IntType

FloatType FloatType

Bool

Unit
An informationless type - An array of this type takes up no space.

data FloatType

A floating point type.

Constructors

Float16

Float32

Float64

data IntType

An integer type, ordered by size. Note that signedness is not a property of the type, but a property of the operations performed on values of these types.

Constructors

Int8

Int16

Int32

Int64

allIntTypes :: [IntType]

A list of all integer types.

allFloatTypes :: [FloatType]

A list of all floating-point types.

allPrimTypes :: [PrimType]

A list of all primitive types.

intValue :: Integral int => IntType -> int -> IntValue

Create an IntValue from a type and an Integer.

intValueType :: IntValue -> IntType

The type of an integer value.

valueIntegral :: Integral int => IntValue -> int

Convert an IntValue to any Integral type.

floatValue :: Real num => FloatType -> num -> FloatValue

Create a FloatValue from a type and a Rational.

floatValueType :: FloatValue -> FloatType

The type of a floating-point value.

primValueType :: PrimValue -> PrimType

The type of a basic value.

blankPrimValue :: PrimType -> PrimValue

A "blank" value of the given primitive type - this is zero, or whatever is close to it. Don't depend on this value, but use it for e.g. creating arrays to be populated by do-loops.

allUnOps :: [UnOp]

A list of all unary operators for all types.

allBinOps :: [BinOp]

A list of all binary operators for all types.

allCmpOps :: [CmpOp]

A list of all comparison operators for all types.

allConvOps :: [ConvOp]

A list of all conversion operators for all types.

doUnOp :: UnOp -> PrimValue -> Maybe PrimValue

Apply an UnOp to an operand. Returns Nothing if the application is mistyped.

doComplement :: IntValue -> IntValue

E.g., ~(~1) = 1.

doAbs :: IntValue -> IntValue

abs(-2) = 2.

doFAbs :: FloatValue -> FloatValue

abs(-2.0) = 2.0.

doSSignum :: IntValue -> IntValue

ssignum(-2) = -1.

doUSignum :: IntValue -> IntValue

usignum(-2) = -1.

doBinOp :: BinOp -> PrimValue -> PrimValue -> Maybe PrimValue

Apply a BinOp to an operand. Returns Nothing if the application is mistyped, or outside the domain (e.g. division by zero).

doAdd :: IntValue -> IntValue -> IntValue

Integer addition.

doMul :: IntValue -> IntValue -> IntValue

Integer multiplication.

doSDiv :: IntValue -> IntValue -> Maybe IntValue

Signed integer division. Rounds towards negativity infinity. Note: this is different from LLVM.

doSMod :: IntValue -> IntValue -> Maybe IntValue

Signed integer modulus; the countepart to SDiv.

doPow :: IntValue -> IntValue -> Maybe IntValue

Signed integer exponentatation.

doConvOp :: ConvOp -> PrimValue -> Maybe PrimValue

Apply a ConvOp to an operand. Returns Nothing if the application is mistyped.

flipConvOp :: ConvOp -> ConvOp

Turn the conversion the other way around. Note that most conversions are lossy, so there is no guarantee the value will round-trip.

doZExt :: IntValue -> IntType -> IntValue

Zero-extend the given integer value to the size of the given type. If the type is smaller than the given value, the result is a truncation.

doSExt :: IntValue -> IntType -> IntValue

Sign-extend the given integer value to the size of the given type. If the type is smaller than the given value, the result is a truncation.

doFPConv :: FloatValue -> FloatType -> FloatValue

Convert the former floating-point type to the latter.

doFPToUI :: FloatValue -> IntType -> IntValue

Convert a floating-point value to the nearest unsigned integer (rounding towards zero).

doFPToSI :: FloatValue -> IntType -> IntValue

Convert a floating-point value to the nearest signed integer (rounding towards zero).

doUIToFP :: IntValue -> FloatType -> FloatValue

Convert an unsigned integer to a floating-point value.

doSIToFP :: IntValue -> FloatType -> FloatValue

Convert a signed integer to a floating-point value.

doCmpOp :: CmpOp -> PrimValue -> PrimValue -> Maybe Bool

Apply a CmpOp to an operand. Returns Nothing if the application is mistyped.

doCmpEq :: PrimValue -> PrimValue -> Bool

Compare any two primtive values for exact equality.

doCmpUlt :: IntValue -> IntValue -> Bool

Unsigned less than.

doCmpUle :: IntValue -> IntValue -> Bool

Unsigned less than or equal.

doCmpSlt :: IntValue -> IntValue -> Bool

Signed less than.

doCmpSle :: IntValue -> IntValue -> Bool

Signed less than or equal.

doFCmpLt :: FloatValue -> FloatValue -> Bool

Floating-point less than.

doFCmpLe :: FloatValue -> FloatValue -> Bool

Floating-point less than or equal.

intToWord64 :: IntValue -> Word64

Translate an IntValue to Word64. This is guaranteed to fit.

intToInt64 :: IntValue -> Int64

Translate an IntValue to Int64. This is guaranteed to fit.

binOpType :: BinOp -> PrimType

The result type of a binary operator.

cmpOpType :: CmpOp -> PrimType

The operand types of a comparison operator.

unOpType :: UnOp -> PrimType

The operand and result type of a unary operator.

convOpType :: ConvOp -> (PrimType, PrimType)

The input and output types of a conversion operator.

primFuns :: Map String ([PrimType], PrimType, [PrimValue] -> Maybe PrimValue)

A mapping from names of primitive functions to their parameter types, their result type, and a function for evaluating them.

zeroIsh :: PrimValue -> Bool

Is the given value kind of zero?

oneIsh :: PrimValue -> Bool

Is the given value kind of one?

negativeIsh :: PrimValue -> Bool

Is the given value kind of negative?

zeroIshInt :: IntValue -> Bool

Is the given integer value kind of zero?

oneIshInt :: IntValue -> Bool

Is the given integer value kind of one?

primBitSize :: PrimType -> Int

The size of a value of a given primitive type in bites.

primByteSize :: Num a => PrimType -> a

The size of a value of a given primitive type in eight-bit bytes.

intByteSize :: Num a => IntType -> a

The size of a value of a given integer type in eight-bit bytes.

floatByteSize :: Num a => FloatType -> a

The size of a value of a given floating-point type in eight-bit bytes.

commutativeBinOp :: BinOp -> Bool

True if the given binary operator is commutative.

convOpFun :: ConvOp -> String

The human-readable name for a ConvOp. This is used to expose the ConvOp in the intrinsics module of a Futhark program.

prettySigned :: Bool -> PrimType -> String

True if signed. Only makes a difference for integer types.

data VName

A name tagged with some integer. Only the integer is used in comparisons, no matter the type of vn.

Constructors

VName !Name !Int

data Name

The abstract (not really) type representing names in the Futhark compiler. Strings, being lists of characters, are very slow, while Texts are based on byte-arrays.

data Uniqueness

The uniqueness attribute of a type. This essentially indicates whether or not in-place modifications are acceptable. With respect to ordering, Unique is greater than Nonunique.

Constructors

Nonunique
May have references outside current function.

Unique
No references outside current function.

defaultEntryPoint :: Name

The name of the default program entry point (main).

nameToString :: Name -> String

Convert a name to the corresponding list of characters.

nameFromString :: String -> Name

Convert a list of characters to the corresponding name.

nameToText :: Name -> Text

Convert a name to the corresponding Text.

nameFromText :: Text -> Name

Convert a Text to the corresponding name.

locStr :: Located a => a -> String

A human-readable location string, of the form filename:lineno:columnno. This follows the GNU coding standards for error messages: https://www.gnu.org/prep/standards/html_node/Errors.html

This function assumes that both start and end position is in the same file (it is not clear what the alternative would even mean).

locStrRel :: (Located a, Located b) => a -> b -> String

Like locStr, but locStrRel prev now prints the location now with the file name left out if the same as prev. This is useful when printing messages that are all in the context of some initially printed location (e.g. the first mention contains the file name; the rest just line and column name).

prettyStacktrace :: Int -> [String] -> String

Given a list of strings representing entries in the stack trace and the index of the frame to highlight, produce a final newline-terminated string for showing to the user. This string should also be preceded by a newline. The most recent stack frame must come first in the list.

baseTag :: VName -> Int

Return the tag contained in the VName.

baseName :: VName -> Name

Return the name contained in the VName.

baseString :: VName -> String

Return the base Name converted to a string.

quote :: String -> String

Enclose a string in the prefered quotes used in error messages. These are picked to not collide with characters permitted in identifiers.

pquote :: Doc -> Doc

As quote, but works on prettyprinted representation.

newtype Attrs

Every statement is associated with a set of attributes, which can have various effects throughout the compiler.

Constructors

Attrs

Fields

unAttrs :: Set Attr

data Attr

A single attribute.

Constructors

AttrName Name

AttrInt Integer

AttrComp Name [Attr]

data ErrorMsgPart a

A part of an error message.

Constructors

ErrorString String
A literal string.

ErrorVal PrimType a
A run-time value.

newtype ErrorMsg a

An error message is a list of error parts, which are concatenated to form the final message.

Constructors

ErrorMsg [ErrorMsgPart a]

data PatElemT dec

An element of a pattern - consisting of a name and an addditional parametric decoration. This decoration is what is expected to contain the type of the resulting variable.

data FlatSlice d

Constructors

FlatSlice d [FlatDimIndex d]

data FlatDimIndex d

Constructors

FlatDimIndex

Fields

d

Number of elements in dimension

d

Stride of dimension

newtype Slice d

A list of DimIndexs, indicating how an array should be sliced. Whenever a function accepts a Slice, that slice should be total, i.e, cover all dimensions of the array. Deviators should be indicated by taking a list of DimIndexes instead.

Constructors

Slice

Fields

unSlice :: [DimIndex d]

data DimIndex d

How to index a single dimension of an array.

Constructors

DimFix d
Fix index in this dimension.

DimSlice d d d
DimSlice start_offset num_elems stride.

data Param dec

A function or lambda parameter.

Constructors

Param

Fields

paramAttrs :: Attrs

Attributes of the parameter. When constructing a parameter, feel free to just pass mempty.

paramName :: VName

Name of the parameter.

paramDec :: dec

Function parameter decoration.

data SubExp

A subexpression is either a scalar constant or a variable. One important property is that evaluation of a subexpression is guaranteed to complete in constant time.

Constructors

Constant PrimValue

Var VName

newtype Certs

A list of names used for certificates in some expressions.

Constructors

Certs

Fields

unCerts :: [VName]

data Ident

An identifier consists of its name and the type of the value bound to the identifier.

Constructors

Ident

Fields

identName :: VName

identType :: Type

data Diet

Information about which parts of a value/type are consumed. For example, we might say that a function taking three arguments of types ([int], *[int], [int]) has diet [Observe, Consume, Observe].

Constructors

Consume
Consumes this value.

Observe
Only observes value in this position, does not consume. A result may alias this.

ObservePrim
As Observe, but the result will not alias, because the parameter does not carry aliases.

type DeclExtType = TypeBase ExtShape Uniqueness

An ExtType with uniqueness information, used for function return types.

type DeclType = TypeBase Shape Uniqueness

A type with shape and uniqueness information, used declaring return- and parameters types.

type ExtType = TypeBase ExtShape NoUniqueness

A type with existentially quantified shapes - used as part of function (and function-like) return types. Generally only makes sense when used in a list.

type Type = TypeBase Shape NoUniqueness

A type with shape information, used for describing the type of variables.

data TypeBase shape u

The type of a value. When comparing types for equality with ==, shapes must match.

Constructors

Prim PrimType

Acc VName Shape [Type] u
Token, index space, element type, and uniqueness.

Array PrimType shape u

Mem Space

data NoUniqueness

A fancier name for () - encodes no uniqueness information.

Constructors

NoUniqueness

type SpaceId = String

A string representing a specific non-default memory space.

data Space

The memory space of a block. If DefaultSpace, this is the "default" space, whatever that is. The exact meaning of the SpaceId depends on the backend used. In GPU kernels, for example, this is used to distinguish between constant, global and shared memory spaces. In GPU-enabled host code, it is used to distinguish between host memory (DefaultSpace) and GPU space.

Constructors

DefaultSpace

Space SpaceId

ScalarSpace [SubExp] PrimType
A special kind of memory that is a statically sized array of some primitive type. Used for private memory on GPUs.

class (Monoid a, Eq a, Ord a) => ArrayShape a where

A class encompassing types containing array shape information.

Methods

shapeRank :: a -> Int

Return the rank of an array with the given size.

subShapeOf :: a -> a -> Bool

Check whether one shape if a subset of another shape.

newtype Rank

The size of an array type as merely the number of dimensions, with no further information.

Constructors

Rank Int

type ExtShape = ShapeBase ExtSize

Like Shape but some of its elements may be bound in a local environment instead. These are denoted with integral indices.

type ExtSize = Ext SubExp

The size of this dimension.

data Ext a

Something that may be existential.

Constructors

Ext Int

Free a

type Shape = ShapeBase SubExp

The size of an array as a list of subexpressions. If a variable, that variable must be in scope where this array is used.

newtype ShapeBase d

The size of an array type as a list of its dimension sizes, with the type of sizes being parametric.

Constructors

Shape

Fields

shapeDims :: [d]

data Commutativity

Whether some operator is commutative or not. The Monoid instance returns the least commutative of its arguments.

Constructors

Noncommutative

Commutative

stripDims :: Int -> ShapeBase d -> ShapeBase d

stripDims n shape strips the outer n dimensions from shape.

dimFix :: DimIndex d -> Maybe d

If the argument is a DimFix, return its component.

sliceIndices :: Slice d -> Maybe [d]

If the slice is all DimFixs, return the components.

sliceDims :: Slice d -> [d]

The dimensions of the array produced by this slice.

unitSlice :: Num d => d -> d -> DimIndex d

A slice with a stride of one.

fixSlice :: Num d => Slice d -> [d] -> [d]

Fix the DimSlices of a slice. The number of indexes must equal the length of sliceDims for the slice.

sliceSlice :: Num d => Slice d -> Slice d -> Slice d

Further slice the DimSlices of a slice. The number of slices must equal the length of sliceDims for the slice.

flatSliceDims :: FlatSlice d -> [d]

flatSliceStrides :: FlatSlice d -> [d]

errorMsgArgTypes :: ErrorMsg a -> [PrimType]

How many non-constant parts does the error message have, and what is their type?

oneAttr :: Attr -> Attrs

Construct Attrs from a single Attr.

inAttrs :: Attr -> Attrs -> Bool

Is the given attribute to be found in the attribute set?

withoutAttrs :: Attrs -> Attrs -> Attrs

x withoutAttrs y gives x except for any attributes also in y.

mapAttrs :: (Attr -> a) -> Attrs -> [a]

Map a function over an attribute set.

class (Show rt, Eq rt, Ord rt, DeclExtTyped rt) => IsRetType rt where

A type representing the return type of a function. In practice, a list of these will be used. It should contain at least the information contained in an ExtType, but may have more, notably an existential context.

Methods

primRetType :: PrimType -> rt

Contruct a return type from a primitive type.

applyRetType :: Typed dec => [rt] -> [Param dec] -> [(SubExp, Type)] -> Maybe [rt]

Given a function return type, the parameters of the function, and the arguments for a concrete call, return the instantiated return type for the concrete call, if valid.

class (Show rt, Eq rt, Ord rt, ExtTyped rt) => IsBodyType rt where

A type representing the return type of a body. It should contain at least the information contained in a list of ExtTypes, but may have more, notably an existential context.

Methods

primBodyType :: PrimType -> rt

Construct a body type from a primitive type.

expectedTypes :: Typed t => [VName] -> [t] -> [SubExp] -> [Type]

Given shape parameter names and types, produce the types of arguments accepted.

class (Show (LetDec l), Show (ExpDec l), Show (BodyDec l), Show (FParamInfo l), Show (LParamInfo l), Show (RetType l), Show (BranchType l), Show (Op l), Eq (LetDec l), Eq (ExpDec l), Eq (BodyDec l), Eq (FParamInfo l), Eq (LParamInfo l), Eq (RetType l), Eq (BranchType l), Eq (Op l), Ord (LetDec l), Ord (ExpDec l), Ord (BodyDec l), Ord (FParamInfo l), Ord (LParamInfo l), Ord (RetType l), Ord (BranchType l), Ord (Op l), IsRetType (RetType l), IsBodyType (BranchType l), Typed (FParamInfo l), Typed (LParamInfo l), Typed (LetDec l), DeclTyped (FParamInfo l)) => RepTypes l

A collection of type families giving various common types for a representation, along with constraints specifying that the types they map to should satisfy some minimal requirements.

Associated Types

type LetDec l :: Type

Decoration for every let-pattern element.

type LetDec l = Type

type ExpDec l :: Type

Decoration for every expression.

type ExpDec l = ()

type BodyDec l :: Type

Decoration for every body.

type BodyDec l = ()

type FParamInfo l :: Type

Decoration for every (non-lambda) function parameter.

type FParamInfo l = DeclType

type LParamInfo l :: Type

Decoration for every lambda function parameter.

type LParamInfo l = Type

type BranchType l :: Type

The return type decoration of branches.

type BranchType l = ExtType

type Op l :: Type

Extensible operation.

type Op l = ()

data Prog rep

An entire Futhark program.

Constructors

Prog

Fields

progConsts :: Stms rep

Top-level constants that are computed at program startup, and which are in scope inside all functions.

progFuns :: [FunDef rep]

The functions comprising the program. All funtions are also available in scope in the definitions of the constants, so be careful not to introduce circular dependencies (not currently checked).

type EntryPoint = (Name, [EntryParam], [EntryPointType])

Information about the inputs and outputs (return value) of an entry point.

data EntryParam

An entry point parameter, comprising its name and original type.

Constructors

EntryParam

Fields

entryParamName :: Name

entryParamType :: EntryPointType

data EntryPointType

Every entry point argument and return value has an annotation indicating how it maps to the original source program type.

Constructors

TypeUnsigned Uniqueness
Is an unsigned integer or array of unsigned integers.

TypeOpaque Uniqueness String Int
A black box type comprising this many core values. The string is a human-readable description with no other semantics.

TypeDirect Uniqueness
Maps directly.

data FunDef rep

Function Declarations

Constructors

FunDef

Fields

funDefEntryPoint :: Maybe EntryPoint

Contains a value if this function is an entry point.

funDefAttrs :: Attrs

funDefName :: Name

funDefRetType :: [RetType rep]

funDefParams :: [FParam rep]

funDefBody :: BodyT rep

data LambdaT rep

Anonymous function for use in a SOAC.

data IfSort

What kind of branch is this? This has no semantic meaning, but provides hints to simplifications.

Constructors

IfNormal
An ordinary branch.

IfFallback
A branch where the "true" case is what we are actually interested in, and the "false" case is only present as a fallback for when the true case cannot be safely evaluated. The compiler is permitted to optimise away the branch if the true case contains only safe statements.

IfEquiv
Both of these branches are semantically equivalent, and it is fine to eliminate one if it turns out to have problems (e.g. contain things we cannot generate code for).

data IfDec rt

Data associated with a branch.

Constructors

IfDec

Fields

ifReturns :: [rt]

ifSort :: IfSort

data LoopForm rep

For-loop or while-loop?

Constructors

ForLoop VName IntType SubExp [(LParam rep, VName)]

WhileLoop VName

data ExpT rep

The root Futhark expression type. The Op constructor contains a rep-specific operation. Do-loops, branches and function calls are special. Everything else is a simple BasicOp.

Constructors

BasicOp BasicOp
A simple (non-recursive) operation.

Apply Name [(SubExp, Diet)] [RetType rep] (Safety, SrcLoc, [SrcLoc])

If SubExp (BodyT rep) (BodyT rep) (IfDec (BranchType rep))

DoLoop [(FParam rep, SubExp)] (LoopForm rep) (BodyT rep)
loop {a} = {v} (for i < n|while b) do b.

WithAcc [WithAccInput rep] (Lambda rep)
Create accumulators backed by the given arrays (which are consumed) and pass them to the lambda, which must return the updated accumulators and possibly some extra values. The accumulators are turned back into arrays. The Shape is the write index space. The corresponding arrays must all have this shape outermost. This construct is not part of BasicOp because we need the rep parameter.

Op (Op rep)

type WithAccInput rep = (Shape, [VName], Maybe (Lambda rep, [SubExp]))

The input to a WithAcc construct. Comprises the index space of the accumulator, the underlying arrays, and possibly a combining function.

data BasicOp

A primitive operation that returns something of known size and does not itself contain any bindings.

Constructors

SubExp SubExp
A variable or constant.

Opaque OpaqueOp SubExp
Semantically and operationally just identity, but is invisible/impenetrable to optimisations (hopefully). This partially a hack to avoid optimisation (so, to work around compiler limitations), but is also used to implement tracing and other operations that are semantically invisible, but have some sort of effect (brrr).

ArrayLit [SubExp] Type
Array literals, e.g., [ [1+x, 3], [2, 1+4] ]. Second arg is the element type of the rows of the array.

UnOp UnOp SubExp
Unary operation.

BinOp BinOp SubExp SubExp
Binary operation.

CmpOp CmpOp SubExp SubExp
Comparison - result type is always boolean.

ConvOp ConvOp SubExp
Conversion "casting".

Assert SubExp (ErrorMsg SubExp) (SrcLoc, [SrcLoc])
Turn a boolean into a certificate, halting the program with the given error message if the boolean is false.

Index VName (Slice SubExp)
The certificates for bounds-checking are part of the Stm.

Update Safety VName (Slice SubExp) SubExp
An in-place update of the given array at the given position. Consumes the array. If Safe, perform a run-time bounds check and ignore the write if out of bounds (like Scatter).

FlatIndex VName (FlatSlice SubExp)

FlatUpdate VName (FlatSlice SubExp) VName

Concat Int VName [VName] SubExp
concat0([1],[2, 3, 4]) = [1, 2, 3, 4]@.

Copy VName
Copy the given array. The result will not alias anything.

Manifest [Int] VName
Manifest an array with dimensions represented in the given order. The result will not alias anything.

Iota SubExp SubExp SubExp IntType
iota(n, x, s) = [x,x+s,..,x+(n-1)*s].

The IntType indicates the type of the array returned and the offset/stride arguments, but not the length argument.

Replicate Shape SubExp
replicate([3][2],1) = [[1,1], [1,1], [1,1]]

Scratch PrimType [SubExp]
Create array of given type and shape, with undefined elements.

Reshape (ShapeChange SubExp) VName
1st arg is the new shape, 2nd arg is the input array *)

Rearrange [Int] VName
Permute the dimensions of the input array. The list of integers is a list of dimensions (0-indexed), which must be a permutation of [0,n-1], where n is the number of dimensions in the input array.

Rotate [SubExp] VName
Rotate the dimensions of the input array. The list of subexpressions specify how much each dimension is rotated. The length of this list must be equal to the rank of the array.

UpdateAcc VName [SubExp] [SubExp]
Update an accumulator at the given index with the given value. Consumes the accumulator and produces a new one.

data OpaqueOp

Apart from being Opaque, what else is going on here?

Constructors

OpaqueNil
No special operation.

OpaqueTrace String
Print the argument, prefixed by this string.

type ShapeChange d = [DimChange d]

A list of DimChanges, indicating the new dimensions of an array.

data DimChange d

The new dimension in a Reshape-like operation. This allows us to disambiguate "real" reshapes, that change the actual shape of the array, from type coercions that are just present to make the types work out. The two constructors are considered equal for purposes of Eq.

Constructors

DimCoercion d
The new dimension is guaranteed to be numerically equal to the old one.

DimNew d
The new dimension is not necessarily numerically equal to the old one.

data BodyT rep

A body consists of a number of bindings, terminating in a result (essentially a tuple literal).

type Result = [SubExpRes]

The result of a body is a sequence of subexpressions.

data SubExpRes

A pairing of a subexpression and some certificates.

Constructors

SubExpRes

Fields

resCerts :: Certs

resSubExp :: SubExp

type Stms rep = Seq (Stm rep)

A sequence of statements.

pattern Let :: Pat rep -> StmAux (ExpDec rep) -> Exp rep -> Stm rep

stmExp :: Stm rep -> Exp rep

Expression.

stmPat :: Stm rep -> Pat rep

Pat.

stmAux :: Stm rep -> StmAux (ExpDec rep)

Auxiliary information statement.

data StmAux dec

Auxilliary Information associated with a statement.

Constructors

StmAux

Fields

stmAuxCerts :: !Certs

stmAuxAttrs :: Attrs

stmAuxDec :: dec

data PatT dec

A pattern is conceptually just a list of names and their types.

oneStm :: Stm rep -> Stms rep

A single statement.

stmsFromList :: [Stm rep] -> Stms rep

Convert a statement list to a statement sequence.

stmsToList :: Stms rep -> [Stm rep]

Convert a statement sequence to a statement list.

stmsHead :: Stms rep -> Maybe (Stm rep, Stms rep)

The first statement in the sequence, if any.

stmsLast :: Stms lore -> Maybe (Stms lore, Stm lore)

The last statement in the sequence, if any.

subExpRes :: SubExp -> SubExpRes

Construct a SubExpRes with no certificates.

varRes :: VName -> SubExpRes

Construct a SubExpRes from a variable name.

subExpsRes :: [SubExp] -> Result

Construct a Result from subexpressions.

varsRes :: [VName] -> Result

Construct a Result from variable names.

subExpResVName :: SubExpRes -> Maybe VName

The VName of a SubExpRes, if it exists.

module Futhark.IR.SOACS.SOAC

data LambdaT rep

Anonymous function for use in a SOAC.

Constructors

Lambda [LParam rep] (BodyT rep) [Type]

data BodyT rep

A body consists of a number of bindings, terminating in a result (essentially a tuple literal).

Constructors

Body (BodyDec rep) (Stms rep) Result

newtype PatT dec

A pattern is conceptually just a list of names and their types.

Constructors

Pat [PatElemT dec]

data PatElemT dec

An element of a pattern - consisting of a name and an addditional parametric decoration. This decoration is what is expected to contain the type of the resulting variable.

Constructors

PatElem VName dec