Futhark.IR.SOACS.Simplify

Documentation

simplifySOACS :: Prog SOACS -> PassM (Prog SOACS)

simplifyLambda :: (HasScope SOACS m, MonadFreshNames m) => Lambda -> m Lambda

simplifyFun :: MonadFreshNames m => SymbolTable (Wise SOACS) -> FunDef SOACS -> m (FunDef SOACS)

simplifyStms :: (HasScope SOACS m, MonadFreshNames m) => Stms SOACS -> m (Stms SOACS)

simplifyConsts :: MonadFreshNames m => Stms SOACS -> m (Stms SOACS)

simpleSOACS :: SimpleOps SOACS

simplifySOAC :: SimplifiableRep rep => SimplifyOp rep (SOAC (Wise rep))

soacRules :: RuleBook (Wise SOACS)

class HasSOAC rep where

Does this rep contain SOACs in its Ops? A rep must be an instance of this class for the simplification rules to work.

Methods

asSOAC :: Op rep -> Maybe (SOAC rep)

soacOp :: SOAC rep -> Op rep

simplifyKnownIterationSOAC :: (Buildable rep, BuilderOps rep, HasSOAC rep) => TopDownRuleOp rep

removeReplicateMapping :: (Aliased rep, Buildable rep, BuilderOps rep, HasSOAC rep) => TopDownRuleOp rep

Remove all arguments to the map that are simply replicates. These can be turned into free variables instead.

removeUnusedSOACInput :: forall rep. (Aliased rep, Buildable rep, BuilderOps rep, HasSOAC rep) => TopDownRuleOp rep

Remove inputs that are not used inside the SOAC.

liftIdentityMapping :: forall rep. (Buildable rep, BuilderOps rep, HasSOAC rep) => TopDownRuleOp rep

simplifyMapIota :: forall rep. (Buildable rep, BuilderOps rep, HasSOAC rep) => TopDownRuleOp rep

data SOACS

The rep for the basic representation.

Orphan instances

TraverseOpStms (Wise SOACS)

Instance details

Methods

traverseOpStms :: Monad m => OpStmsTraverser m (Op (Wise SOACS)) (Wise SOACS)

BuilderOps (Wise SOACS)

Instance details

Methods

mkExpDecB :: (MonadBuilder m, Rep m ~ Wise SOACS) => Pat (Wise SOACS) -> Exp (Wise SOACS) -> m (ExpDec (Wise SOACS))

mkBodyB :: (MonadBuilder m, Rep m ~ Wise SOACS) => Stms (Wise SOACS) -> Result -> m (Body (Wise SOACS))

mkLetNamesB :: (MonadBuilder m, Rep m ~ Wise SOACS) => [VName] -> Exp (Wise SOACS) -> m (Stm (Wise SOACS))