Futhark.IR.SegOp

Description

Segmented operations. These correspond to perfect map nests on top of something, except that the maps are conceptually only over iotas (so there will be explicit indexing inside them).

Documentation

data SegOp lvl rep

A SegOp is semantically a perfectly nested stack of maps, on top of some bottommost computation (scalar computation, reduction, scan, or histogram). The SegSpace encodes the original map structure.

All SegOps are parameterised by the representation of their body, as well as a *level*. The *level* is a representation-specific bit of information. For example, in GPU backends, it is used to indicate whether the SegOp is expected to run at the thread-level or the group-level.

Constructors

SegMap lvl SegSpace [Type] (KernelBody rep)

SegRed lvl SegSpace [SegBinOp rep] [Type] (KernelBody rep)
The KernelSpace must always have at least two dimensions, implying that the result of a SegRed is always an array.

SegScan lvl SegSpace [SegBinOp rep] [Type] (KernelBody rep)

SegHist lvl SegSpace [HistOp rep] [Type] (KernelBody rep)

data SegVirt

Do we need group-virtualisation when generating code for the segmented operation? In most cases, we do, but for some simple kernels, we compute the full number of groups in advance, and then virtualisation is an unnecessary (but generally very small) overhead. This only really matters for fairly trivial but very wide map kernels where each thread performs constant-time work on scalars.

Constructors

SegVirt

SegNoVirt

SegNoVirtFull
Not only do we not need virtualisation, but we _guarantee_ that all physical threads participate in the work. This can save some checks in code generation.

segLevel :: SegOp lvl rep -> lvl

The level of a SegOp.

segBody :: SegOp lvl rep -> KernelBody rep

The body of a SegOp.

segSpace :: SegOp lvl rep -> SegSpace

The space of a SegOp.

typeCheckSegOp :: Checkable rep => (lvl -> TypeM rep ()) -> SegOp lvl (Aliases rep) -> TypeM rep ()

Type check a SegOp, given a checker for its level.

data SegSpace

Index space of a SegOp.

Constructors

SegSpace

Fields

segFlat :: VName

Flat physical index corresponding to the dimensions (at code generation used for a thread ID or similar).

unSegSpace :: [(VName, SubExp)]

scopeOfSegSpace :: SegSpace -> Scope rep

A Scope containing all the identifiers brought into scope by this SegSpace.

segSpaceDims :: SegSpace -> [SubExp]

The sizes spanned by the indexes of the SegSpace.

Details

data HistOp rep

An operator for SegHist.

Constructors

HistOp

Fields

histShape :: Shape

histRaceFactor :: SubExp

histDest :: [VName]

histNeutral :: [SubExp]

histOpShape :: Shape

In case this operator is semantically a vectorised operator (corresponding to a perfect map nest in the SOACS representation), these are the logical "dimensions". This is used to generate more efficient code.

histOp :: Lambda rep

histType :: HistOp rep -> [Type]

The type of a histogram produced by a HistOp. This can be different from the type of the histDests in case we are dealing with a segmented histogram.

splitHistResults :: [HistOp rep] -> [SubExp] -> [([SubExp], [SubExp])]

Split reduction results returned by a KernelBody into those that correspond to indexes for the HistOps, and those that correspond to value.

data SegBinOp rep

An operator for SegScan and SegRed.

Constructors

SegBinOp

Fields

segBinOpComm :: Commutativity

segBinOpLambda :: Lambda rep

segBinOpNeutral :: [SubExp]

segBinOpShape :: Shape

In case this operator is semantically a vectorised operator (corresponding to a perfect map nest in the SOACS representation), these are the logical "dimensions". This is used to generate more efficient code.

segBinOpResults :: [SegBinOp rep] -> Int

How many reduction results are produced by these SegBinOps?

segBinOpChunks :: [SegBinOp rep] -> [a] -> [[a]]

Split some list into chunks equal to the number of values returned by each SegBinOp

data KernelBody rep

The body of a SegOp.

Constructors

KernelBody

Fields

kernelBodyDec :: BodyDec rep

kernelBodyStms :: Stms rep

kernelBodyResult :: [KernelResult]

aliasAnalyseKernelBody :: (ASTRep rep, CanBeAliased (Op rep)) => AliasTable -> KernelBody rep -> KernelBody (Aliases rep)

Perform alias analysis on a KernelBody.

consumedInKernelBody :: Aliased rep => KernelBody rep -> Names

The variables consumed in the kernel body.

data ResultManifest

Metadata about whether there is a subtle point to this KernelResult. This is used to protect things like tiling, which might otherwise be removed by the simplifier because they're semantically redundant. This has no semantic effect and can be ignored at code generation.

Constructors

ResultNoSimplify
Don't simplify this one!

ResultMaySimplify
Go nuts.

ResultPrivate
The results produced are only used within the same physical thread later on, and can thus be kept in registers.

data KernelResult

A KernelBody does not return an ordinary Result. Instead, it returns a list of these.

Constructors

Returns ResultManifest Certs SubExp
Each "worker" in the kernel returns this. Whether this is a result-per-thread or a result-per-group depends on where the SegOp occurs.

WriteReturns Certs Shape VName [(Slice SubExp, SubExp)]

ConcatReturns Certs SplitOrdering SubExp SubExp VName

TileReturns Certs [(SubExp, SubExp)] VName

RegTileReturns Certs [(SubExp, SubExp, SubExp)] VName

kernelResultCerts :: KernelResult -> Certs

Get the certs for this KernelResult.

kernelResultSubExp :: KernelResult -> SubExp

Get the root SubExp corresponding values for a KernelResult.

data SplitOrdering

How an array is split into chunks.

Constructors

SplitContiguous

SplitStrided SubExp

Generic traversal

data SegOpMapper lvl frep trep m

Like Mapper, but just for SegOps.

Constructors

SegOpMapper

Fields

mapOnSegOpSubExp :: SubExp -> m SubExp

mapOnSegOpLambda :: Lambda frep -> m (Lambda trep)

mapOnSegOpBody :: KernelBody frep -> m (KernelBody trep)

mapOnSegOpVName :: VName -> m VName

mapOnSegOpLevel :: lvl -> m lvl

identitySegOpMapper :: Monad m => SegOpMapper lvl rep rep m

A mapper that simply returns the SegOp verbatim.

mapSegOpM :: (Applicative m, Monad m) => SegOpMapper lvl frep trep m -> SegOp lvl frep -> m (SegOp lvl trep)

Apply a SegOpMapper to the given SegOp.

traverseSegOpStms :: Monad m => OpStmsTraverser m (SegOp lvl rep) rep

A helper for defining TraverseOpStms.

Simplification

simplifySegOp :: (SimplifiableRep rep, BodyDec rep ~ (), Simplifiable lvl) => SegOp lvl (Wise rep) -> SimpleM rep (SegOp lvl (Wise rep), Stms (Wise rep))

Simplify the given SegOp.

class HasSegOp rep where

Does this rep contain SegOps in its Ops? A rep must be an instance of this class for the simplification rules to work.

Associated Types

type SegOpLevel rep

Methods

asSegOp :: Op rep -> Maybe (SegOp (SegOpLevel rep) rep)

segOp :: SegOp (SegOpLevel rep) rep -> Op rep

segOpRules :: (HasSegOp rep, BuilderOps rep, Buildable rep) => RuleBook rep

Simplification rules for simplifying SegOps.

Memory

segOpReturns :: (Mem rep inner, Monad m, HasScope rep m) => SegOp lvl somerep -> m [ExpReturns]

Like segOpType, but for memory representations.