Futhark.IR.Syntax

Description

Definition of the Futhark core language IR

For actually constructing ASTs, see Futhark.Construct.

Types and values

The core language type system is much more restricted than the core language. This is a theme that repeats often. The only types that are supported in the core language are various primitive types PrimType which can be combined in arrays (ignore Mem for now). Types are represented as TypeBase, which is parameterised by the shape of the array and whether we keep uniqueness information. The Type alias, which is the most commonly used, uses Shape and NoUniqueness.

This means that the records, tuples, and sum types of the source language are represented merely as collections of primitives and arrays. This is implemented in Futhark.Internalise, but the specifics are not important for writing passes on the core language. What is important is that many constructs that conceptually return tuples instead return multiple values. This is not merely syntactic sugar for a tuple: each of those values are eventually bound to distinct variables. The prettyprinter for the IR will typically print such collections of values or types in curly braces.

The system of primitive types is interesting in itself. See Futhark.IR.Primitive.

Overall AST design

Internally, the Futhark compiler core intermediate representation resembles a traditional compiler for an imperative language more than it resembles, say, a Haskell or ML compiler. All functions are monomorphic (except for sizes), first-order, and defined at the top level. Notably, the IR does not use continuation-passing style (CPS) at any time. Instead it uses Administrative Normal Form (ANF), where all subexpressions SubExp are either constants PrimValue or variables VName. Variables are represented as a human-readable Name (which doesn't matter to the compiler) as well as a numeric tag, which is what the compiler actually looks at. All variable names when prettyprinted are of the form foo_123. Function names are just Names, though.

The body of a function (FunDef) is a Body, which consists of a sequence of statements (Stms) and a Result. Execution of a Body consists of executing all of the statements, then returning the values of the variables indicated by the result.

A statement (Stm) consists of a Pat alongside an expression ExpT. A pattern is a sequence of name/type pairs.

For example, the source language expression let z = x + y - 1 in z would in the core language be represented (in prettyprinted form) as something like:

let {a_12} = x_10 + y_11
let {b_13} = a_12 - 1
in {b_13}

Representations

Most AST types (Stm, ExpT, Prog, etc) are parameterised by a type parameter rep. The representation specifies how to fill out various polymorphic parts of the AST. For example, ExpT has a constructor Op whose payload depends on rep, via the use of a type family called Op (a kind of type-level function) which is applied to the rep. The SOACS representation (Futhark.IR.SOACS) thus uses a rep called SOACS, and defines that Op SOACS is a SOAC, while the Kernels representation (Futhark.IR.Kernels) defines Op Kernels as some kind of kernel construct. Similarly, various other decorations (e.g. what information we store in a PatElemT) are also type families.

The full list of possible decorations is defined as part of the type class RepTypes (although other type families are also used elsewhere in the compiler on an ad hoc basis).

Essentially, the rep type parameter functions as a kind of proxy, saving us from having to parameterise the AST type with all the different forms of decorations that we desire (it would easily become a type with a dozen type parameters).

Defining a new representation (or rep) thus requires you to define an empty datatype and implement a handful of type class instances for it. See the source of Futhark.IR.Seq for what is likely the simplest example.

Documentation

module Language.Futhark.Core

pretty :: Pretty a => a -> String

Prettyprint a value, wrapped to 80 characters.

module Futhark.IR.Rep

module Futhark.IR.Syntax.Core

Types

data Uniqueness

The uniqueness attribute of a type. This essentially indicates whether or not in-place modifications are acceptable. With respect to ordering, Unique is greater than Nonunique.

Constructors

Nonunique
May have references outside current function.

Unique
No references outside current function.

data NoUniqueness

A fancier name for () - encodes no uniqueness information.

Constructors

NoUniqueness

newtype Rank

The size of an array type as merely the number of dimensions, with no further information.

Constructors

Rank Int

class (Monoid a, Eq a, Ord a) => ArrayShape a where

A class encompassing types containing array shape information.

Methods

shapeRank :: a -> Int

Return the rank of an array with the given size.

subShapeOf :: a -> a -> Bool

Check whether one shape if a subset of another shape.

data Space

The memory space of a block. If DefaultSpace, this is the "default" space, whatever that is. The exact meaning of the SpaceId depends on the backend used. In GPU kernels, for example, this is used to distinguish between constant, global and shared memory spaces. In GPU-enabled host code, it is used to distinguish between host memory (DefaultSpace) and GPU space.

Constructors

DefaultSpace

Space SpaceId

ScalarSpace [SubExp] PrimType
A special kind of memory that is a statically sized array of some primitive type. Used for private memory on GPUs.

data TypeBase shape u

The type of a value. When comparing types for equality with ==, shapes must match.

Constructors

Prim PrimType

Acc VName Shape [Type] u
Token, index space, element type, and uniqueness.

Array PrimType shape u

Mem Space

data Diet

Information about which parts of a value/type are consumed. For example, we might say that a function taking three arguments of types ([int], *[int], [int]) has diet [Observe, Consume, Observe].

Constructors

Consume
Consumes this value.

Observe
Only observes value in this position, does not consume. A result may alias this.

ObservePrim
As Observe, but the result will not alias, because the parameter does not carry aliases.

Abstract syntax tree

data Ident

An identifier consists of its name and the type of the value bound to the identifier.

Constructors

Ident

Fields

identName :: VName

identType :: Type

data SubExp

A subexpression is either a scalar constant or a variable. One important property is that evaluation of a subexpression is guaranteed to complete in constant time.

Constructors

Constant PrimValue

Var VName

type PatElem rep = PatElemT (LetDec rep)

A type alias for namespace control.

data PatElemT dec

An element of a pattern - consisting of a name and an addditional parametric decoration. This decoration is what is expected to contain the type of the resulting variable.

Constructors

PatElem

Fields

patElemName :: VName

The name being bound.

patElemDec :: dec

Pat element decoration.

newtype PatT dec

A pattern is conceptually just a list of names and their types.

Constructors

Pat

Fields

patElems :: [PatElemT dec]

type Pat rep = PatT (LetDec rep)

A type alias for namespace control.

data StmAux dec

Auxilliary Information associated with a statement.

Constructors

StmAux

Fields

stmAuxCerts :: !Certs

stmAuxAttrs :: Attrs

stmAuxDec :: dec

data Stm rep

A local variable binding.

Constructors

Let

Fields

stmPat :: Pat rep

Pat.

stmAux :: StmAux (ExpDec rep)

Auxiliary information statement.

stmExp :: Exp rep

Expression.

type Stms rep = Seq (Stm rep)

A sequence of statements.

data SubExpRes

A pairing of a subexpression and some certificates.

Constructors

SubExpRes

Fields

resCerts :: Certs

resSubExp :: SubExp

type Result = [SubExpRes]

The result of a body is a sequence of subexpressions.

data BodyT rep

A body consists of a number of bindings, terminating in a result (essentially a tuple literal).

Constructors

Body

Fields

bodyDec :: BodyDec rep

bodyStms :: Stms rep

bodyResult :: Result

type Body = BodyT

Type alias for namespace reasons.

data BasicOp

A primitive operation that returns something of known size and does not itself contain any bindings.

Constructors

SubExp SubExp
A variable or constant.

Opaque OpaqueOp SubExp
Semantically and operationally just identity, but is invisible/impenetrable to optimisations (hopefully). This partially a hack to avoid optimisation (so, to work around compiler limitations), but is also used to implement tracing and other operations that are semantically invisible, but have some sort of effect (brrr).

ArrayLit [SubExp] Type
Array literals, e.g., [ [1+x, 3], [2, 1+4] ]. Second arg is the element type of the rows of the array.

UnOp UnOp SubExp
Unary operation.

BinOp BinOp SubExp SubExp
Binary operation.

CmpOp CmpOp SubExp SubExp
Comparison - result type is always boolean.

ConvOp ConvOp SubExp
Conversion "casting".

Assert SubExp (ErrorMsg SubExp) (SrcLoc, [SrcLoc])
Turn a boolean into a certificate, halting the program with the given error message if the boolean is false.

Index VName (Slice SubExp)
The certificates for bounds-checking are part of the Stm.

Update Safety VName (Slice SubExp) SubExp
An in-place update of the given array at the given position. Consumes the array. If Safe, perform a run-time bounds check and ignore the write if out of bounds (like Scatter).

FlatIndex VName (FlatSlice SubExp)

FlatUpdate VName (FlatSlice SubExp) VName

Concat Int VName [VName] SubExp
concat0([1],[2, 3, 4]) = [1, 2, 3, 4]@.

Copy VName
Copy the given array. The result will not alias anything.

Manifest [Int] VName
Manifest an array with dimensions represented in the given order. The result will not alias anything.

Iota SubExp SubExp SubExp IntType
iota(n, x, s) = [x,x+s,..,x+(n-1)*s].

The IntType indicates the type of the array returned and the offset/stride arguments, but not the length argument.

Replicate Shape SubExp
replicate([3][2],1) = [[1,1], [1,1], [1,1]]

Scratch PrimType [SubExp]
Create array of given type and shape, with undefined elements.

Reshape (ShapeChange SubExp) VName
1st arg is the new shape, 2nd arg is the input array *)

Rearrange [Int] VName
Permute the dimensions of the input array. The list of integers is a list of dimensions (0-indexed), which must be a permutation of [0,n-1], where n is the number of dimensions in the input array.

Rotate [SubExp] VName
Rotate the dimensions of the input array. The list of subexpressions specify how much each dimension is rotated. The length of this list must be equal to the rank of the array.

UpdateAcc VName [SubExp] [SubExp]
Update an accumulator at the given index with the given value. Consumes the accumulator and produces a new one.

data UnOp

Various unary operators. It is a bit ad-hoc what is a unary operator and what is a built-in function. Perhaps these should all go away eventually.

Constructors

Not
E.g., ! True == False.

Complement IntType
E.g., ~(~1) = 1.

Abs IntType
abs(-2) = 2.

FAbs FloatType
fabs(-2.0) = 2.0.

SSignum IntType
Signed sign function: ssignum(-2) = -1.

USignum IntType
Unsigned sign function: usignum(2) = 1.

FSignum FloatType
Floating-point sign function.

data BinOp

Binary operators. These correspond closely to the binary operators in LLVM. Most are parametrised by their expected input and output types.

Constructors

Add IntType Overflow
Integer addition.

FAdd FloatType
Floating-point addition.

Sub IntType Overflow
Integer subtraction.

FSub FloatType
Floating-point subtraction.

Mul IntType Overflow
Integer multiplication.

FMul FloatType
Floating-point multiplication.

UDiv IntType Safety
Unsigned integer division. Rounds towards negativity infinity. Note: this is different from LLVM.

UDivUp IntType Safety
Unsigned integer division. Rounds towards positive infinity.

SDiv IntType Safety
Signed integer division. Rounds towards negativity infinity. Note: this is different from LLVM.

SDivUp IntType Safety
Signed integer division. Rounds towards positive infinity.

FDiv FloatType
Floating-point division.

FMod FloatType
Floating-point modulus.

UMod IntType Safety
Unsigned integer modulus; the countepart to UDiv.

SMod IntType Safety
Signed integer modulus; the countepart to SDiv.

SQuot IntType Safety
Signed integer division. Rounds towards zero. This corresponds to the sdiv instruction in LLVM and integer division in C.

SRem IntType Safety
Signed integer division. Rounds towards zero. This corresponds to the srem instruction in LLVM and integer modulo in C.

SMin IntType
Returns the smallest of two signed integers.

UMin IntType
Returns the smallest of two unsigned integers.

FMin FloatType
Returns the smallest of two floating-point numbers.

SMax IntType
Returns the greatest of two signed integers.

UMax IntType
Returns the greatest of two unsigned integers.

FMax FloatType
Returns the greatest of two floating-point numbers.

Shl IntType
Left-shift.

LShr IntType
Logical right-shift, zero-extended.

AShr IntType
Arithmetic right-shift, sign-extended.

And IntType
Bitwise and.

Or IntType
Bitwise or.

Xor IntType
Bitwise exclusive-or.

Pow IntType
Integer exponentiation.

FPow FloatType
Floating-point exponentiation.

LogAnd
Boolean and - not short-circuiting.

LogOr
Boolean or - not short-circuiting.

data CmpOp

Comparison operators are like BinOps, but they always return a boolean value. The somewhat ugly constructor names are straight out of LLVM.

Constructors

CmpEq PrimType
All types equality.

CmpUlt IntType
Unsigned less than.

CmpUle IntType
Unsigned less than or equal.

CmpSlt IntType
Signed less than.

CmpSle IntType
Signed less than or equal.

FCmpLt FloatType
Floating-point less than.

FCmpLe FloatType
Floating-point less than or equal.

CmpLlt
Boolean less than.

CmpLle
Boolean less than or equal.

data ConvOp

Conversion operators try to generalise the from t0 x to t1 instructions from LLVM.

Constructors

ZExt IntType IntType
Zero-extend the former integer type to the latter. If the new type is smaller, the result is a truncation.

SExt IntType IntType
Sign-extend the former integer type to the latter. If the new type is smaller, the result is a truncation.

FPConv FloatType FloatType
Convert value of the former floating-point type to the latter. If the new type is smaller, the result is a truncation.

FPToUI FloatType IntType
Convert a floating-point value to the nearest unsigned integer (rounding towards zero).

FPToSI FloatType IntType
Convert a floating-point value to the nearest signed integer (rounding towards zero).

UIToFP IntType FloatType
Convert an unsigned integer to a floating-point value.

SIToFP IntType FloatType
Convert a signed integer to a floating-point value.

IToB IntType
Convert an integer to a boolean value. Zero becomes false; anything else is true.

BToI IntType
Convert a boolean to an integer. True is converted to 1 and False to 0.

data OpaqueOp

Apart from being Opaque, what else is going on here?

Constructors

OpaqueNil
No special operation.

OpaqueTrace String
Print the argument, prefixed by this string.

data DimChange d

The new dimension in a Reshape-like operation. This allows us to disambiguate "real" reshapes, that change the actual shape of the array, from type coercions that are just present to make the types work out. The two constructors are considered equal for purposes of Eq.

Constructors

DimCoercion d
The new dimension is guaranteed to be numerically equal to the old one.

DimNew d
The new dimension is not necessarily numerically equal to the old one.

type ShapeChange d = [DimChange d]

A list of DimChanges, indicating the new dimensions of an array.

type WithAccInput rep = (Shape, [VName], Maybe (Lambda rep, [SubExp]))

The input to a WithAcc construct. Comprises the index space of the accumulator, the underlying arrays, and possibly a combining function.

data ExpT rep

The root Futhark expression type. The Op constructor contains a rep-specific operation. Do-loops, branches and function calls are special. Everything else is a simple BasicOp.

Constructors

BasicOp BasicOp
A simple (non-recursive) operation.

Apply Name [(SubExp, Diet)] [RetType rep] (Safety, SrcLoc, [SrcLoc])

If SubExp (BodyT rep) (BodyT rep) (IfDec (BranchType rep))

DoLoop [(FParam rep, SubExp)] (LoopForm rep) (BodyT rep)
loop {a} = {v} (for i < n|while b) do b.

WithAcc [WithAccInput rep] (Lambda rep)
Create accumulators backed by the given arrays (which are consumed) and pass them to the lambda, which must return the updated accumulators and possibly some extra values. The accumulators are turned back into arrays. The Shape is the write index space. The corresponding arrays must all have this shape outermost. This construct is not part of BasicOp because we need the rep parameter.

Op (Op rep)

type Exp = ExpT

A type alias for namespace control.

data LoopForm rep

For-loop or while-loop?

Constructors

ForLoop VName IntType SubExp [(LParam rep, VName)]

WhileLoop VName

data IfDec rt

Data associated with a branch.

Constructors

IfDec

Fields

ifReturns :: [rt]

ifSort :: IfSort

data IfSort

What kind of branch is this? This has no semantic meaning, but provides hints to simplifications.

Constructors

IfNormal
An ordinary branch.

IfFallback
A branch where the "true" case is what we are actually interested in, and the "false" case is only present as a fallback for when the true case cannot be safely evaluated. The compiler is permitted to optimise away the branch if the true case contains only safe statements.

IfEquiv
Both of these branches are semantically equivalent, and it is fine to eliminate one if it turns out to have problems (e.g. contain things we cannot generate code for).

data Safety

Whether something is safe or unsafe (mostly function calls, and in the context of whether operations are dynamically checked). When we inline an Unsafe function, we remove all safety checks in its body. The Ord instance picks Unsafe as being less than Safe.

For operations like integer division, a safe division will not explode the computer in case of division by zero, but instead return some unspecified value. This always involves a run-time check, so generally the unsafe variant is what the compiler will insert, but guarded by an explicit assertion elsewhere. Safe operations are useful when the optimiser wants to move e.g. a division to a location where the divisor may be zero, but where the result will only be used when it is non-zero (so it doesn't matter what result is provided with a zero divisor, as long as the program keeps running).

Constructors

Unsafe

Safe

data LambdaT rep

Anonymous function for use in a SOAC.

Constructors

Lambda

Fields

lambdaParams :: [LParam rep]

lambdaBody :: BodyT rep

lambdaReturnType :: [Type]

type Lambda = LambdaT

Type alias for namespacing reasons.

Definitions

data Param dec

A function or lambda parameter.

Constructors

Param

Fields

paramAttrs :: Attrs

Attributes of the parameter. When constructing a parameter, feel free to just pass mempty.

paramName :: VName

Name of the parameter.

paramDec :: dec

Function parameter decoration.

type FParam rep = Param (FParamInfo rep)

A function and loop parameter.

type LParam rep = Param (LParamInfo rep)

A lambda parameter.

data FunDef rep

Function Declarations

Constructors

FunDef

Fields

funDefEntryPoint :: Maybe EntryPoint

Contains a value if this function is an entry point.

funDefAttrs :: Attrs

funDefName :: Name

funDefRetType :: [RetType rep]

funDefParams :: [FParam rep]

funDefBody :: BodyT rep

type EntryPoint = (Name, [EntryParam], [EntryPointType])

Information about the inputs and outputs (return value) of an entry point.

data EntryParam

An entry point parameter, comprising its name and original type.

Constructors

EntryParam

Fields

entryParamName :: Name

entryParamType :: EntryPointType

data EntryPointType

Every entry point argument and return value has an annotation indicating how it maps to the original source program type.

Constructors

TypeUnsigned Uniqueness
Is an unsigned integer or array of unsigned integers.

TypeOpaque Uniqueness String Int
A black box type comprising this many core values. The string is a human-readable description with no other semantics.

TypeDirect Uniqueness
Maps directly.

data Prog rep

An entire Futhark program.

Constructors

Prog

Fields

progConsts :: Stms rep

Top-level constants that are computed at program startup, and which are in scope inside all functions.

progFuns :: [FunDef rep]

The functions comprising the program. All funtions are also available in scope in the definitions of the constants, so be careful not to introduce circular dependencies (not currently checked).

Utils

oneStm :: Stm rep -> Stms rep

A single statement.

stmsFromList :: [Stm rep] -> Stms rep

Convert a statement list to a statement sequence.

stmsToList :: Stms rep -> [Stm rep]

Convert a statement sequence to a statement list.

stmsHead :: Stms rep -> Maybe (Stm rep, Stms rep)

The first statement in the sequence, if any.

stmsLast :: Stms lore -> Maybe (Stms lore, Stm lore)

The last statement in the sequence, if any.

subExpRes :: SubExp -> SubExpRes

Construct a SubExpRes with no certificates.

subExpsRes :: [SubExp] -> Result

Construct a Result from subexpressions.

varRes :: VName -> SubExpRes

Construct a SubExpRes from a variable name.

varsRes :: [VName] -> Result

Construct a Result from variable names.

subExpResVName :: SubExpRes -> Maybe VName

The VName of a SubExpRes, if it exists.