Futhark.IR.Syntax.Core

Description

The most primitive ("core") aspects of the AST. Split out of Futhark.IR.Syntax in order for Futhark.IR.Rep to use these definitions. This module is re-exported from Futhark.IR.Syntax and there should be no reason to include it explicitly.

Documentation

module Language.Futhark.Core

module Futhark.IR.Primitive

Types

data Commutativity

Whether some operator is commutative or not. The Monoid instance returns the least commutative of its arguments.

Constructors

Noncommutative

Commutative

data Uniqueness

The uniqueness attribute of a type. This essentially indicates whether or not in-place modifications are acceptable. With respect to ordering, Unique is greater than Nonunique.

Constructors

Nonunique
May have references outside current function.

Unique
No references outside current function.

data NoUniqueness

A fancier name for () - encodes no uniqueness information.

Constructors

NoUniqueness

newtype ShapeBase d

The size of an array type as a list of its dimension sizes, with the type of sizes being parametric.

Constructors

Shape

Fields

shapeDims :: [d]

type Shape = ShapeBase SubExp

The size of an array as a list of subexpressions. If a variable, that variable must be in scope where this array is used.

stripDims :: Int -> ShapeBase d -> ShapeBase d

stripDims n shape strips the outer n dimensions from shape.

data Ext a

Something that may be existential.

Constructors

Ext Int

Free a

type ExtSize = Ext SubExp

The size of this dimension.

type ExtShape = ShapeBase ExtSize

Like Shape but some of its elements may be bound in a local environment instead. These are denoted with integral indices.

newtype Rank

The size of an array type as merely the number of dimensions, with no further information.

Constructors

Rank Int

class (Monoid a, Eq a, Ord a) => ArrayShape a where

A class encompassing types containing array shape information.

Methods

shapeRank :: a -> Int

Return the rank of an array with the given size.

subShapeOf :: a -> a -> Bool

Check whether one shape if a subset of another shape.

data Space

The memory space of a block. If DefaultSpace, this is the "default" space, whatever that is. The exact meaning of the SpaceId depends on the backend used. In GPU kernels, for example, this is used to distinguish between constant, global and shared memory spaces. In GPU-enabled host code, it is used to distinguish between host memory (DefaultSpace) and GPU space.

Constructors

DefaultSpace

Space SpaceId

ScalarSpace [SubExp] PrimType
A special kind of memory that is a statically sized array of some primitive type. Used for private memory on GPUs.

type SpaceId = String

A string representing a specific non-default memory space.

data TypeBase shape u

The type of a value. When comparing types for equality with ==, shapes must match.

Constructors

Prim PrimType

Acc VName Shape [Type] u
Token, index space, element type, and uniqueness.

Array PrimType shape u

Mem Space

type Type = TypeBase Shape NoUniqueness

A type with shape information, used for describing the type of variables.

type ExtType = TypeBase ExtShape NoUniqueness

A type with existentially quantified shapes - used as part of function (and function-like) return types. Generally only makes sense when used in a list.

type DeclType = TypeBase Shape Uniqueness

A type with shape and uniqueness information, used declaring return- and parameters types.

type DeclExtType = TypeBase ExtShape Uniqueness

An ExtType with uniqueness information, used for function return types.

data Diet

Information about which parts of a value/type are consumed. For example, we might say that a function taking three arguments of types ([int], *[int], [int]) has diet [Observe, Consume, Observe].

Constructors

Consume
Consumes this value.

Observe
Only observes value in this position, does not consume. A result may alias this.

ObservePrim
As Observe, but the result will not alias, because the parameter does not carry aliases.

newtype ErrorMsg a

An error message is a list of error parts, which are concatenated to form the final message.

Constructors

ErrorMsg [ErrorMsgPart a]

data ErrorMsgPart a

A part of an error message.

Constructors

ErrorString String
A literal string.

ErrorVal PrimType a
A run-time value.

errorMsgArgTypes :: ErrorMsg a -> [PrimType]

How many non-constant parts does the error message have, and what is their type?

Attributes

data Attr

A single attribute.

Constructors

AttrName Name

AttrInt Integer

AttrComp Name [Attr]

newtype Attrs

Every statement is associated with a set of attributes, which can have various effects throughout the compiler.

Constructors

Attrs

Fields

unAttrs :: Set Attr

oneAttr :: Attr -> Attrs

Construct Attrs from a single Attr.

inAttrs :: Attr -> Attrs -> Bool

Is the given attribute to be found in the attribute set?

withoutAttrs :: Attrs -> Attrs -> Attrs

x withoutAttrs y gives x except for any attributes also in y.

mapAttrs :: (Attr -> a) -> Attrs -> [a]

Map a function over an attribute set.

Values

data PrimValue

Non-array values.

Constructors

IntValue !IntValue

FloatValue !FloatValue

BoolValue !Bool

UnitValue
The only value of type Unit.

Abstract syntax tree

data Ident

An identifier consists of its name and the type of the value bound to the identifier.

Constructors

Ident

Fields

identName :: VName

identType :: Type

newtype Certs

A list of names used for certificates in some expressions.

Constructors

Certs

Fields

unCerts :: [VName]

data SubExp

A subexpression is either a scalar constant or a variable. One important property is that evaluation of a subexpression is guaranteed to complete in constant time.

Constructors

Constant PrimValue

Var VName

data Param dec

A function or lambda parameter.

Constructors

Param

Fields

paramAttrs :: Attrs

Attributes of the parameter. When constructing a parameter, feel free to just pass mempty.

paramName :: VName

Name of the parameter.

paramDec :: dec

Function parameter decoration.

data DimIndex d

How to index a single dimension of an array.

Constructors

DimFix d
Fix index in this dimension.

DimSlice d d d
DimSlice start_offset num_elems stride.

newtype Slice d

A list of DimIndexs, indicating how an array should be sliced. Whenever a function accepts a Slice, that slice should be total, i.e, cover all dimensions of the array. Deviators should be indicated by taking a list of DimIndexes instead.

Constructors

Slice

Fields

unSlice :: [DimIndex d]

dimFix :: DimIndex d -> Maybe d

If the argument is a DimFix, return its component.

sliceIndices :: Slice d -> Maybe [d]

If the slice is all DimFixs, return the components.

sliceDims :: Slice d -> [d]

The dimensions of the array produced by this slice.

unitSlice :: Num d => d -> d -> DimIndex d

A slice with a stride of one.

fixSlice :: Num d => Slice d -> [d] -> [d]

Fix the DimSlices of a slice. The number of indexes must equal the length of sliceDims for the slice.

sliceSlice :: Num d => Slice d -> Slice d -> Slice d

Further slice the DimSlices of a slice. The number of slices must equal the length of sliceDims for the slice.

data PatElemT dec

An element of a pattern - consisting of a name and an addditional parametric decoration. This decoration is what is expected to contain the type of the resulting variable.

Constructors

PatElem

Fields

patElemName :: VName

The name being bound.

patElemDec :: dec

Pat element decoration.

Flat (LMAD) slices

data FlatSlice d

Constructors

FlatSlice d [FlatDimIndex d]

data FlatDimIndex d

Constructors

FlatDimIndex

Fields

d

Number of elements in dimension

d

Stride of dimension

flatSliceDims :: FlatSlice d -> [d]

flatSliceStrides :: FlatSlice d -> [d]