Futhark.IR.TypeCheck

Description

The type checker checks whether the program is type-consistent.

Interface

checkProg :: Checkable rep => Prog (Aliases rep) -> Either (TypeError rep) ()

Type check a program containing arbitrary type information, yielding either a type error or a program with complete type information.

data TypeError rep

A type error.

Constructors

Error [String] (ErrorCase rep)

data ErrorCase rep

Information about an error during type checking. The Show instance for this type produces a human-readable description.

Constructors

TypeError String

UnexpectedType (Exp rep) Type [Type]

ReturnTypeError Name [ExtType] [ExtType]

DupDefinitionError Name

DupParamError Name VName

DupPatError VName

InvalidPatError (Pat (Aliases rep)) [ExtType] (Maybe String)

UnknownVariableError VName

UnknownFunctionError Name

ParameterMismatch (Maybe Name) [Type] [Type]

SlicingError Int Int

BadAnnotation String Type Type

ReturnAliased Name VName

UniqueReturnAliased Name

NotAnArray VName Type

PermutationError [Int] Int (Maybe VName)

Extensionality

data TypeM rep a

The type checker runs in this monad.

bad :: ErrorCase rep -> TypeM rep a

context :: String -> TypeM rep a -> TypeM rep a

Add information about what is being type-checked to the current context. Liberal use of this combinator makes it easier to track type errors, as the strings are added to type errors signalled via bad.

message :: Pretty a => String -> a -> String

class (ASTRep rep, CanBeAliased (Op rep), CheckableOp rep) => Checkable rep where

The class of representations that can be type-checked.

Minimal complete definition

Nothing

Methods

checkExpDec :: ExpDec rep -> TypeM rep ()

default checkExpDec :: ExpDec rep ~ () => ExpDec rep -> TypeM rep ()

checkBodyDec :: BodyDec rep -> TypeM rep ()

default checkBodyDec :: BodyDec rep ~ () => BodyDec rep -> TypeM rep ()

checkFParamDec :: VName -> FParamInfo rep -> TypeM rep ()

default checkFParamDec :: FParamInfo rep ~ DeclType => VName -> FParamInfo rep -> TypeM rep ()

checkLParamDec :: VName -> LParamInfo rep -> TypeM rep ()

default checkLParamDec :: LParamInfo rep ~ Type => VName -> LParamInfo rep -> TypeM rep ()

checkLetBoundDec :: VName -> LetDec rep -> TypeM rep ()

default checkLetBoundDec :: LetDec rep ~ Type => VName -> LetDec rep -> TypeM rep ()

checkRetType :: [RetType rep] -> TypeM rep ()

default checkRetType :: RetType rep ~ DeclExtType => [RetType rep] -> TypeM rep ()

matchPat :: Pat (Aliases rep) -> Exp (Aliases rep) -> TypeM rep ()

default matchPat :: Pat (Aliases rep) -> Exp (Aliases rep) -> TypeM rep ()

primFParam :: VName -> PrimType -> TypeM rep (FParam (Aliases rep))

default primFParam :: FParamInfo rep ~ DeclType => VName -> PrimType -> TypeM rep (FParam (Aliases rep))

matchReturnType :: [RetType rep] -> Result -> TypeM rep ()

default matchReturnType :: RetType rep ~ DeclExtType => [RetType rep] -> Result -> TypeM rep ()

matchBranchType :: [BranchType rep] -> Body (Aliases rep) -> TypeM rep ()

default matchBranchType :: BranchType rep ~ ExtType => [BranchType rep] -> Body (Aliases rep) -> TypeM rep ()

matchLoopResult :: [FParam (Aliases rep)] -> Result -> TypeM rep ()

default matchLoopResult :: FParamInfo rep ~ DeclType => [FParam (Aliases rep)] -> Result -> TypeM rep ()

class ASTRep rep => CheckableOp rep where

Methods

checkOp :: OpWithAliases (Op rep) -> TypeM rep ()

Used at top level; can be locally changed with checkOpWith.

lookupVar :: VName -> TypeM rep (NameInfo (Aliases rep))

lookupAliases :: Checkable rep => VName -> TypeM rep Names

checkOpWith :: (OpWithAliases (Op rep) -> TypeM rep ()) -> TypeM rep a -> TypeM rep a

Checkers

require :: Checkable rep => [Type] -> SubExp -> TypeM rep ()

require ts se causes a '(TypeError vn)' if the type of se is not a subtype of one of the types in ts.

requireI :: Checkable rep => [Type] -> VName -> TypeM rep ()

Variant of require working on variable names.

requirePrimExp :: Checkable rep => PrimType -> PrimExp VName -> TypeM rep ()

checkSubExp :: Checkable rep => SubExp -> TypeM rep Type

checkCerts :: Checkable rep => Certs -> TypeM rep ()

checkExp :: Checkable rep => Exp (Aliases rep) -> TypeM rep ()

checkStms :: Checkable rep => Stms (Aliases rep) -> TypeM rep a -> TypeM rep a

checkStm :: Checkable rep => Stm (Aliases rep) -> TypeM rep a -> TypeM rep a

checkType :: Checkable rep => TypeBase Shape u -> TypeM rep ()

checkExtType :: Checkable rep => TypeBase ExtShape u -> TypeM rep ()

matchExtPat :: Checkable rep => Pat (Aliases rep) -> [ExtType] -> TypeM rep ()

matchExtBranchType :: Checkable rep => [ExtType] -> Body (Aliases rep) -> TypeM rep ()

argType :: Arg -> Type

argAliases :: Arg -> Names

Remove all aliases from the Arg.

noArgAliases :: Arg -> Arg

checkArg :: Checkable rep => SubExp -> TypeM rep Arg

checkSOACArrayArgs :: Checkable rep => SubExp -> [VName] -> TypeM rep [Arg]

checkLambda :: Checkable rep => Lambda (Aliases rep) -> [Arg] -> TypeM rep ()

checkBody :: Checkable rep => Body (Aliases rep) -> TypeM rep [Names]

consume :: Checkable rep => Names -> TypeM rep ()

Proclaim that we have written to the given variables.

consumeOnlyParams :: [(VName, Names)] -> TypeM rep a -> TypeM rep a

Permit consumption of only the specified names. If one of these names is consumed, the consumption will be rewritten to be a consumption of the corresponding alias set. Consumption of anything else will result in a type error.

binding :: Checkable rep => Scope (Aliases rep) -> TypeM rep a -> TypeM rep a

alternative :: TypeM rep a -> TypeM rep b -> TypeM rep (a, b)