Futhark.Internalise.FreeVars

Description

Facilities for computing free variables in an expression, which we need for both lambda-lifting and defunctionalisation.

Documentation

freeVars :: Exp -> NameSet

Compute the set of free variables of an expression.

without :: NameSet -> Set VName -> NameSet

Set subtraction.

ident :: Ident -> NameSet

A NameSet with a single Nonunique name.

size :: VName -> NameSet

sizes :: Set VName -> NameSet

newtype NameSet

A set of names where we also track uniqueness.

Constructors

NameSet

Fields

unNameSet :: Map VName StructType

patVars :: Pat -> NameSet

Extract all the variable names bound in a pattern.