Futhark.Internalise.Monad

Documentation

data InternaliseM a

runInternaliseM :: MonadFreshNames m => Bool -> InternaliseM () -> m (Stms SOACS, [FunDef SOACS])

throwError :: MonadError e m => e -> m a

type VarSubsts = Map VName [SubExp]

A mapping from external variable names to the corresponding internalised subexpressions.

data InternaliseEnv

Constructors

InternaliseEnv

Fields

envSubsts :: VarSubsts

envDoBoundsChecks :: Bool

envSafe :: Bool

envAttrs :: Attrs

type FunInfo = ([VName], [DeclType], [FParam], [(SubExp, Type)] -> Maybe [DeclExtType])

substitutingVars :: VarSubsts -> InternaliseM a -> InternaliseM a

lookupSubst :: VName -> InternaliseM (Maybe [SubExp])

addFunDef :: FunDef SOACS -> InternaliseM ()

Add a function definition to the program being constructed.

lookupFunction :: VName -> InternaliseM FunInfo

lookupFunction' :: VName -> InternaliseM (Maybe FunInfo)

lookupConst :: VName -> InternaliseM (Maybe [SubExp])

bindFunction :: VName -> FunDef SOACS -> FunInfo -> InternaliseM ()

bindConstant :: VName -> FunDef SOACS -> InternaliseM ()

localConstsScope :: InternaliseM a -> InternaliseM a

assert :: String -> SubExp -> ErrorMsg SubExp -> SrcLoc -> InternaliseM Certs

Construct an Assert statement, but taking attributes into account. Always use this function, and never construct Assert directly in the internaliser!

Convenient reexports

module Futhark.Tools