Futhark.Internalise.TypesValues

Internalising types

internaliseReturnType :: StructRetType -> [TypeBase shape u] -> InternaliseM [TypeBase ExtShape Uniqueness]

internaliseLambdaReturnType :: TypeBase (DimDecl VName) () -> [TypeBase shape u] -> InternaliseM [TypeBase Shape NoUniqueness]

internaliseEntryReturnType :: StructRetType -> InternaliseM [[TypeBase ExtShape Uniqueness]]

As internaliseReturnType, but returns components of a top-level tuple type piecemeal.

internaliseType :: TypeBase (DimDecl VName) () -> InternaliseM [TypeBase ExtShape Uniqueness]

internaliseParamTypes :: [TypeBase (DimDecl VName) ()] -> InternaliseM [[TypeBase Shape Uniqueness]]

internaliseLoopParamType :: TypeBase (DimDecl VName) () -> [TypeBase shape u] -> InternaliseM [TypeBase Shape Uniqueness]

internalisePrimType :: PrimType -> PrimType

Convert an external primitive to an internal primitive.

internalisedTypeSize :: TypeBase (DimDecl VName) als -> InternaliseM Int

How many core language values are needed to represent one source language value of the given type?

internaliseSumType :: Map Name [StructType] -> InternaliseM ([TypeBase ExtShape Uniqueness], Map Name (Int, [Int]))

Internalising values

internalisePrimValue :: PrimValue -> PrimValue

Convert an external primitive value to an internal primitive value.