Futhark.Optimise.InPlaceLowering.SubstituteIndices

Description

This module exports facilities for transforming array accesses in a list of Stms (intended to be the bindings in a body). The idea is that you can state that some variable x is in fact an array indexing v[i0,i1,...].

Documentation

substituteIndices :: (MonadFreshNames m, BuilderOps rep, Buildable rep, Aliased rep, LParamInfo rep ~ Type) => IndexSubstitutions -> Stms rep -> m (IndexSubstitutions, Stms rep)

Perform the substitution.

type IndexSubstitution = (Certs, VName, Type, Slice SubExp)

type IndexSubstitutions = [(VName, IndexSubstitution)]