Futhark.Optimise.Simplify.Engine

Description

Perform general rule-based simplification based on data dependency information. This module will:

Perform common-subexpression elimination (CSE).

Hoist expressions out of loops (including lambdas) and branches. This is done as aggressively as possible.

Apply simplification rules (see Futhark.Optimise.Simplification.Rules).

If you just want to run the simplifier as simply as possible, you may prefer to use the Futhark.Optimise.Simplify module.

Monadic interface

data SimpleM rep a

runSimpleM :: SimpleM rep a -> SimpleOps rep -> Env rep -> VNameSource -> ((a, Bool), VNameSource)

data SimpleOps rep

Constructors

SimpleOps

Fields

mkExpDecS :: SymbolTable (Wise rep) -> Pat (Wise rep) -> Exp (Wise rep) -> SimpleM rep (ExpDec (Wise rep))

mkBodyS :: SymbolTable (Wise rep) -> Stms (Wise rep) -> Result -> SimpleM rep (Body (Wise rep))

protectHoistedOpS :: Protect (Builder (Wise rep))

Make a hoisted Op safe. The SubExp is a boolean that is true when the value of the statement will actually be used.

opUsageS :: Op (Wise rep) -> UsageTable

simplifyOpS :: SimplifyOp rep (Op (Wise rep))

type SimplifyOp rep op = op -> SimpleM rep (op, Stms (Wise rep))

bindableSimpleOps :: (SimplifiableRep rep, Buildable rep) => SimplifyOp rep (Op (Wise rep)) -> SimpleOps rep

data Env rep

emptyEnv :: RuleBook (Wise rep) -> HoistBlockers rep -> Env rep

data HoistBlockers rep

Constructors

HoistBlockers

Fields

blockHoistPar :: BlockPred (Wise rep)

Blocker for hoisting out of parallel loops.

blockHoistSeq :: BlockPred (Wise rep)

Blocker for hoisting out of sequential loops.

blockHoistBranch :: BlockPred (Wise rep)

Blocker for hoisting out of branches.

isAllocation :: Stm (Wise rep) -> Bool

neverBlocks :: BlockPred rep

noExtraHoistBlockers :: HoistBlockers rep

neverHoist :: HoistBlockers rep

type BlockPred rep = SymbolTable rep -> UsageTable -> Stm rep -> Bool

orIf :: BlockPred rep -> BlockPred rep -> BlockPred rep

hasFree :: ASTRep rep => Names -> BlockPred rep

isConsumed :: BlockPred rep

isFalse :: Bool -> BlockPred rep

isOp :: BlockPred rep

isNotSafe :: ASTRep rep => BlockPred rep

asksEngineEnv :: (Env rep -> a) -> SimpleM rep a

askVtable :: SimpleM rep (SymbolTable (Wise rep))

localVtable :: (SymbolTable (Wise rep) -> SymbolTable (Wise rep)) -> SimpleM rep a -> SimpleM rep a

Building blocks

type SimplifiableRep rep = (ASTRep rep, Simplifiable (LetDec rep), Simplifiable (FParamInfo rep), Simplifiable (LParamInfo rep), Simplifiable (RetType rep), Simplifiable (BranchType rep), TraverseOpStms (Wise rep), CanBeWise (Op rep), IndexOp (OpWithWisdom (Op rep)), BuilderOps (Wise rep), IsOp (Op rep))

class Simplifiable e where

Methods

simplify :: SimplifiableRep rep => e -> SimpleM rep e

simplifyFun :: SimplifiableRep rep => FunDef (Wise rep) -> SimpleM rep (FunDef (Wise rep))

simplifyStms :: SimplifiableRep rep => Stms (Wise rep) -> SimpleM rep (Stms (Wise rep))

simplifyStmsWithUsage :: SimplifiableRep rep => UsageTable -> Stms (Wise rep) -> SimpleM rep (Stms (Wise rep))

simplifyLambda :: SimplifiableRep rep => Lambda (Wise rep) -> SimpleM rep (Lambda (Wise rep), Stms (Wise rep))

simplifyLambdaNoHoisting :: SimplifiableRep rep => Lambda (Wise rep) -> SimpleM rep (Lambda (Wise rep))

bindLParams :: SimplifiableRep rep => [LParam (Wise rep)] -> SimpleM rep a -> SimpleM rep a

simplifyBody :: SimplifiableRep rep => BlockPred (Wise rep) -> UsageTable -> [Usages] -> Body (Wise rep) -> SimpleM rep (Stms (Wise rep), Body (Wise rep))

Simplify a single body.

data SymbolTable rep

hoistStms :: SimplifiableRep rep => RuleBook (Wise rep) -> BlockPred (Wise rep) -> Stms (Wise rep) -> SimpleM rep (a, UsageTable) -> SimpleM rep (a, Stms (Wise rep), Stms (Wise rep))

blockIf :: SimplifiableRep rep => BlockPred (Wise rep) -> Stms (Wise rep) -> SimpleM rep (a, UsageTable) -> SimpleM rep (a, Stms (Wise rep), Stms (Wise rep))

enterLoop :: SimpleM rep a -> SimpleM rep a

Indicate in the symbol table that we have descended into a loop.

module Futhark.Optimise.Simplify.Rep