Futhark.Optimise.Simplify.Rule

Description

This module defines the concept of a simplification rule for bindings. The intent is that you pass some context (such as symbol table) and a binding, and is given back a sequence of bindings that compute the same result, but are "better" in some sense.

These rewrite rules are "local", in that they do not maintain any state or look at the program as a whole. Compare this to the fusion algorithm in Futhark.Optimise.Fusion.Fusion, which must be implemented as its own pass.

The rule monad

data RuleM rep a

The monad in which simplification rules are evaluated.

cannotSimplify :: RuleM rep a

liftMaybe :: Maybe a -> RuleM rep a

Rule definition

data Rule rep

An efficient way of encoding whether a simplification rule should even be attempted.

Constructors

Simplify (RuleM rep ())
Give it a shot.

Skip
Don't bother.

data SimplificationRule rep a

A simplification rule takes some argument and a statement, and tries to simplify the statement.

Constructors

RuleGeneric (RuleGeneric rep a)

RuleBasicOp (RuleBasicOp rep a)

RuleIf (RuleIf rep a)

RuleDoLoop (RuleDoLoop rep a)

RuleOp (RuleOp rep a)

type RuleGeneric rep a = a -> Stm rep -> Rule rep

type RuleBasicOp rep a = a -> Pat rep -> StmAux (ExpDec rep) -> BasicOp -> Rule rep

type RuleIf rep a = a -> Pat rep -> StmAux (ExpDec rep) -> (SubExp, BodyT rep, BodyT rep, IfDec (BranchType rep)) -> Rule rep

type RuleDoLoop rep a = a -> Pat rep -> StmAux (ExpDec rep) -> ([(FParam rep, SubExp)], LoopForm rep, BodyT rep) -> Rule rep

Top-down rules

type TopDown rep = SymbolTable rep

Context for a rule applied during top-down traversal of the program. Takes a symbol table as argument.

type TopDownRule rep = SimplificationRule rep (TopDown rep)

type TopDownRuleGeneric rep = RuleGeneric rep (TopDown rep)

type TopDownRuleBasicOp rep = RuleBasicOp rep (TopDown rep)

type TopDownRuleIf rep = RuleIf rep (TopDown rep)

type TopDownRuleDoLoop rep = RuleDoLoop rep (TopDown rep)

type TopDownRuleOp rep = RuleOp rep (TopDown rep)

Bottom-up rules

type BottomUp rep = (SymbolTable rep, UsageTable)

Context for a rule applied during bottom-up traversal of the program. Takes a symbol table and usage table as arguments.

type BottomUpRule rep = SimplificationRule rep (BottomUp rep)

type BottomUpRuleGeneric rep = RuleGeneric rep (BottomUp rep)

type BottomUpRuleBasicOp rep = RuleBasicOp rep (BottomUp rep)

type BottomUpRuleIf rep = RuleIf rep (BottomUp rep)

type BottomUpRuleDoLoop rep = RuleDoLoop rep (BottomUp rep)

type BottomUpRuleOp rep = RuleOp rep (BottomUp rep)

Assembling rules

data RuleBook rep

A collection of both top-down and bottom-up rules.

ruleBook :: [TopDownRule m] -> [BottomUpRule m] -> RuleBook m

Construct a rule book from a collection of rules.

Applying rules

topDownSimplifyStm :: (MonadFreshNames m, HasScope rep m) => RuleBook rep -> SymbolTable rep -> Stm rep -> m (Maybe (Stms rep))

simplifyStm lookup stm performs simplification of the binding stm. If simplification is possible, a replacement list of bindings is returned, that bind at least the same names as the original binding (and possibly more, for intermediate results).

bottomUpSimplifyStm :: (MonadFreshNames m, HasScope rep m) => RuleBook rep -> (SymbolTable rep, UsageTable) -> Stm rep -> m (Maybe (Stms rep))

simplifyStm uses stm performs simplification of the binding stm. If simplification is possible, a replacement list of bindings is returned, that bind at least the same names as the original binding (and possibly more, for intermediate results). The first argument is the set of names used after this binding.