Futhark.Optimise.Simplify.Rules.Simple

Description

Particularly simple simplification rules.

Documentation

type TypeLookup = SubExp -> Maybe Type

A function that, given a subexpression, returns its type.

type VarLookup rep = VName -> Maybe (Exp rep, Certs)

A function that, given a variable name, returns its definition.

applySimpleRules :: VarLookup rep -> TypeLookup -> BasicOp -> Maybe (BasicOp, Certs)

Try to simplify the given BasicOp, returning a new BasicOp and certificates that it must depend on.