Futhark.Optimise.TileLoops.Shared

Documentation

type TileM = ReaderT (Scope GPU) (State VNameSource)

type Env = (WithEnv, IxFnEnv)

scratch :: MonadBuilder m => String -> PrimType -> [SubExp] -> m VName

index :: MonadBuilder m => String -> VName -> [VName] -> m VName

update :: MonadBuilder m => String -> VName -> [VName] -> VName -> m VName

update' :: MonadBuilder m => String -> VName -> [VName] -> SubExp -> m VName

rebindLambda :: Lambda GPU -> [VName] -> [VName] -> Stms GPU

forLoop' :: SubExp -> [VName] -> (VName -> [VName] -> Builder GPU (Body GPU)) -> Builder GPU [VName]

forLoop :: SubExp -> [VName] -> (VName -> [VName] -> Builder GPU (Body GPU)) -> Builder GPU VName

segMap1D :: String -> SegLevel -> ResultManifest -> (VName -> Builder GPU Result) -> Builder GPU [VName]

segMap2D :: String -> SegLevel -> ResultManifest -> (SubExp, SubExp) -> ((VName, VName) -> Builder GPU Result) -> Builder GPU [VName]

segMap3D :: String -> SegLevel -> ResultManifest -> (SubExp, SubExp, SubExp) -> ((VName, VName, VName) -> Builder GPU Result) -> Builder GPU [VName]

segScatter2D :: String -> SubExp -> VName -> SegLevel -> (SubExp, SubExp) -> ((VName, VName) -> Builder GPU (SubExp, SubExp)) -> Builder GPU [VName]

type VarianceTable = Map VName Names

The variance table keeps a mapping from a variable name (something produced by a Stm) to the kernel thread indices that name depends on. If a variable is not present in this table, that means it is bound outside the kernel (and so can be considered invariant to all dimensions).

varianceInStms :: VarianceTable -> Stms GPU -> VarianceTable

isTileableRedomap :: Stm GPU -> Maybe (SubExp, [VName], (Commutativity, Lambda GPU, [SubExp], Lambda GPU))

changeEnv :: Env -> VName -> Exp GPU -> TileM Env

kkLoopBody :: Env -> ((SubExp, SubExp, SubExp, SubExp, SubExp, SubExp, SubExp, SubExp), SegLevel, [Int], (VName, SubExp, VName, SubExp, SubExp), (SubExp, SubExp), (VName, VName), (Stm GPU, VName, PrimType, Stm GPU, VName, PrimType), (Lambda GPU, Lambda GPU)) -> VName -> (VName, VName, VName) -> Bool -> Builder GPU [VName]