Futhark.Pass

Description

Definition of a polymorphic (generic) pass that can work with programs of any rep.

Documentation

data PassM a

The monad in which passes execute.

runPassM :: MonadFreshNames m => PassM a -> m (a, Log)

Execute a PassM action, yielding logging information and either an error text or a result.

liftEither :: Show err => Either err a -> PassM a

Turn an Either computation into a PassM. If the Either is Left, the result is a CompilerBug.

liftEitherM :: Show err => PassM (Either err a) -> PassM a

Turn an Either monadic computation into a PassM. If the Either is Left, the result is an exception.

data Pass fromrep torep

A compiler pass transforming a Prog of a given rep to a Prog of another rep.

Constructors

Pass

Fields

passName :: String

Name of the pass. Keep this short and simple. It will be used to automatically generate a command-line option name via passLongOption.

passDescription :: String

A slightly longer description, which will show up in the command-line help text.

passFunction :: Prog fromrep -> PassM (Prog torep)

passLongOption :: Pass fromrep torep -> String

Take the name of the pass, turn spaces into dashes, and make all characters lowercase.

parPass :: (a -> PassM b) -> [a] -> PassM [b]

Apply a PassM operation in parallel to multiple elements, joining together the name sources and logs, and propagating any error properly.

intraproceduralTransformation :: (Scope rep -> Stms rep -> PassM (Stms rep)) -> Prog rep -> PassM (Prog rep)

Like intraproceduralTransformationWithConsts, but do not change the top-level constants, and simply pass along their Scope.

intraproceduralTransformationWithConsts :: (Stms fromrep -> PassM (Stms torep)) -> (Stms torep -> FunDef fromrep -> PassM (FunDef torep)) -> Prog fromrep -> PassM (Prog torep)

Apply some operation to the top-level constants. Then applies an operation to all the function function definitions, which are also given the transformed constants so they can be brought into scope. The function definition transformations are run in parallel (with parPass), since they cannot affect each other.