Futhark.Pass.ExplicitAllocations

Description

A generic transformation for adding memory allocations to a Futhark program. Specialised by specific representations in submodules.

Documentation

explicitAllocationsGeneric :: Allocable fromrep torep inner => (Op fromrep -> AllocM fromrep torep (Op torep)) -> (Exp torep -> AllocM fromrep torep [ExpHint]) -> Pass fromrep torep

explicitAllocationsInStmsGeneric :: (MonadFreshNames m, HasScope torep m, Allocable fromrep torep inner) => (Op fromrep -> AllocM fromrep torep (Op torep)) -> (Exp torep -> AllocM fromrep torep [ExpHint]) -> Stms fromrep -> m (Stms torep)

data ExpHint

Constructors

NoHint

Hint IxFun Space

defaultExpHints :: (Monad m, ASTRep rep) => Exp rep -> m [ExpHint]

type Allocable fromrep torep inner = (PrettyRep fromrep, PrettyRep torep, Mem torep inner, LetDec torep ~ LetDecMem, FParamInfo fromrep ~ DeclType, LParamInfo fromrep ~ Type, BranchType fromrep ~ ExtType, RetType fromrep ~ DeclExtType, BodyDec fromrep ~ (), BodyDec torep ~ (), ExpDec torep ~ (), SizeSubst inner, BuilderOps torep)

data AllocM fromrep torep a

Monad for adding allocations to an entire program.

data AllocEnv fromrep torep

Constructors

AllocEnv

Fields

chunkMap :: ChunkMap

aggressiveReuse :: Bool

Aggressively try to reuse memory in do-loops - should be True inside kernels, False outside.

allocSpace :: Space

When allocating memory, put it in this memory space. This is primarily used to ensure that group-wide statements store their results in local memory.

envConsts :: Set VName

The set of names that are known to be constants at kernel compile time.

allocInOp :: Op fromrep -> AllocM fromrep torep (Op torep)

envExpHints :: Exp torep -> AllocM fromrep torep [ExpHint]

class SizeSubst op where

Minimal complete definition

opSizeSubst

Methods

opSizeSubst :: PatT dec -> op -> ChunkMap

opIsConst :: op -> Bool

allocInStms :: Allocable fromrep torep inner => Stms fromrep -> AllocM fromrep torep a -> AllocM fromrep torep a

allocForArray :: Allocable fromrep torep inner => Type -> Space -> AllocM fromrep torep VName

Allocate memory for a value of the given type.

simplifiable :: (SimplifiableRep rep, ExpDec rep ~ (), BodyDec rep ~ (), Mem rep inner) => (OpWithWisdom inner -> UsageTable) -> (OpWithWisdom inner -> SimpleM rep (OpWithWisdom inner, Stms (Wise rep))) -> SimpleOps rep

arraySizeInBytesExp :: Type -> PrimExp VName

mkLetNamesB' :: (LetDec (Rep m) ~ LetDecMem, Mem (Rep m) inner, MonadBuilder m, ExpDec (Rep m) ~ ()) => ExpDec (Rep m) -> [VName] -> Exp (Rep m) -> m (Stm (Rep m))

mkLetNamesB'' :: (BuilderOps rep, Mem rep inner, LetDec rep ~ LetDecMem, OpReturns (OpWithWisdom inner), ExpDec rep ~ (), Rep m ~ Wise rep, HasScope (Wise rep) m, MonadBuilder m, CanBeWise inner) => [VName] -> Exp (Wise rep) -> m (Stm (Wise rep))

dimAllocationSize :: ChunkMap -> SubExp -> SubExp

The subexpression giving the number of elements we should allocate space for. See ChunkMap comment.

type ChunkMap = Map VName SubExp

A mapping from chunk names to their maximum size. XXX FIXME HACK: This is part of a hack to add loop-invariant allocations to reduce kernels, because memory expansion does not use range analysis yet (it should).

Module re-exports

module Futhark.MonadFreshNames

module Futhark.Pass

module Futhark.Tools