Futhark.Pass.ExtractKernels.Distribution

Documentation

type Target = (PatT Type, Result)

data Targets

First pair element is the very innermost ("current") target. In the list, the outermost target comes first. Invariant: Every element of a pattern must be present as the result of the immediately enclosing target. This is ensured by pushInnerTarget by removing unused pattern elements.

ppTargets :: Targets -> String

singleTarget :: Target -> Targets

outerTarget :: Targets -> Target

innerTarget :: Targets -> Target

pushInnerTarget :: Target -> Targets -> Targets

popInnerTarget :: Targets -> Maybe (Target, Targets)

targetsScope :: DistRep rep => Targets -> Scope rep

data LoopNesting

Constructors

MapNesting

Fields

loopNestingPat :: PatT Type

loopNestingAux :: StmAux ()

loopNestingWidth :: SubExp

loopNestingParamsAndArrs :: [(Param Type, VName)]

ppLoopNesting :: LoopNesting -> String

scopeOfLoopNesting :: LParamInfo rep ~ Type => LoopNesting -> Scope rep

data Nesting

Constructors

Nesting

Fields

nestingLetBound :: Names

nestingLoop :: LoopNesting

type Nestings = (Nesting, [Nesting])

ppNestings :: Nestings -> String

letBindInInnerNesting :: Names -> Nestings -> Nestings

singleNesting :: Nesting -> Nestings

pushInnerNesting :: Nesting -> Nestings -> Nestings

type KernelNest = (LoopNesting, [LoopNesting])

Note: first element is *outermost* nesting. This is different from the similar types elsewhere!

ppKernelNest :: KernelNest -> String

newKernel :: LoopNesting -> KernelNest

innermostKernelNesting :: KernelNest -> LoopNesting

Retrieve the innermost kernel nesting.

pushKernelNesting :: Target -> LoopNesting -> KernelNest -> KernelNest

Add new outermost nesting, pushing the current outermost to the list, also taking care to swap patterns if necessary.

pushInnerKernelNesting :: Target -> LoopNesting -> KernelNest -> KernelNest

Add new innermost nesting, pushing the current outermost to the list. It is important that the Target has the right order (non-permuted compared to what is expected by the outer nests).

scopeOfKernelNest :: LParamInfo rep ~ Type => KernelNest -> Scope rep

kernelNestLoops :: KernelNest -> [LoopNesting]

kernelNestWidths :: KernelNest -> [SubExp]

boundInKernelNest :: KernelNest -> Names

boundInKernelNests :: KernelNest -> [Names]

flatKernel :: MonadFreshNames m => KernelNest -> m ([(VName, SubExp)], [KernelInput])

Flatten a kernel nesting to:

The index space.

The kernel inputs - note that some of these may be unused.

constructKernel :: (DistRep rep, MonadFreshNames m, LocalScope rep m) => MkSegLevel rep m -> KernelNest -> Body rep -> m (Stm rep, Stms rep)

tryDistribute :: (DistRep rep, MonadFreshNames m, LocalScope rep m, MonadLogger m) => MkSegLevel rep m -> Nestings -> Targets -> Stms rep -> m (Maybe (Targets, Stms rep))

tryDistributeStm :: (MonadFreshNames m, HasScope t m, ASTRep rep) => Nestings -> Targets -> Stm rep -> m (Maybe (Result, Targets, KernelNest))