Futhark.Pass.ExtractKernels.ISRWIM

Description

Interchanging scans with inner maps.

Documentation

iswim :: (MonadBuilder m, Rep m ~ SOACS) => Pat -> SubExp -> Lambda -> [(SubExp, VName)] -> Maybe (m ())

Interchange Scan With Inner Map. Tries to turn a scan(map) into a @map(scan)

irwim :: (MonadBuilder m, Rep m ~ SOACS) => Pat -> SubExp -> Commutativity -> Lambda -> [(SubExp, VName)] -> Maybe (m ())

Interchange Reduce With Inner Map. Tries to turn a reduce(map) into a @map(reduce)

rwimPossible :: Lambda -> Maybe (Pat, Certs, SubExp, Lambda)

Does this reduce operator contain an inner map, and if so, what does that map look like?