Futhark.Pipeline

Description

Definition of the core compiler driver building blocks. The spine of the compiler is the FutharkM monad, although note that individual passes are pure functions, and do not use the FutharkM monad (see Futhark.Pass).

Running the compiler involves producing an initial IR program (see Futhark.Compiler.Program), running a Pipeline to produce a final program (still in IR), then running an Action, which is usually a code generator.

Documentation

data Pipeline fromrep torep

A compiler pipeline is conceptually a function from programs to programs, where the actual representation may change. Pipelines can be composed using their Category instance.

data PipelineConfig

Configuration object for running a compiler pipeline.

Constructors

PipelineConfig

Fields

pipelineVerbose :: Bool

pipelineValidate :: Bool

data Action rep

A compilation always ends with some kind of action.

Constructors

Action

Fields

actionName :: String

actionDescription :: String

actionProcedure :: Prog rep -> FutharkM ()

data FutharkM a

The main Futhark compiler driver monad - basically some state tracking on top if IO.

runFutharkM :: FutharkM a -> Verbosity -> IO (Either CompilerError a)

Run a FutharkM action.

data Verbosity

How much information to print to stderr while the compiler is running.

Constructors

NotVerbose
Silence is golden.

Verbose
Print messages about which pass is running.

VeryVerbose
Also print logs from individual passes.

module Futhark.Error

onePass :: Checkable torep => Pass fromrep torep -> Pipeline fromrep torep

Construct a pipeline from a single compiler pass.

passes :: Checkable rep => [Pass rep rep] -> Pipeline rep rep

Create a pipeline from a list of passes.

runPipeline :: Pipeline fromrep torep -> PipelineConfig -> Prog fromrep -> FutharkM (Prog torep)

Run the pipeline on the given program.