Futhark.Pkg.Info

Description

Obtaining information about packages over THE INTERNET!

Package info

data PkgInfo m

Information about a package. The name of the package is stored separately.

Constructors

PkgInfo

Fields

pkgVersions :: Map SemVer (PkgRevInfo m)

pkgLookupCommit :: Maybe Text -> m (PkgRevInfo m)

Look up information about a specific commit, or HEAD in case of Nothing.

lookupPkgRev :: SemVer -> PkgInfo m -> Maybe (PkgRevInfo m)

Lookup information about a given version of a package.

pkgInfo :: (MonadIO m, MonadLogger m, MonadFail m) => PkgPath -> m (Either Text (PkgInfo m))

Retrieve information about a package based on its package path. This uses Semantic Import Versioning when interacting with repositories. For example, a package github.comuserrepo will match version 0.* or 1.* tags only, a package github.comuserrepo/v2 will match 2.* tags, and so forth..

data PkgRevInfo m

Information about a version of a single package. The version number is stored separately.

Constructors

PkgRevInfo

Fields

pkgRevZipballUrl :: Text

pkgRevZipballDir :: FilePath

The directory inside the zipball containing the lib directory, in which the package files themselves are stored (Based on the package path).

pkgRevCommit :: Text

The commit ID can be used for verification ("freezing"), by storing what it was at the time this version was last selected.

pkgRevGetManifest :: GetManifest m

pkgRevTime :: UTCTime

Timestamp for when the revision was made (rarely used).

data GetManifest m

The manifest is stored as a monadic action, because we want to fetch them on-demand. It would be a waste to fetch it information for every version of every package if we only actually need a small subset of them.

downloadZipball :: (MonadLogger m, MonadIO m, MonadFail m) => PkgRevInfo m -> m Archive

Download the zip archive corresponding to a specific package version.

Package registry

data PkgRegistry m

A package registry is a mapping from package paths to information about the package. It is unlikely that any given registry is global; rather small registries are constructed on-demand based on the package paths referenced by the user, and may also be combined monoidically. In essence, the PkgRegistry is just a cache.

class (MonadIO m, MonadLogger m, MonadFail m) => MonadPkgRegistry m where

Monads that support a stateful package registry. These are also required to be instances of MonadIO because most package registry operations involve network operations.

Minimal complete definition

getPkgRegistry, putPkgRegistry

Methods

getPkgRegistry :: m (PkgRegistry m)

putPkgRegistry :: PkgRegistry m -> m ()

modifyPkgRegistry :: (PkgRegistry m -> PkgRegistry m) -> m ()

lookupPackage :: MonadPkgRegistry m => PkgPath -> m (PkgInfo m)

Given a package path, look up information about that package.

lookupPackageRev :: MonadPkgRegistry m => PkgPath -> SemVer -> m (PkgRevInfo m)

Look up information about a specific version of a package.

lookupNewestRev :: MonadPkgRegistry m => PkgPath -> m SemVer

Find the newest version of a package.