Futhark.Pkg.Solve

Description

Dependency solver

This is a relatively simple problem due to the choice of the Minimum Package Version algorithm. In fact, the only failure mode is referencing an unknown package or revision.

Documentation

solveDeps :: MonadPkgRegistry m => PkgRevDeps -> m BuildList

Run the solver, producing both a package registry containing a cache of the lookups performed, as well as a build list.

solveDepsPure :: PkgRevDepInfo -> PkgRevDeps -> Either Text BuildList

Perform package resolution with only pre-known information. This is useful for testing.

type PkgRevDepInfo = Map (PkgPath, SemVer) PkgRevDeps

A mapping of package revisions to the dependencies of that package. Can be considered a PkgRegistry without the option of obtaining more information from the Internet. Probably useful only for testing the solver.