Futhark.Pkg.Types

Description

Types (and a few other simple definitions) for futhark-pkg.

Documentation

type PkgPath = Text

A package path is a unique identifier for a package, for example github.comuserfoo.

pkgPathFilePath :: PkgPath -> FilePath

Turn a package path (which always uses forward slashes) into a file path in the local file system (which might use different slashes).

newtype PkgRevDeps

The dependencies of a (revision of a) package is a mapping from package paths to minimum versions (and an optional hash pinning).

Constructors

PkgRevDeps (Map PkgPath (SemVer, Maybe Text))

prettySemVer :: SemVer -> Text

Convert a SemVer back to its textual representation.

data SemVer

An (Ideal) version number that conforms to Semantic Versioning. This is a prescriptive parser, meaning it follows the SemVer standard.

Legal semvers are of the form: MAJOR.MINOR.PATCH-PREREL+META

Example: 1.2.3-r1+commithash

Extra Rules:

Pre-release versions have lower precedence than normal versions.

Build metadata does not affect version precedence.

PREREL and META strings may only contain ASCII alphanumerics and hyphens.

For more information, see http://semver.org

Constructors

SemVer

Fields

_svMajor :: !Word

_svMinor :: !Word

_svPatch :: !Word

_svPreRel :: ![VChunk]

_svMeta :: !(Maybe Text)

data VUnit

A single unit of a Version. May be digits or a string of characters. Groups of these are called VChunks, and are the identifiers separated by periods in the source.

Constructors

Digits Word

Str Text

Versions

commitVersion :: Text -> Text -> SemVer

commitVersion timestamp commit constructs a commit version.

isCommitVersion :: SemVer -> Maybe Text

Versions of the form (0,0,0)-timestamp+hash are treated specially, as a reference to the commit identified uniquely with hash (typically the Git commit ID). This function detects such versions.

parseVersion :: Text -> Either (ParseErrorBundle Text Void) SemVer

Unfortunately, Data.Versions has a buggy semver parser that collapses consecutive zeroes in the metadata field. So, we define our own parser here. It's a little simpler too, since we don't need full semver.

Package manifests

data PkgManifest

A structure corresponding to a futhark.pkg file, including comments. It is an invariant that duplicate required packages do not occcur (the parser will verify this).

Constructors

PkgManifest

Fields

manifestPkgPath :: Commented (Maybe PkgPath)

The name of the package.

manifestRequire :: Commented [Either Comment Required]

manifestEndComments :: [Comment]

newPkgManifest :: Maybe PkgPath -> PkgManifest

Possibly given a package path, construct an otherwise-empty manifest file.

pkgRevDeps :: PkgManifest -> PkgRevDeps

The required packages listed in a package manifest.

pkgDir :: PkgManifest -> Maybe FilePath

Where in the corresponding repository archive we can expect to find the package files.

addRequiredToManifest :: Required -> PkgManifest -> (PkgManifest, Maybe Required)

Add new required package to the package manifest. If the package was already present, return the old version.

removeRequiredFromManifest :: PkgPath -> PkgManifest -> Maybe (PkgManifest, Required)

Remove a required package from the manifest. Returns Nothing if the package was not found in the manifest, and otherwise the new manifest and the Required that was present.

prettyPkgManifest :: PkgManifest -> Text

Prettyprint a package manifest such that it can be written to a futhark.pkg file.

type Comment = Text

A line comment.

data Commented a

Wraps a value with an annotation of preceding line comments. This is important to our goal of being able to programmatically modify the futhark.pkg file while keeping comments intact.

Constructors

Commented

Fields

comments :: [Comment]

commented :: a

data Required

An entry in the required section of a futhark.pkg file.

Constructors

Required

Fields

requiredPkg :: PkgPath

Name of the required package.

requiredPkgRev :: SemVer

The minimum revision.

requiredHash :: Maybe Text

An optional hash indicating what this revision looked like the last time we saw it. Used for integrity checking.

futharkPkg :: FilePath

The name of the file containing the futhark-pkg manifest.

Parsing package manifests

parsePkgManifest :: FilePath -> Text -> Either (ParseErrorBundle Text Void) PkgManifest

Parse a text as a PkgManifest. The FilePath is used for any error messages.

parsePkgManifestFromFile :: FilePath -> IO PkgManifest

Read contents of file and pass it to parsePkgManifest.

errorBundlePretty

Arguments

:: (VisualStream s, TraversableStream s, ShowErrorComponent e)

=> ParseErrorBundle s e
Parse error bundle to display

-> String
Textual rendition of the bundle

Pretty-print a ParseErrorBundle. All ParseErrors in the bundle will be pretty-printed in order together with the corresponding offending lines by doing a single efficient pass over the input stream. The rendered String always ends with a newline.

Since: megaparsec-7.0.0

Build list

newtype BuildList

A mapping from package paths to their chosen revisions. This is the result of the version solver.

Constructors

BuildList

Fields

unBuildList :: Map PkgPath SemVer

prettyBuildList :: BuildList -> Text

Prettyprint a build list; one package per line and newline-terminated.