Futhark.Script

Description

FutharkScript is a (tiny) subset of Futhark used to write small expressions that are evaluated by server executables. The futhark literate command is the main user.

Server

data ScriptServer

Like a Server, but keeps a bit more state to make FutharkScript more convenient.

withScriptServer :: ServerCfg -> (ScriptServer -> IO a) -> IO a

Start a server, execute an action, then shut down the server. Similar to withServer.

withScriptServer' :: MonadIO m => Server -> (ScriptServer -> m a) -> m a

Run an action with a ScriptServer produced by an existing Server, without shutting it down at the end.

Expressions, values, and types

data Func

A function called in a Call expression can be either a Futhark function or a builtin function.

Constructors

FuncFut EntryName

FuncBuiltin Text

data Exp

A FutharkScript expression. This is a simple AST that might not correspond exactly to what the user wrote (e.g. no parentheses or source locations). This is fine for small expressions, which is all this is meant for.

Constructors

Call Func [Exp]

Const Value

Tuple [Exp]

Record [(Text, Exp)]

StringLit Text

Let [VarName] Exp Exp

ServerVar TypeName VarName
Server-side variable, *not* Futhark variable (these are handled in Call).

parseExp :: Parser () -> Parser Exp

Parse a FutharkScript expression, given a whitespace parser.

parseExpFromText :: FilePath -> Text -> Either Text Exp

Parse a FutharkScript expression with normal whitespace handling.

varsInExp :: Exp -> Set EntryName

The set of Futhark variables that are referenced by the expression - these will have to be entry points in the Futhark program.

data ScriptValueType

The type of a ScriptValue - either a value type or a function type.

Constructors

STValue TypeName

STFun [TypeName] [TypeName]
Ins, then outs.

data ScriptValue v

A ScriptValue is either a base value or a partially applied function. We don't have real first-class functions in FutharkScript, but we sort of have closures.

Constructors

SValue TypeName v

SFun EntryName [TypeName] [TypeName] [ScriptValue v]
Ins, then outs. Yes, this is the opposite of more or less everywhere else.

scriptValueType :: ScriptValue v -> ScriptValueType

The type of a ScriptValue.

serverVarsInValue :: ExpValue -> Set VarName

The set of server-side variables in the value.

data ValOrVar

A Haskell-level value or a variable on the server.

Constructors

VVal Value

VVar VarName

type ExpValue = Compound (ScriptValue ValOrVar)

The intermediate values produced by an expression - in particular, these may not be on the server.

Evaluation

type EvalBuiltin m = Text -> [CompoundValue] -> m CompoundValue

How to evaluate a builtin function.

evalExp :: forall m. (MonadError Text m, MonadIO m) => EvalBuiltin m -> ScriptServer -> Exp -> m ExpValue

Evaluate a FutharkScript expression relative to some running server.

getExpValue :: (MonadError Text m, MonadIO m) => ScriptServer -> ExpValue -> m CompoundValue

Read actual values from the server. Fails for values that have no well-defined external representation.

evalExpToGround :: (MonadError Text m, MonadIO m) => EvalBuiltin m -> ScriptServer -> Exp -> m (Either (Compound ScriptValueType) CompoundValue)

Like evalExp, but requires all values to be non-functional. If the value has a bad type, return that type instead. Other evaluation problems (e.g. type failures) raise errors.

valueToExp :: ExpValue -> Exp

Convert a value into a corresponding expression.

freeValue :: (MonadError Text m, MonadIO m) => ScriptServer -> ExpValue -> m ()

Release all the server-side variables in the value. Yes, FutharkScript has manual memory management...