Futhark.Test.Spec

Description

Definition and parsing of a test specification.

Documentation

testSpecFromProgram :: FilePath -> IO (Either String ProgramTest)

Read the test specification from the given Futhark program.

testSpecFromProgramOrDie :: FilePath -> IO ProgramTest

Like testSpecFromProgram, but exits the process on error.

testSpecsFromPaths :: [FilePath] -> IO (Either String [(FilePath, ProgramTest)])

Read test specifications from the given paths, which can be a files or directories containing .fut files and further directories.

testSpecsFromPathsOrDie :: [FilePath] -> IO [(FilePath, ProgramTest)]

Like testSpecsFromPaths, but kills the process on errors.

testSpecFromFile :: FilePath -> IO (Either String ProgramTest)

Read a test specification from a file. Expects only a single block, and no comment prefixes.

testSpecFromFileOrDie :: FilePath -> IO ProgramTest

Like testSpecFromFile, but kills the process on errors.

data ProgramTest

Description of a test to be carried out on a Futhark program. The Futhark program is stored separately.

Constructors

ProgramTest

Fields

testDescription :: Text

testTags :: [Text]

testAction :: TestAction

data StructureTest

A structure test specifies a compilation pipeline, as well as metrics for the program coming out the other end.

Constructors

StructureTest StructurePipeline AstMetrics

data StructurePipeline

How a program can be transformed.

Constructors

GpuPipeline

SOACSPipeline

SeqMemPipeline

GpuMemPipeline

NoPipeline

data WarningTest

A warning test requires that a warning matching the regular expression is produced. The program must also compile succesfully.

Constructors

ExpectedWarning Text Regex

data TestAction

How to test a program.

Constructors

CompileTimeFailure ExpectedError

RunCases [InputOutputs] [StructureTest] [WarningTest]

data ExpectedError

The error expected for a negative test.

Constructors

AnyError

ThisError Text Regex

data InputOutputs

Input and output pairs for some entry point(s).

Constructors

InputOutputs

Fields

iosEntryPoint :: Text

iosTestRuns :: [TestRun]

data TestRun

A condition for execution, input, and expected result.

Constructors

TestRun

Fields

runTags :: [String]

runInput :: Values

runExpectedResult :: ExpectedResult Success

runIndex :: Int

runDescription :: String

data ExpectedResult values

How a test case is expected to terminate.

Constructors

Succeeds (Maybe values)
Execution suceeds, with or without expected result values.

RunTimeFailure ExpectedError
Execution fails with this error.

data Success

The result expected from a succesful execution.

Constructors

SuccessValues Values
These values are expected.

SuccessGenerateValues
Compute expected values from executing a known-good reference implementation.

data Values

Several values - either literally, or by reference to a file, or to be generated on demand. All paths are relative to test program.

Constructors

Values [Value]

InFile FilePath

GenValues [GenValue]

ScriptValues Exp

ScriptFile FilePath

data GenValue

Constructors

GenValue ValueType
Generate a value of the given rank and primitive type. Scalars are considered 0-ary arrays.

GenPrim Value
A fixed non-randomised primitive value.

genValueType :: GenValue -> String

A prettyprinted representation of type of value produced by a GenValue.