Futhark.Test.Values

Description

This module provides an efficient value representation as well as parsing and comparison functions.

Documentation

module Futhark.Data

module Futhark.Data.Compare

module Futhark.Data.Reader

data Compound v

The structure of a compound value, parameterised over the actual values. For most cases you probably want CompoundValue.

Constructors

ValueRecord (Map Text (Compound v))

ValueTuple [Compound v]
Must not be single value.

ValueAtom v

type CompoundValue = Compound Value

Like a Value, but also grouped in compound ways that are not supported by raw values. You cannot parse or read these in standard ways, and they cannot be elements of arrays.

mkCompound :: [Compound v] -> Compound v

Create a tuple for a non-unit list, and otherwise a ValueAtom

unCompound :: Compound v -> [Compound v]

If the value is a tuple, extract the components, otherwise return a singleton list of the value.

Orphan instances

Pretty Value

Instance details

Methods

ppr :: Value -> Doc

pprPrec :: Int -> Value -> Doc

pprList :: [Value] -> Doc

Pretty ValueType

Instance details

Methods

ppr :: ValueType -> Doc

pprPrec :: Int -> ValueType -> Doc

pprList :: [ValueType] -> Doc