Futhark.Transform.Substitute

Description

This module contains facilities for replacing variable names in syntactic constructs.

Documentation

type Substitutions = Map VName VName

The substitutions to be made are given by a mapping from names to names.

class Substitute a where

A type that is an instance of this class supports substitution of any names contained within.

Methods

substituteNames :: Map VName VName -> a -> a

substituteNames m e replaces the variable names in e with new names, based on the mapping in m. It is assumed that all names in e are unique, i.e. there is no shadowing.

type Substitutable rep = (RepTypes rep, Substitute (ExpDec rep), Substitute (BodyDec rep), Substitute (LetDec rep), Substitute (FParamInfo rep), Substitute (LParamInfo rep), Substitute (RetType rep), Substitute (BranchType rep), Substitute (Op rep))

Representations in which all annotations support name substitution.