Futhark.Util

Description

Non-Futhark-specific utilities. If you find yourself writing general functions on generic data structures, consider putting them here.

Sometimes it is also preferable to copy a small function rather than introducing a large dependency. In this case, make sure to note where you got it from (and make sure that the license is compatible).

Documentation

nubOrd :: Ord a => [a] -> [a]

Like nub, but without the quadratic runtime.

nubByOrd :: (a -> a -> Ordering) -> [a] -> [a]

Like nubBy, but without the quadratic runtime.

mapAccumLM :: Monad m => (acc -> x -> m (acc, y)) -> acc -> [x] -> m (acc, [y])

Like mapAccumL, but monadic.

maxinum :: (Num a, Ord a, Foldable f) => f a -> a

Like maximum, but returns zero for an empty list.

chunk :: Int -> [a] -> [[a]]

chunk n a splits a into n-size-chunks. If the length of a is not divisible by n, the last chunk will have fewer than n elements (but it will never be empty).

chunks :: [Int] -> [a] -> [[a]]

chunks ns a splits a into chunks determined by the elements of ns. It must hold that sum ns == length a, or the resulting list may contain too few chunks, or not all elements of a.

pairs :: [a] -> [(a, a)]

pairs l chunks the list into pairs of consecutive elements, ignoring any excess element. Example: pairs [a,b,c,d] == [(a,b),(c,d)].

unpairs :: [(a, a)] -> [a]

The opposite of pairs: unpairs [(a,b),(c,d)] = [a,b,c,d].

dropAt :: Int -> Int -> [a] -> [a]

dropAt i n drops n elements starting at element i.

takeLast :: Int -> [a] -> [a]

takeLast n l takes the last n elements of l.

dropLast :: Int -> [a] -> [a]

dropLast n l drops the last n elements of l.

mapEither :: (a -> Either b c) -> [a] -> ([b], [c])

A combination of map and partitionEithers.

maybeNth :: Integral int => int -> [a] -> Maybe a

Return the list element at the given index, if the index is valid.

maybeHead :: [a] -> Maybe a

Return the first element of the list, if it exists.

splitFromEnd :: Int -> [a] -> ([a], [a])

Like splitAt, but from the end.

splitAt3 :: Int -> Int -> [a] -> ([a], [a], [a])

Like splitAt, but produces three lists.

focusNth :: Integral int => int -> [a] -> Maybe ([a], a, [a])

Return the list element at the given index, if the index is valid, along with the elements before and after.

hashText :: Text -> Text

Compute a hash of a text that is stable across OS versions. Returns the hash as a text as well, ready for human consumption.

unixEnvironment :: [(String, String)]

The Unix environment when the Futhark compiler started.

isEnvVarAtLeast :: String -> Int -> Bool

True if the environment variable, viewed as an integer, has at least this numeric value. Returns False if variable is unset or not numeric.

fancyTerminal :: Bool

Are we running in a terminal capable of fancy commands and visualisation?

runProgramWithExitCode :: FilePath -> [String] -> ByteString -> IO (Either IOException (ExitCode, String, String))

Like readProcessWithExitCode, but also wraps exceptions when the indicated binary cannot be launched, or some other exception is thrown. Also does shenanigans to handle improperly encoded outputs.

directoryContents :: FilePath -> IO [FilePath]

Every non-directory file contained in a directory tree.

roundFloat :: Float -> Float

Round a single-precision floating point number correctly.

ceilFloat :: Float -> Float

Round a single-precision floating point number upwards correctly.

floorFloat :: Float -> Float

Round a single-precision floating point number downwards correctly.

roundDouble :: Double -> Double

Round a double-precision floating point number correctly.

ceilDouble :: Double -> Double

Round a double-precision floating point number upwards correctly.

floorDouble :: Double -> Double

Round a double-precision floating point number downwards correctly.

lgamma :: Double -> Double

The system-level lgamma() function.

lgammaf :: Float -> Float

The system-level lgammaf() function.

tgamma :: Double -> Double

The system-level tgamma() function.

tgammaf :: Float -> Float

The system-level tgammaf() function.

hypot :: Double -> Double -> Double

The system-level hypot function.

hypotf :: Float -> Float -> Float

The system-level hypotf function.

fromPOSIX :: FilePath -> FilePath

Some bad operating systems do not use forward slash as directory separator - this is where we convert Futhark includes (which always use forward slash) to native paths.

toPOSIX :: FilePath -> FilePath

Turn a POSIX filepath into a filepath for the native system.

trim :: String -> String

Remove leading and trailing whitespace from a string. Not an efficient implementation!

pmapIO :: Maybe Int -> (a -> IO b) -> [a] -> IO [b]

Run various IO actions concurrently, possibly with a bound on the number of threads. The list must be finite. The ordering of the result list is not deterministic - add your own sorting if needed. If any of the actions throw an exception, then that exception is propagated to this function.

readFileSafely :: FilePath -> IO (Maybe (Either String Text))

Read a file, returning Nothing if the file does not exist, and Left if some other error occurs.

convFloat :: (RealFloat from, RealFloat to) => from -> to

Convert between different floating-point types, preserving infinities and NaNs.

type UserString = String

As the user typed it.

type EncodedString = String

Encoded form.

zEncodeString :: UserString -> EncodedString

Z-encode a string using a slightly simplified variant of GHC Z-encoding. The encoded string is a valid identifier in most programming languages.

atMostChars :: Int -> String -> String

invertMap :: (Ord v, Ord k) => Map k v -> Map v (Set k)

traverseFold :: (Monoid m, Traversable t, Applicative f) => (a -> f m) -> t a -> f m

fixPoint :: Eq a => (a -> a) -> a -> a