Futhark.Util.IntegralExp

Description

It is occasionally useful to define generic functions that can not only compute their result as an integer, but also as a symbolic expression in the form of an AST.

There are some Haskell hacks for this - it is for example not hard to define an instance of Num that constructs an AST. However, this falls down for some other interesting classes, like Integral, which requires both the problematic method fromInteger, and also that the type is an instance of Enum.

We can always just define hobbled instances that call error for those methods that are impractical, but this is ugly.

Hence, this module defines similes to standard Haskell numeric typeclasses that have been modified to make generic functions slightly easier to write.

Documentation

class Num e => IntegralExp e where

A twist on the Integral type class that is more friendly to symbolic representations.

Minimal complete definition

quot, rem, div, mod, sgn

Methods

quot :: e -> e -> e

rem :: e -> e -> e

div :: e -> e -> e

mod :: e -> e -> e

sgn :: e -> Maybe Int

divUp :: e -> e -> e

Like div, but rounds towards positive infinity.

newtype Wrapped a

This wrapper allows you to use a type that is an instance of the true class whenever the simile class is required.

Constructors

Wrapped

Fields

wrappedValue :: a