Futhark.Util.Log

Description

Opaque type for an operations log that provides fast O(1) appends.

Documentation

data Log

An efficiently catenable sequence of log entries.

toText :: Log -> Text

Transform a log into text. Every log entry becomes its own line (or possibly more, in case of multi-line entries).

class ToLog a where

Typeclass for things that can be turned into a single-entry log.

Methods

toLog :: a -> Log

class (Applicative m, Monad m) => MonadLogger m where

Typeclass for monads that support logging.

Minimal complete definition

addLog

Methods

logMsg :: ToLog a => a -> m ()

Add one log entry.

addLog :: Log -> m ()

Append an entire log.