Futhark.Util.Options

Description

Common code for parsing command line options based on getopt.

Documentation

type FunOptDescr cfg = OptDescr (Either (IO ()) (cfg -> cfg))

A command line option that either purely updates a configuration, or performs an IO action (and stops).

mainWithOptions :: cfg -> [FunOptDescr cfg] -> String -> ([String] -> cfg -> Maybe (IO ())) -> String -> [String] -> IO ()

Generate a main action that parses the given command line options (while always adding commonOptions).

commonOptions :: String -> String -> [FunOptDescr cfg] -> [FunOptDescr cfg]

Common definitions for -v and -h, given the list of all other options.

optionsError :: String -> IO ()

Terminate the program with this error message (but don't report it as an ICE, as happens with error).