Futhark.Util.Table

Description

Basic table building for prettier futhark-test output.

Documentation

buildTable :: [[Entry]] -> Int -> String

Builds a table from a list of entries and a padding amount that determines padding from the right side of the widest entry in each column.

mkEntry :: String -> (String, [SGR])

Makes a table entry with the default SGR mode.

type Entry = (String, [SGR])

A table entry. Consists of the content as well a list of SGR commands to color/stylelize the entry.