Language.Futhark.Core

Description

This module contains very basic definitions for Futhark - so basic, that they can be shared between the internal and external representation.

Documentation

data Uniqueness

The uniqueness attribute of a type. This essentially indicates whether or not in-place modifications are acceptable. With respect to ordering, Unique is greater than Nonunique.

Constructors

Nonunique
May have references outside current function.

Unique
No references outside current function.

Location utilities

data SrcLoc

Source location type. Source location are all equal, which allows AST nodes to be compared modulo location information.

data Loc

Location type, consisting of a beginning position and an end position.

class Located a where

Located values have a location.

Minimal complete definition

locOf

Methods

locOf :: a -> Loc

locOfList :: [a] -> Loc

srclocOf :: Located a => a -> SrcLoc

The SrcLoc of a Located value.

locStr :: Located a => a -> String

A human-readable location string, of the form filename:lineno:columnno. This follows the GNU coding standards for error messages: https://www.gnu.org/prep/standards/html_node/Errors.html

This function assumes that both start and end position is in the same file (it is not clear what the alternative would even mean).

locStrRel :: (Located a, Located b) => a -> b -> String

Like locStr, but locStrRel prev now prints the location now with the file name left out if the same as prev. This is useful when printing messages that are all in the context of some initially printed location (e.g. the first mention contains the file name; the rest just line and column name).

prettyStacktrace :: Int -> [String] -> String

Given a list of strings representing entries in the stack trace and the index of the frame to highlight, produce a final newline-terminated string for showing to the user. This string should also be preceded by a newline. The most recent stack frame must come first in the list.

Name handling

data Name

The abstract (not really) type representing names in the Futhark compiler. Strings, being lists of characters, are very slow, while Texts are based on byte-arrays.

nameToString :: Name -> String

Convert a name to the corresponding list of characters.

nameFromString :: String -> Name

Convert a list of characters to the corresponding name.

nameToText :: Name -> Text

Convert a name to the corresponding Text.

nameFromText :: Text -> Name

Convert a Text to the corresponding name.

data VName

A name tagged with some integer. Only the integer is used in comparisons, no matter the type of vn.

Constructors

VName !Name !Int

baseTag :: VName -> Int

Return the tag contained in the VName.

baseName :: VName -> Name

Return the name contained in the VName.

baseString :: VName -> String

Return the base Name converted to a string.

quote :: String -> String

Enclose a string in the prefered quotes used in error messages. These are picked to not collide with characters permitted in identifiers.

pquote :: Doc -> Doc

As quote, but works on prettyprinted representation.

Special identifiers

defaultEntryPoint :: Name

The name of the default program entry point (main).

Number re-export

data Int8

data Int16

data Int32

data Int64

data Word8

data Word16

data Word32

data Word64

data Half