Language.Futhark.Interpreter

Description

An interpreter operating on type-checked source Futhark terms. Relatively slow.

Documentation

data Ctx

The interpreter context. All evaluation takes place with respect to a context, and it can be extended with more definitions, which is how the REPL works.

Constructors

Ctx

Fields

ctxEnv :: Env

ctxImports :: Map FilePath Env

data Env

The actual type- and value environment.

data InterpreterError

An error occurred during interpretation due to an error in the user program. Actual interpreter errors will be signaled with an IO exception (error).

initialCtx :: Ctx

The initial environment contains definitions of the various intrinsic functions.

interpretExp :: Ctx -> Exp -> F ExtOp Value

interpretDec :: Ctx -> Dec -> F ExtOp Ctx

interpretImport :: Ctx -> (FilePath, Prog) -> F ExtOp Ctx

interpretFunction :: Ctx -> VName -> [Value] -> Either String (F ExtOp Value)

Execute the named function on the given arguments; may fail horribly if these are ill-typed.

data ExtOp a

Constructors

ExtOpTrace String String a

ExtOpBreak Loc BreakReason (NonEmpty StackFrame) a

ExtOpError InterpreterError

data BreakReason

What is the reason for this break point?

Constructors

BreakPoint
An explicit breakpoint in the program.

BreakNaN
A

data StackFrame

Constructors

StackFrame

Fields

stackFrameLoc :: Loc

stackFrameCtx :: Ctx

typeCheckerEnv :: Env -> Env

data Value

A fully evaluated Futhark value.

Constructors

ValuePrim !PrimValue

ValueArray ValueShape !(Array Int Value)

ValueRecord (Map Name Value)

fromTuple :: Value -> Maybe [Value]

isEmptyArray :: Value -> Bool

Does the value correspond to an empty array?

prettyEmptyArray :: TypeBase () () -> Value -> String

String representation of an empty array with the provided element type. This is pretty ad-hoc - don't expect good results unless the element type is a primitive.