Language.Futhark.Parser

Description

Interface to the Futhark parser.

Documentation

parseFuthark :: FilePath -> Text -> Either ParseError UncheckedProg

Parse an entire Futhark program from the given Text, using the FilePath as the source name for error messages.

parseExp :: FilePath -> Text -> Either ParseError UncheckedExp

Parse an Futhark expression from the given String, using the FilePath as the source name for error messages.

parseModExp :: FilePath -> Text -> Either ParseError (ModExpBase NoInfo Name)

Parse a Futhark module expression from the given String, using the FilePath as the source name for error messages.

parseType :: FilePath -> Text -> Either ParseError UncheckedTypeExp

Parse an Futhark type from the given String, using the FilePath as the source name for error messages.

parseValue :: FilePath -> Text -> Either ParseError Value

Parse any Futhark value from the given String, using the FilePath as the source name for error messages.

parseValues :: FilePath -> Text -> Either ParseError [Value]

Parse several Futhark values (separated by anything) from the given String, using the FilePath as the source name for error messages.

parseDecOrExpIncrM :: Monad m => m Text -> FilePath -> Text -> m (Either ParseError (Either UncheckedDec UncheckedExp))

Parse either an expression or a declaration incrementally; favouring declarations in case of ambiguity.

data ParseError

A parse error. Use show to get a human-readable description.

Constructors

ParseError String

scanTokensText :: Pos -> Text -> Either String ([L Token], Pos)

Given a starting position, produce tokens from the given text (or a lexer error). Returns the final position.

data L a

A value tagged with a source location.

Constructors

L SrcLoc a

data Token

A lexical token. It does not itself contain position information, so in practice the parser will consume tokens tagged with a source position.

Constructors

ID Name

INDEXING Name

QUALINDEXING [Name] Name

QUALPAREN [Name] Name

SYMBOL BinOp [Name] Name

CONSTRUCTOR Name

PROJ_INTFIELD Name

INTLIT Integer

STRINGLIT Text

I8LIT Int8

I16LIT Int16

I32LIT Int32

I64LIT Int64

U8LIT Word8

U16LIT Word16

U32LIT Word32

U64LIT Word64

FLOATLIT Double

F16LIT Half

F32LIT Float

F64LIT Double

CHARLIT Char

COLON

COLON_GT

BACKSLASH

APOSTROPHE

APOSTROPHE_THEN_HAT

APOSTROPHE_THEN_TILDE

BACKTICK

HASH_LBRACKET

DOT

TWO_DOTS

TWO_DOTS_LT

TWO_DOTS_GT

THREE_DOTS

LPAR

RPAR

RPAR_THEN_LBRACKET

LBRACKET

RBRACKET

LCURLY

RCURLY

COMMA

UNDERSCORE

RIGHT_ARROW

QUESTION_MARK

EQU

ASTERISK

NEGATE

BANG

DOLLAR

LTH

HAT

TILDE

PIPE

IF

THEN

ELSE

DEF

LET

LOOP

IN

FOR

DO

WITH

ASSERT

TRUE

FALSE

WHILE

INCLUDE

IMPORT

ENTRY

TYPE

MODULE

VAL

OPEN

LOCAL

MATCH

CASE

DOC String

EOF