Language.Futhark.Pretty

Description

Futhark prettyprinter. This module defines Pretty instances for the AST defined in Language.Futhark.Syntax.

Documentation

pretty :: Pretty a => a -> String

Prettyprint a value, wrapped to 80 characters.

prettyTuple :: Pretty a => [a] -> String

Prettyprint a list enclosed in curly braces.

leadingOperator :: Name -> BinOp

Given an operator name, return the operator that determines its syntactical properties.

class IsName v where

A class for types that are variable names in the Futhark source language. This is used instead of a mere Pretty instance because in the compiler frontend we want to print VNames differently depending on whether the FUTHARK_COMPILER_DEBUGGING environment variable is set, yet in the backend we want to always print VNames with the tag. To avoid erroneously using the Pretty instance for VNames, we in fact only define it inside the modules for the core language (as an orphan instance).

Methods

pprName :: v -> Doc

prettyName :: IsName v => v -> String

Prettyprint a name to a string.

class Annot f where

Class for type constructors that represent annotations. Used in the prettyprinter to either print the original AST, or the computed decoration.

Methods

unAnnot :: f a -> Maybe a

Extract value, if any.

Orphan instances

Pretty Liftedness

Instance details

Methods

ppr :: Liftedness -> Doc

pprPrec :: Int -> Liftedness -> Doc

pprList :: [Liftedness] -> Doc

Pretty PatLit

Instance details

Methods

ppr :: PatLit -> Doc

pprPrec :: Int -> PatLit -> Doc

pprList :: [PatLit] -> Doc

Pretty Value

Instance details

Methods

ppr :: Value -> Doc

pprPrec :: Int -> Value -> Doc

pprList :: [Value] -> Doc

Pretty PrimValue

Instance details

Methods

ppr :: PrimValue -> Doc

pprPrec :: Int -> PrimValue -> Doc

pprList :: [PrimValue] -> Doc

(Eq vn, IsName vn) => Pretty (TypeParamBase vn)

Instance details

Methods

ppr :: TypeParamBase vn -> Doc

pprPrec :: Int -> TypeParamBase vn -> Doc

pprList :: [TypeParamBase vn] -> Doc

IsName vn => Pretty (SizeBinder vn)

Instance details

Methods

ppr :: SizeBinder vn -> Doc

pprPrec :: Int -> SizeBinder vn -> Doc

pprList :: [SizeBinder vn] -> Doc

IsName vn => Pretty (QualName vn)

Instance details

Methods

ppr :: QualName vn -> Doc

pprPrec :: Int -> QualName vn -> Doc

pprList :: [QualName vn] -> Doc

(Eq vn, IsName vn) => Pretty (TypeArgExp vn)

Instance details

Methods

ppr :: TypeArgExp vn -> Doc

pprPrec :: Int -> TypeArgExp vn -> Doc

pprList :: [TypeArgExp vn] -> Doc

(Eq vn, IsName vn) => Pretty (TypeExp vn)

Instance details

Methods

ppr :: TypeExp vn -> Doc

pprPrec :: Int -> TypeExp vn -> Doc

pprList :: [TypeExp vn] -> Doc

IsName vn => Pretty (DimExp vn)

Instance details

Methods

ppr :: DimExp vn -> Doc

pprPrec :: Int -> DimExp vn -> Doc

pprList :: [DimExp vn] -> Doc

Pretty (ShapeDecl dim) => Pretty (TypeArg dim)

Instance details

Methods

ppr :: TypeArg dim -> Doc

pprPrec :: Int -> TypeArg dim -> Doc

pprList :: [TypeArg dim] -> Doc

Pretty (ShapeDecl Bool)

Instance details

Methods

ppr :: ShapeDecl Bool -> Doc

pprPrec :: Int -> ShapeDecl Bool -> Doc

pprList :: [ShapeDecl Bool] -> Doc

Pretty (ShapeDecl Int64)

Instance details

Methods

ppr :: ShapeDecl Int64 -> Doc

pprPrec :: Int -> ShapeDecl Int64 -> Doc

pprList :: [ShapeDecl Int64] -> Doc

Pretty (ShapeDecl ())

Instance details

Methods

ppr :: ShapeDecl () -> Doc

pprPrec :: Int -> ShapeDecl () -> Doc

pprList :: [ShapeDecl ()] -> Doc

IsName vn => Pretty (ShapeDecl (DimDecl vn))

Instance details

Methods

ppr :: ShapeDecl (DimDecl vn) -> Doc

pprPrec :: Int -> ShapeDecl (DimDecl vn) -> Doc

pprList :: [ShapeDecl (DimDecl vn)] -> Doc

IsName vn => Pretty (DimDecl vn)

Instance details

Methods

ppr :: DimDecl vn -> Doc

pprPrec :: Int -> DimDecl vn -> Doc

pprList :: [DimDecl vn] -> Doc

IsName vn => Pretty (AttrInfo vn)

Instance details

Methods

ppr :: AttrInfo vn -> Doc

pprPrec :: Int -> AttrInfo vn -> Doc

pprList :: [AttrInfo vn] -> Doc

IsName vn => Pretty (AttrAtom vn)

Instance details

Methods

ppr :: AttrAtom vn -> Doc

pprPrec :: Int -> AttrAtom vn -> Doc

pprList :: [AttrAtom vn] -> Doc

(Eq vn, IsName vn, Annot f) => Pretty (ProgBase f vn)

Instance details

Methods

ppr :: ProgBase f vn -> Doc

pprPrec :: Int -> ProgBase f vn -> Doc

pprList :: [ProgBase f vn] -> Doc

(Eq vn, IsName vn, Annot f) => Pretty (DecBase f vn)

Instance details

Methods

ppr :: DecBase f vn -> Doc

pprPrec :: Int -> DecBase f vn -> Doc

pprList :: [DecBase f vn] -> Doc

(Eq vn, IsName vn, Annot f) => Pretty (ModParamBase f vn)

Instance details

Methods

ppr :: ModParamBase f vn -> Doc

pprPrec :: Int -> ModParamBase f vn -> Doc

pprList :: [ModParamBase f vn] -> Doc

(Eq vn, IsName vn, Annot f) => Pretty (ModBindBase f vn)

Instance details

Methods

ppr :: ModBindBase f vn -> Doc

pprPrec :: Int -> ModBindBase f vn -> Doc

pprList :: [ModBindBase f vn] -> Doc

(Eq vn, IsName vn, Annot f) => Pretty (ModExpBase f vn)

Instance details

Methods

ppr :: ModExpBase f vn -> Doc

pprPrec :: Int -> ModExpBase f vn -> Doc

pprList :: [ModExpBase f vn] -> Doc

(Eq vn, IsName vn, Annot f) => Pretty (SigBindBase f vn)

Instance details

Methods

ppr :: SigBindBase f vn -> Doc

pprPrec :: Int -> SigBindBase f vn -> Doc

pprList :: [SigBindBase f vn] -> Doc

(Eq vn, IsName vn, Annot f) => Pretty (SigExpBase f vn)

Instance details

Methods

ppr :: SigExpBase f vn -> Doc

pprPrec :: Int -> SigExpBase f vn -> Doc

pprList :: [SigExpBase f vn] -> Doc

(Eq vn, IsName vn, Annot f) => Pretty (SpecBase f vn)

Instance details

Methods

ppr :: SpecBase f vn -> Doc

pprPrec :: Int -> SpecBase f vn -> Doc

pprList :: [SpecBase f vn] -> Doc

(Eq vn, IsName vn, Annot f) => Pretty (TypeBindBase f vn)

Instance details

Methods

ppr :: TypeBindBase f vn -> Doc

pprPrec :: Int -> TypeBindBase f vn -> Doc

pprList :: [TypeBindBase f vn] -> Doc

(Eq vn, IsName vn, Annot f) => Pretty (ValBindBase f vn)

Instance details

Methods

ppr :: ValBindBase f vn -> Doc

pprPrec :: Int -> ValBindBase f vn -> Doc

pprList :: [ValBindBase f vn] -> Doc

(Eq vn, IsName vn, Annot f) => Pretty (PatBase f vn)

Instance details

Methods

ppr :: PatBase f vn -> Doc

pprPrec :: Int -> PatBase f vn -> Doc

pprList :: [PatBase f vn] -> Doc

(Eq vn, IsName vn, Annot f) => Pretty (LoopFormBase f vn)

Instance details

Methods

ppr :: LoopFormBase f vn -> Doc

pprPrec :: Int -> LoopFormBase f vn -> Doc

pprList :: [LoopFormBase f vn] -> Doc

(Eq vn, IsName vn, Annot f) => Pretty (CaseBase f vn)

Instance details

Methods

ppr :: CaseBase f vn -> Doc

pprPrec :: Int -> CaseBase f vn -> Doc

pprList :: [CaseBase f vn] -> Doc

(Eq vn, IsName vn, Annot f) => Pretty (FieldBase f vn)

Instance details

Methods

ppr :: FieldBase f vn -> Doc

pprPrec :: Int -> FieldBase f vn -> Doc

pprList :: [FieldBase f vn] -> Doc

(Eq vn, IsName vn, Annot f) => Pretty (ExpBase f vn)

Instance details

Methods

ppr :: ExpBase f vn -> Doc

pprPrec :: Int -> ExpBase f vn -> Doc

pprList :: [ExpBase f vn] -> Doc

(Eq vn, IsName vn, Annot f) => Pretty (AppExpBase f vn)

Instance details

Methods

ppr :: AppExpBase f vn -> Doc

pprPrec :: Int -> AppExpBase f vn -> Doc

pprList :: [AppExpBase f vn] -> Doc

(Eq vn, IsName vn, Annot f) => Pretty (DimIndexBase f vn)

Instance details

Methods

ppr :: DimIndexBase f vn -> Doc

pprPrec :: Int -> DimIndexBase f vn -> Doc

pprList :: [DimIndexBase f vn] -> Doc

IsName vn => Pretty (IdentBase f vn)

Instance details

Methods

ppr :: IdentBase f vn -> Doc

pprPrec :: Int -> IdentBase f vn -> Doc

pprList :: [IdentBase f vn] -> Doc

(Eq vn, IsName vn, Annot f) => Pretty (TypeDeclBase f vn)

Instance details

Methods

ppr :: TypeDeclBase f vn -> Doc

pprPrec :: Int -> TypeDeclBase f vn -> Doc

pprList :: [TypeDeclBase f vn] -> Doc

Pretty (ShapeDecl dim) => Pretty (TypeBase dim as)

Instance details

Methods

ppr :: TypeBase dim as -> Doc

pprPrec :: Int -> TypeBase dim as -> Doc

pprList :: [TypeBase dim as] -> Doc

Pretty (ShapeDecl dim) => Pretty (ScalarTypeBase dim as)

Instance details

Methods

ppr :: ScalarTypeBase dim as -> Doc

pprPrec :: Int -> ScalarTypeBase dim as -> Doc

pprList :: [ScalarTypeBase dim as] -> Doc

Pretty (ShapeDecl dim) => Pretty (RetTypeBase dim as)

Instance details

Methods

ppr :: RetTypeBase dim as -> Doc

pprPrec :: Int -> RetTypeBase dim as -> Doc

pprList :: [RetTypeBase dim as] -> Doc