Language.Futhark.Prop

Description

This module provides various simple ways to query and manipulate fundamental Futhark terms, such as types and values. The intent is to keep Futhark.Language.Syntax simple, and put whatever embellishments we need here.

Various

data Intrinsic

The nature of something predefined. For functions, these can either be monomorphic or overloaded. An overloaded builtin is a list valid types it can be instantiated with, to the parameter and result type, with Nothing representing the overloaded parameter type.

Constructors

IntrinsicMonoFun [PrimType] PrimType

IntrinsicOverloadedFun [PrimType] [Maybe PrimType] (Maybe PrimType)

IntrinsicPolyFun [TypeParamBase VName] [StructType] (RetTypeBase (DimDecl VName) ())

IntrinsicType Liftedness [TypeParamBase VName] StructType

IntrinsicEquality

intrinsics :: Map VName Intrinsic

A map of all built-ins.

maxIntrinsicTag :: Int

The largest tag used by an intrinsic - this can be used to determine whether a VName refers to an intrinsic or a user-defined name.

namesToPrimTypes :: Map Name PrimType

Names of primitive types to types. This is only valid if no shadowing is going on, but useful for tools.

qualName :: v -> QualName v

Create a name with no qualifiers from a name.

qualify :: v -> QualName v -> QualName v

Add another qualifier (at the head) to a qualified name.

typeName :: VName -> TypeName

Create a type name name with no qualifiers from a VName.

valueType :: Value -> ValueType

The type of the value.

primValueType :: PrimValue -> PrimType

The type of a basic value.

leadingOperator :: Name -> BinOp

Given an operator name, return the operator that determines its syntactical properties.

progImports :: ProgBase f vn -> [(String, SrcLoc)]

The modules imported by a Futhark program.

decImports :: DecBase f vn -> [(String, SrcLoc)]

The modules imported by a single declaration.

progModuleTypes :: Ord vn => ProgBase f vn -> Set vn

The set of module types used in any exported (non-local) declaration.

identifierReference :: String -> Maybe ((String, String, Maybe FilePath), String)

Extract a leading ((name, namespace, file), remainder) from a documentation comment string. These are formatted as `name`@namespace[@file]. Let us hope that this pattern does not occur anywhere else.

prettyStacktrace :: Int -> [String] -> String

Given a list of strings representing entries in the stack trace and the index of the frame to highlight, produce a final newline-terminated string for showing to the user. This string should also be preceded by a newline. The most recent stack frame must come first in the list.

Queries on expressions

typeOf :: ExpBase Info VName -> PatType

The type of an Futhark term. The aliasing will refer to itself, if the term is a non-tuple-typed variable.

valBindTypeScheme :: ValBindBase Info VName -> ([TypeParamBase VName], StructType)

The type scheme of a value binding, comprising the type parameters and the actual type.

valBindBound :: ValBindBase Info VName -> [VName]

The names that are brought into scope by this value binding (not including its own parameter names, but including any existential sizes).

funType :: [PatBase Info VName] -> StructRetType -> StructType

The type of a function with the given parameters and return type.

Queries on patterns and params

patIdents :: (Functor f, Ord vn) => PatBase f vn -> Set (IdentBase f vn)

The set of identifiers bound in a pattern.

patNames :: (Functor f, Ord vn) => PatBase f vn -> Set vn

The set of names bound in a pattern.

patternMap :: Functor f => PatBase f VName -> Map VName (IdentBase f VName)

A mapping from names bound in a map to their identifier.

patternType :: PatBase Info VName -> PatType

The type of values bound by the pattern.

patternStructType :: PatBase Info VName -> StructType

The type matched by the pattern, including shape declarations if present.

patternParam :: PatBase Info VName -> (PName, StructType)

When viewed as a function parameter, does this pattern correspond to a named parameter of some type?

patternOrderZero :: PatBase Info vn -> Bool

patternOrderZero pat is True if all of the types in the given pattern have order 0.

patternDimNames :: PatBase Info VName -> Set VName

Extract all the shape names that occur in a given pattern.

Queries on types

uniqueness :: TypeBase shape as -> Uniqueness

Return the uniqueness of a type.

unique :: TypeBase shape as -> Bool

unique t is True if the type of the argument is unique.

aliases :: Monoid as => TypeBase shape as -> as

Return the set of all variables mentioned in the aliasing of a type.

diet :: TypeBase shape as -> Diet

diet t returns a description of how a function parameter of type t might consume its argument.

arrayRank :: TypeBase dim as -> Int

Return the dimensionality of a type. For non-arrays, this is zero. For a one-dimensional array it is one, for a two-dimensional it is two, and so forth.

arrayShape :: TypeBase dim as -> ShapeDecl dim

Return the shape of a type - for non-arrays, this is mempty.

nestedDims :: TypeBase (DimDecl VName) as -> [DimDecl VName]

Return any free shape declarations in the type, with duplicates removed.

orderZero :: TypeBase dim as -> Bool

orderZero t is True if the argument type has order 0, i.e., it is not a function type, does not contain a function type as a subcomponent, and may not be instantiated with a function type.

unfoldFunType :: TypeBase dim as -> ([TypeBase dim as], TypeBase dim as)

Extract the parameter types and return type from a type. If the type is not an arrow type, the list of parameter types is empty.

foldFunType :: Monoid as => [TypeBase dim as] -> RetTypeBase dim as -> TypeBase dim as

foldFunType ts ret creates a function type (Arrow) that takes ts as parameters and returns ret.

typeVars :: Monoid as => TypeBase dim as -> Set VName

The type names mentioned in a type.

typeDimNames :: TypeBase (DimDecl VName) als -> Set VName

Extract all the shape names that occur free in a given type.

Operations on types

peelArray :: Int -> TypeBase dim as -> Maybe (TypeBase dim as)

peelArray n t returns the type resulting from peeling the first n array dimensions from t. Returns Nothing if t has less than n dimensions.

stripArray :: Int -> TypeBase dim as -> TypeBase dim as

stripArray n t removes the n outermost layers of the array. Essentially, it is the type of indexing an array of type t with n indexes.

arrayOf :: Monoid as => TypeBase dim as -> ShapeDecl dim -> Uniqueness -> TypeBase dim as

arrayOf t s u constructs an array type. The convenience compared to using the Array constructor directly is that t can itself be an array. If t is an n-dimensional array, and s is a list of length n, the resulting type is of an n+m dimensions. The uniqueness of the new array will be u, no matter the uniqueness of t.

toStructural :: TypeBase dim as -> TypeBase () ()

Convert any type to one that has rank information, no alias information, and no embedded names.

toStruct :: TypeBase dim as -> TypeBase dim ()

Remove aliasing information from a type.

fromStruct :: TypeBase dim as -> TypeBase dim Aliasing

Replace no aliasing with an empty alias set.

setAliases :: TypeBase dim asf -> ast -> TypeBase dim ast

t `setAliases` als returns t, but with als substituted for any already present aliasing.

addAliases :: TypeBase dim asf -> (asf -> ast) -> TypeBase dim ast

t `addAliases` f returns t, but with any already present aliasing replaced by f applied to that aliasing.

setUniqueness :: TypeBase dim as -> Uniqueness -> TypeBase dim as

Set the uniqueness attribute of a type. If the type is a record or sum type, the uniqueness of its components will be modified.

noSizes :: TypeBase (DimDecl vn) as -> TypeBase () as

Change the shape of a type to be just the rank.

traverseDims :: forall f fdim tdim als. Applicative f => (Set VName -> DimPos -> fdim -> f tdim) -> TypeBase fdim als -> f (TypeBase tdim als)

Perform a traversal (possibly including replacement) on sizes that are parameters in a function type, but also including the type immediately passed to the function. Also passes along a set of the parameter names inside the type that have come in scope at the occurrence of the dimension.

data DimPos

Where does this dimension occur?

Constructors

PosImmediate
Immediately in the argument to traverseDims.

PosParam
In a function parameter type.

PosReturn
In a function return type.

mustBeExplicit :: StructType -> Set VName

Figure out which of the sizes in a binding type must be passed explicitly, because their first use is as something else than just an array dimension.

mustBeExplicitInType :: StructType -> Set VName

Figure out which of the sizes in a parameter type must be passed explicitly, because their first use is as something else than just an array dimension. mustBeExplicit is like this function, but first decomposes into parameter types.

tupleRecord :: [TypeBase dim as] -> ScalarTypeBase dim as

Create a record type corresponding to a tuple with the given element types.

isTupleRecord :: TypeBase dim as -> Maybe [TypeBase dim as]

Does this type corespond to a tuple? If so, return the elements of that tuple.

areTupleFields :: Map Name a -> Maybe [a]

Does this record map correspond to a tuple?

tupleFields :: [a] -> Map Name a

Construct a record map corresponding to a tuple.

tupleFieldNames :: [Name]

Increasing field names for a tuple (starts at 0).

sortFields :: Map Name a -> [(Name, a)]

Sort fields by their name; taking care to sort numeric fields by their numeric value. This ensures that tuples and tuple-like records match.

sortConstrs :: Map Name a -> [(Name, a)]

Sort the constructors of a sum type in some well-defined (but not otherwise significant) manner.

isTypeParam :: TypeParamBase vn -> Bool

Is this a TypeParamType?

isSizeParam :: TypeParamBase vn -> Bool

Is this a TypeParamDim?

combineTypeShapes :: (Monoid as, ArrayDim dim) => TypeBase dim as -> TypeBase dim as -> TypeBase dim as

Combine the shape information of types as much as possible. The first argument is the orignal type and the second is the type of the transformed expression. This is necessary since the original type may contain additional information (e.g., shape restrictions) from the user given annotation.

matchDims :: forall as m d1 d2. (Monoid as, Monad m) => ([VName] -> d1 -> d2 -> m d1) -> TypeBase d1 as -> TypeBase d2 as -> m (TypeBase d1 as)

Match the dimensions of otherwise assumed-equal types. The combining function is also passed the names bound within the type (from named parameters or return types).

Values of these types are produces by the parser. They use unadorned names and have no type information, apart from that which is syntactically required.

data NoInfo a

No information functor. Usually used for placeholder type- or aliasing information.

Constructors

NoInfo

type UncheckedType = TypeBase (ShapeDecl Name) ()

A type with no aliasing information but shape annotations.

type UncheckedTypeExp = TypeExp Name

An expression with no type annotations.

type UncheckedIdent = IdentBase NoInfo Name

An identifier with no type annotations.

type UncheckedTypeDecl = TypeDeclBase NoInfo Name

A type declaration with no expanded type.

type UncheckedDimIndex = DimIndexBase NoInfo Name

An index with no type annotations.

type UncheckedSlice = SliceBase NoInfo Name

A slice with no type annotations.

type UncheckedExp = ExpBase NoInfo Name

An expression with no type annotations.

type UncheckedModExp = ModExpBase NoInfo Name

A module expression with no type annotations.

type UncheckedSigExp = SigExpBase NoInfo Name

A module type expression with no type annotations.

type UncheckedTypeParam = TypeParamBase Name

A type parameter with no type annotations.

type UncheckedPat = PatBase NoInfo Name

A pattern with no type annotations.

type UncheckedValBind = ValBindBase NoInfo Name

A function declaration with no type annotations.

type UncheckedDec = DecBase NoInfo Name

A declaration with no type annotations.

type UncheckedSpec = SpecBase NoInfo Name

A spec with no type annotations.

type UncheckedProg = ProgBase NoInfo Name

A Futhark program with no type annotations.

type UncheckedCase = CaseBase NoInfo Name

A case (of a match expression) with no type annotations.