Language.Futhark.Query

Description

Facilities for answering queries about a program, such as "what appears at this source location", or "where is this name bound". The intent is that this is used as a building block for IDE-like functionality.

Documentation

data BoundTo

What a name is bound to.

Constructors

BoundTerm StructType Loc

BoundModule Loc

BoundModuleType Loc

BoundType Loc

boundLoc :: BoundTo -> Loc

Where was a bound variable actually bound? That is, what is the location of its definition?

data AtPos

Information about what is at the given source location.

Constructors

AtName (QualName VName) (Maybe BoundTo) Loc

atPos :: Imports -> Pos -> Maybe AtPos

Information about what's at the given source position. Returns Nothing if there is nothing there, including if the source position is invalid.

data Pos

Position type.

Constructors

Pos !FilePath !Int !Int !Int
Source file name, line, column, and character offset.

Line numbering starts at 1, column offset starts at 1, and character offset starts at 0.