Language.Futhark.Semantic

Description

Definitions of various semantic objects (*not* the Futhark semantics themselves).

Documentation

data ImportName

Canonical reference to a Futhark code file. Does not include the .fut extension. This is most often a path relative to the current working directory of the compiler.

mkInitialImport :: FilePath -> ImportName

Create an import name immediately from a file path specified by the user.

mkImportFrom :: ImportName -> String -> SrcLoc -> ImportName

We resolve '..' paths here and assume that no shenanigans are going on with symbolic links. If there is, too bad. Don't do that.

includeToFilePath :: ImportName -> FilePath

Create a .fut file corresponding to an ImportName.

includeToString :: ImportName -> String

Produce a human-readable canonicalized string from an ImportName.

data FileModule

The result of type checking some file. Can be passed to further invocations of the type checker.

Constructors

FileModule

Fields

fileAbs :: TySet

Abstract types.

fileEnv :: Env

fileProg :: Prog

type Imports = [(String, FileModule)]

A mapping from import names to imports. The ordering is significant.

data Namespace

The space inhabited by a name.

Constructors

Term
Functions and values.

Type

Signature

data Env

Modules produces environment with this representation.

Constructors

Env

Fields

envVtable :: Map VName BoundV

envTypeTable :: Map VName TypeBinding

envSigTable :: Map VName MTy

envModTable :: Map VName Mod

envNameMap :: NameMap

type TySet = Map (QualName VName) Liftedness

A mapping of abstract types to their liftedness.

data FunSig

A parametric functor consists of a set of abstract types, the environment of its parameter, and the resulting module type.

Constructors

FunSig

Fields

funSigAbs :: TySet

funSigMod :: Mod

funSigMty :: MTy

type NameMap = Map (Namespace, Name) (QualName VName)

A mapping from names (which always exist in some namespace) to a unique (tagged) name.

data BoundV

Type parameters, list of parameter types (optinally named), and return type. The type parameters are in scope in both parameter types and the return type. Non-functional values have only a return type.

Constructors

BoundV [TypeParam] StructType

data Mod

Representation of a module, which is either a plain environment, or a parametric module ("functor" in SML).

Constructors

ModEnv Env

ModFun FunSig

data TypeBinding

A binding from a name to its definition as a type. We allow a return type here to support type abbreviations that hide some inner sizes (these must necessarily be Lifted or SizeLifted).

Constructors

TypeAbbr Liftedness [TypeParam] StructRetType

data MTy

Representation of a module type.

Constructors

MTy

Fields

mtyAbs :: TySet

Abstract types in the module type.

mtyMod :: Mod