Language.Futhark.Syntax

Description

The Futhark source language AST definition. Many types, such as ExpBase, are parametrised by type and name representation. E.g. in a value of type ExpBase f vn, annotations are wrapped in the functor f, and all names are of type vn. See the https://futhark.readthedocs.org@ for a language reference, or this module may be a little hard to understand.

Documentation

module Language.Futhark.Core

pretty :: Pretty a => a -> String

Prettyprint a value, wrapped to 80 characters.

Types

data Uniqueness

The uniqueness attribute of a type. This essentially indicates whether or not in-place modifications are acceptable. With respect to ordering, Unique is greater than Nonunique.

Constructors

Nonunique
May have references outside current function.

Unique
No references outside current function.

data IntType

An integer type, ordered by size. Note that signedness is not a property of the type, but a property of the operations performed on values of these types.

Constructors

Int8

Int16

Int32

Int64

data FloatType

A floating point type.

Constructors

Float16

Float32

Float64

data PrimType

Low-level primitive types.

Constructors

Signed IntType

Unsigned IntType

FloatType FloatType

Bool

class Eq dim => ArrayDim dim where

A type class for things that can be array dimensions.

Methods

unifyDims :: dim -> dim -> Maybe dim

unifyDims x y combines x and y to contain their maximum common information, and fails if they conflict.

data DimDecl vn

Declaration of a dimension size.

Constructors

NamedDim (QualName vn)
The size of the dimension is this name, which must be in scope. In a return type, this will give rise to an assertion.

ConstDim Int
The size is a constant.

AnyDim (Maybe vn)
No known size - but still possibly given a unique name, so we can recognise e.g. type square [n] = [n][n]i32 and make square [] do the right thing. If Nothing, then this is a name distinct from any other. The type checker should _never_ produce these - they are a (hopefully temporary) thing introduced by defunctorisation and monomorphisation.

newtype ShapeDecl dim

The size of an array type is a list of its dimension sizes. If Nothing, that dimension is of a (statically) unknown size.

Constructors

ShapeDecl

Fields

shapeDims :: [dim]

shapeRank :: ShapeDecl dim -> Int

The number of dimensions contained in a shape.

stripDims :: Int -> ShapeDecl dim -> Maybe (ShapeDecl dim)

stripDims n shape strips the outer n dimensions from shape, returning Nothing if this would result in zero or fewer dimensions.

data TypeName

A type name consists of qualifiers (for error messages) and a VName (for equality checking).

Constructors

TypeName

Fields

typeQuals :: [VName]

typeLeaf :: VName

typeNameFromQualName :: QualName VName -> TypeName

Convert a QualName to a TypeName.

qualNameFromTypeName :: TypeName -> QualName VName

Convert a TypeName to a QualName.

data TypeBase dim as

An expanded Futhark type is either an array, or something that can be an element of an array. When comparing types for equality, function parameter names are ignored. This representation permits some malformed types (arrays of functions), but importantly rules out arrays-of-arrays.

Constructors

Scalar (ScalarTypeBase dim as)

Array as Uniqueness (ScalarTypeBase dim ()) (ShapeDecl dim)

data TypeArg dim

An argument passed to a type constructor.

Constructors

TypeArgDim dim SrcLoc

TypeArgType (TypeBase dim ()) SrcLoc

data DimExp vn

A dimension declaration expression for use in a TypeExp.

Constructors

DimExpNamed (QualName vn) SrcLoc
The size of the dimension is this name, which must be in scope.

DimExpConst Int SrcLoc
The size is a constant.

DimExpAny
No dimension declaration.

data TypeExp vn

An unstructured type with type variables and possibly shape declarations - this is what the user types in the source program. These are used to construct TypeBases in the type checker.

Constructors

TEVar (QualName vn) SrcLoc

TETuple [TypeExp vn] SrcLoc

TERecord [(Name, TypeExp vn)] SrcLoc

TEArray (TypeExp vn) (DimExp vn) SrcLoc

TEUnique (TypeExp vn) SrcLoc

TEApply (TypeExp vn) (TypeArgExp vn) SrcLoc

TEArrow (Maybe vn) (TypeExp vn) (TypeExp vn) SrcLoc

TESum [(Name, [TypeExp vn])] SrcLoc

TEDim [vn] (TypeExp vn) SrcLoc

data TypeArgExp vn

A type argument expression passed to a type constructor.

Constructors

TypeArgExpDim (DimExp vn) SrcLoc

TypeArgExpType (TypeExp vn)

data PName

The name (if any) of a function parameter. The Eq and Ord instances always compare values of this type equal.

Constructors

Named VName

Unnamed

data ScalarTypeBase dim as

Types that can be elements of arrays. This representation does allow arrays of records of functions, which is nonsensical, but it convolutes the code too much if we try to statically rule it out.

Constructors

Prim PrimType

TypeVar as Uniqueness TypeName [TypeArg dim]

Record (Map Name (TypeBase dim as))

Sum (Map Name [TypeBase dim as])

Arrow as PName (TypeBase dim as) (RetTypeBase dim as)
The aliasing corresponds to the lexical closure of the function.

data RetTypeBase dim as

Types that can appear to the right of a function arrow. This just means they can be existentially quantified.

Constructors

RetType

Fields

retDims :: [VName]

retType :: TypeBase dim as

type PatType = TypeBase (DimDecl VName) Aliasing

A type with aliasing information and shape annotations, used for describing the type patterns and expressions.

type StructType = TypeBase (DimDecl VName) ()

A "structural" type with shape annotations and no aliasing information, used for declarations.

type StructRetType = RetTypeBase (DimDecl VName) ()

The return type version of StructType.

type PatRetType = RetTypeBase (DimDecl VName) Aliasing

The return type version of PatType.

type ValueType = TypeBase Int64 ()

A value type contains full, manifest size information.

data Diet

Information about which parts of a value/type are consumed.

Constructors

RecordDiet (Map Name Diet)
Consumes these fields in the record.

SumDiet (Map Name [Diet])
Consume these parts of the constructors.

FuncDiet Diet Diet
A function that consumes its argument(s) like this. The final Diet should always be Observe, as there is no way for a function to consume its return value.

Consume
Consumes this value.

Observe
Only observes value in this position, does not consume.

data TypeDeclBase f vn

A declaration of the type of something.

Constructors

TypeDecl

Fields

declaredType :: TypeExp vn

The type declared by the user.

expandedType :: f StructType

The type deduced by the type checker.

Values

data IntValue

An integer value.

Constructors

Int8Value !Int8

Int16Value !Int16

Int32Value !Int32

Int64Value !Int64

data FloatValue

A floating-point value.

Constructors

Float16Value !Half

Float32Value !Float

Float64Value !Double

data PrimValue

Non-array values.

Constructors

SignedValue !IntValue

UnsignedValue !IntValue

FloatValue !FloatValue

BoolValue !Bool

class IsPrimValue v where

A class for converting ordinary Haskell values to primitive Futhark values.

Methods

primValue :: v -> PrimValue

data Value

Simple Futhark values. Values are fully evaluated and their type is always unambiguous.

Constructors

PrimValue !PrimValue

ArrayValue !(Array Int Value) ValueType
It is assumed that the array is 0-indexed. The type is the full type.

Abstract syntax tree

data AttrInfo vn

The payload of an attribute.

Constructors

AttrAtom (AttrAtom vn) SrcLoc

AttrComp Name [AttrInfo vn] SrcLoc

data AttrAtom vn

The value of an AttrAtom.

Constructors

AtomName Name

AtomInt Integer

data BinOp

Default binary operators.

Constructors

Backtick
A pseudo-operator standing in for any normal identifier used as an operator (they all have the same fixity). Binary Ops for Numbers

Plus

Minus

Pow

Times

Divide

Mod

Quot

Rem

ShiftR

ShiftL

Band

Xor

Bor

LogAnd

LogOr

Equal

NotEqual

Less

Leq

Greater

Geq

PipeRight
|>

PipeLeft
<| Misc

data IdentBase f vn

An identifier consists of its name and the type of the value bound to the identifier.

Constructors

Ident

Fields

identName :: vn

identType :: f PatType

identSrcLoc :: SrcLoc

data Inclusiveness a

Whether a bound for an end-point of a DimSlice or a range literal is inclusive or exclusive.

Constructors

DownToExclusive a

ToInclusive a
May be "down to" if step is negative.

UpToExclusive a

data DimIndexBase f vn

An indexing of a single dimension.

Constructors

DimFix (ExpBase f vn)

DimSlice (Maybe (ExpBase f vn)) (Maybe (ExpBase f vn)) (Maybe (ExpBase f vn))

type SliceBase f vn = [DimIndexBase f vn]

A slicing of an array (potentially multiple dimensions).

data SizeBinder vn

A binding of a size in a pattern (essentially a size parameter in a let expression).

Constructors

SizeBinder

Fields

sizeName :: !vn

sizeLoc :: !SrcLoc

data AppExpBase f vn

An "application expression" is a semantic (not syntactic) grouping of expressions that have "funcall-like" semantics, mostly meaning that they can return existential sizes. In our type theory, these are all thought to be bound to names (*Administrative Normal Form*), but as this is not practical in a real language, we instead use an annotation (AppRes) that stores the information we need, so we can pretend that an application expression was really bound to a name.

Constructors

Apply (ExpBase f vn) (ExpBase f vn) (f (Diet, Maybe VName)) SrcLoc
The Maybe VName is a possible existential size that is instantiated by this argument. May have duplicates across the program, but they will all produce the same value (the expressions will be identical).

Coerce (ExpBase f vn) (TypeDeclBase f vn) SrcLoc
Size coercion: e :> t.

Range (ExpBase f vn) (Maybe (ExpBase f vn)) (Inclusiveness (ExpBase f vn)) SrcLoc

LetPat [SizeBinder vn] (PatBase f vn) (ExpBase f vn) (ExpBase f vn) SrcLoc

LetFun vn ([TypeParamBase vn], [PatBase f vn], Maybe (TypeExp vn), f StructRetType, ExpBase f vn) (ExpBase f vn) SrcLoc

If (ExpBase f vn) (ExpBase f vn) (ExpBase f vn) SrcLoc

DoLoop [VName] (PatBase f vn) (ExpBase f vn) (LoopFormBase f vn) (ExpBase f vn) SrcLoc

BinOp (QualName vn, SrcLoc) (f PatType) (ExpBase f vn, f (StructType, Maybe VName)) (ExpBase f vn, f (StructType, Maybe VName)) SrcLoc

LetWith (IdentBase f vn) (IdentBase f vn) (SliceBase f vn) (ExpBase f vn) (ExpBase f vn) SrcLoc

Index (ExpBase f vn) (SliceBase f vn) SrcLoc

Match (ExpBase f vn) (NonEmpty (CaseBase f vn)) SrcLoc
A match expression.

data AppRes

An annotation inserted by the type checker on constructs that are "function calls" (either literally or conceptually). This annotation encodes the result type, as well as any existential sizes that are generated here.

Constructors

AppRes

Fields

appResType :: PatType

appResExt :: [VName]

data ExpBase f vn

The Futhark expression language.

This allows us to encode whether or not the expression has been type-checked in the Haskell type of the expression. Specifically, the parser will produce expressions of type Exp NoInfo Name, and the type checker will convert these to Exp Info VName, in which type information is always present and all names are unique.

Constructors

Literal PrimValue SrcLoc

IntLit Integer (f PatType) SrcLoc
A polymorphic integral literal.

FloatLit Double (f PatType) SrcLoc
A polymorphic decimal literal.

StringLit [Word8] SrcLoc
A string literal is just a fancy syntax for an array of bytes.

Var (QualName vn) (f PatType) SrcLoc

Parens (ExpBase f vn) SrcLoc
A parenthesized expression.

QualParens (QualName vn, SrcLoc) (ExpBase f vn) SrcLoc

TupLit [ExpBase f vn] SrcLoc
Tuple literals, e.g., {1+3, {x, y+z}}.

RecordLit [FieldBase f vn] SrcLoc
Record literals, e.g. {x=2,y=3,z}.

ArrayLit [ExpBase f vn] (f PatType) SrcLoc
Array literals, e.g., [ [1+x, 3], [2, 1+4] ]. Second arg is the row type of the rows of the array.

Attr (AttrInfo vn) (ExpBase f vn) SrcLoc
An attribute applied to the following expression.

Project Name (ExpBase f vn) (f PatType) SrcLoc

Negate (ExpBase f vn) SrcLoc
Numeric negation (ugly special case; Haskell did it first).

Not (ExpBase f vn) SrcLoc
Logical and bitwise negation.

Assert (ExpBase f vn) (ExpBase f vn) (f String) SrcLoc
Fail if the first expression does not return true, and return the value of the second expression if it does.

Constr Name [ExpBase f vn] (f PatType) SrcLoc
An n-ary value constructor.

Update (ExpBase f vn) (SliceBase f vn) (ExpBase f vn) SrcLoc

RecordUpdate (ExpBase f vn) [Name] (ExpBase f vn) (f PatType) SrcLoc

Lambda [PatBase f vn] (ExpBase f vn) (Maybe (TypeExp vn)) (f (Aliasing, StructRetType)) SrcLoc

OpSection (QualName vn) (f PatType) SrcLoc
+; first two types are operands, third is result.

OpSectionLeft (QualName vn) (f PatType) (ExpBase f vn) (f (PName, StructType, Maybe VName), f (PName, StructType)) (f PatRetType, f [VName]) SrcLoc
2+; first type is operand, second is result.

OpSectionRight (QualName vn) (f PatType) (ExpBase f vn) (f (PName, StructType), f (PName, StructType, Maybe VName)) (f PatRetType) SrcLoc
+2; first type is operand, second is result.

ProjectSection [Name] (f PatType) SrcLoc
Field projection as a section: (.x.y.z).

IndexSection (SliceBase f vn) (f PatType) SrcLoc
Array indexing as a section: (.[i,j]).

Ascript (ExpBase f vn) (TypeDeclBase f vn) SrcLoc
Type ascription: e : t.

AppExp (AppExpBase f vn) (f AppRes)

data FieldBase f vn

An entry in a record literal.

Constructors

RecordFieldExplicit Name (ExpBase f vn) SrcLoc

RecordFieldImplicit vn (f PatType) SrcLoc

data CaseBase f vn

A case in a match expression.

Constructors

CasePat (PatBase f vn) (ExpBase f vn) SrcLoc

data LoopFormBase f vn

Whether the loop is a for-loop or a while-loop.

Constructors

For (IdentBase f vn) (ExpBase f vn)

ForIn (PatBase f vn) (ExpBase f vn)

While (ExpBase f vn)

data PatLit

A literal in a pattern.

Constructors

PatLitInt Integer

PatLitFloat Double

PatLitPrim PrimValue

data PatBase f vn

A pattern as used most places where variables are bound (function parameters, let expressions, etc).

Constructors

TuplePat [PatBase f vn] SrcLoc

RecordPat [(Name, PatBase f vn)] SrcLoc

PatParens (PatBase f vn) SrcLoc

Id vn (f PatType) SrcLoc

Wildcard (f PatType) SrcLoc

PatAscription (PatBase f vn) (TypeDeclBase f vn) SrcLoc

PatLit PatLit (f PatType) SrcLoc

PatConstr Name (f PatType) [PatBase f vn] SrcLoc

PatAttr (AttrInfo vn) (PatBase f vn) SrcLoc

Module language

data SpecBase f vn

A spec is a component of a module type.

Constructors

ValSpec

Fields

specName :: vn

specTypeParams :: [TypeParamBase vn]

specType :: TypeDeclBase f vn

specDoc :: Maybe DocComment

specLocation :: SrcLoc

TypeAbbrSpec (TypeBindBase f vn)

TypeSpec Liftedness vn [TypeParamBase vn] (Maybe DocComment) SrcLoc
Abstract type.

ModSpec vn (SigExpBase f vn) (Maybe DocComment) SrcLoc

IncludeSpec (SigExpBase f vn) SrcLoc

data SigExpBase f vn

A module type expression.

Constructors

SigVar (QualName vn) (f (Map VName VName)) SrcLoc

SigParens (SigExpBase f vn) SrcLoc

SigSpecs [SpecBase f vn] SrcLoc

SigWith (SigExpBase f vn) (TypeRefBase f vn) SrcLoc

SigArrow (Maybe vn) (SigExpBase f vn) (SigExpBase f vn) SrcLoc

data TypeRefBase f vn

A type refinement.

Constructors

TypeRef (QualName vn) [TypeParamBase vn] (TypeDeclBase f vn) SrcLoc

data SigBindBase f vn

Module type binding.

Constructors

SigBind

Fields

sigName :: vn

sigExp :: SigExpBase f vn

sigDoc :: Maybe DocComment

sigLoc :: SrcLoc

data ModExpBase f vn

Module expression.

Constructors

ModVar (QualName vn) SrcLoc

ModParens (ModExpBase f vn) SrcLoc

ModImport FilePath (f FilePath) SrcLoc
The contents of another file as a module.

ModDecs [DecBase f vn] SrcLoc

ModApply (ModExpBase f vn) (ModExpBase f vn) (f (Map VName VName)) (f (Map VName VName)) SrcLoc
Functor application. The first mapping is from parameter names to argument names, while the second maps names in the constructed module to the names inside the functor.

ModAscript (ModExpBase f vn) (SigExpBase f vn) (f (Map VName VName)) SrcLoc

ModLambda (ModParamBase f vn) (Maybe (SigExpBase f vn, f (Map VName VName))) (ModExpBase f vn) SrcLoc

data ModBindBase f vn

A module binding.

Constructors

ModBind

Fields

modName :: vn

modParams :: [ModParamBase f vn]

modSignature :: Maybe (SigExpBase f vn, f (Map VName VName))

modExp :: ModExpBase f vn

modDoc :: Maybe DocComment

modLocation :: SrcLoc

data ModParamBase f vn

A module parameter.

Constructors

ModParam

Fields

modParamName :: vn

modParamType :: SigExpBase f vn

modParamAbs :: f [VName]

modParamLocation :: SrcLoc

Definitions

data DocComment

Documentation strings, including source location.

Constructors

DocComment String SrcLoc

data ValBindBase f vn

Function Declarations

Constructors

ValBind

Fields

valBindEntryPoint :: Maybe (f EntryPoint)

Just if this function is an entry point. If so, it also contains the externally visible interface. Note that this may not strictly be well-typed after some desugaring operations, as it may refer to abstract types that are no longer in scope.

valBindName :: vn

valBindRetDecl :: Maybe (TypeExp vn)

valBindRetType :: f StructRetType

If valBindParams is null, then the retDims are brought into scope at this point.

valBindTypeParams :: [TypeParamBase vn]

valBindParams :: [PatBase f vn]

valBindBody :: ExpBase f vn

valBindDoc :: Maybe DocComment

valBindAttrs :: [AttrInfo vn]

valBindLocation :: SrcLoc

data EntryPoint

Information about the external interface exposed by an entry point. The important thing is that that we remember the original source-language types, without desugaring them at all. The annoying thing is that we do not require type annotations on entry points, so the types can be either ascribed or inferred.

Constructors

EntryPoint

Fields

entryParams :: [EntryParam]

entryReturn :: EntryType

data EntryType

Part of the type of an entry point. Has an actual type, and maybe also an ascribed type expression.

Constructors

EntryType

Fields

entryType :: StructType

entryAscribed :: Maybe (TypeExp VName)

data EntryParam

A parameter of an entry point.

Constructors

EntryParam

Fields

entryParamName :: Name

entryParamType :: EntryType

data Liftedness

The liftedness of a type parameter. By the Ord instance, Unlifted < SizeLifted < Lifted.

Constructors

Unlifted
May only be instantiated with a zero-order type of (possibly symbolically) known size.

SizeLifted
May only be instantiated with a zero-order type, but the size can be varying.

Lifted
May be instantiated with a functional type.

data TypeBindBase f vn

Type Declarations

Constructors

TypeBind

Fields

typeAlias :: vn

typeLiftedness :: Liftedness

typeParams :: [TypeParamBase vn]

typeExp :: TypeExp vn

typeElab :: f StructRetType

typeDoc :: Maybe DocComment

typeBindLocation :: SrcLoc

data TypeParamBase vn

A type parameter.

Constructors

TypeParamDim vn SrcLoc
A type parameter that must be a size.

TypeParamType Liftedness vn SrcLoc
A type parameter that must be a type.

typeParamName :: TypeParamBase vn -> vn

The name of a type parameter.

data ProgBase f vn

The program described by a single Futhark file. May depend on other files.

Constructors

Prog

Fields

progDoc :: Maybe DocComment

progDecs :: [DecBase f vn]

data DecBase f vn

A top-level binding.

Constructors

ValDec (ValBindBase f vn)

TypeDec (TypeBindBase f vn)

SigDec (SigBindBase f vn)

ModDec (ModBindBase f vn)

OpenDec (ModExpBase f vn) SrcLoc

LocalDec (DecBase f vn) SrcLoc

ImportDec FilePath (f FilePath) SrcLoc

Miscellaneous

class (Show vn, Show (f VName), Show (f (Diet, Maybe VName)), Show (f String), Show (f [VName]), Show (f ([VName], [VName])), Show (f PatType), Show (f (PatType, [VName])), Show (f (StructType, [VName])), Show (f (StructRetType, [VName])), Show (f EntryPoint), Show (f StructType), Show (f StructRetType), Show (f PatRetType), Show (f (StructType, Maybe VName)), Show (f (PName, StructType)), Show (f (PName, StructType, Maybe VName)), Show (f (Aliasing, StructRetType)), Show (f (Map VName VName)), Show (f AppRes)) => Showable f vn

Convenience class for deriving Show instances for the AST.

data NoInfo a

No information functor. Usually used for placeholder type- or aliasing information.

Constructors

NoInfo

newtype Info a

Some information. The dual to NoInfo

Constructors

Info

Fields

unInfo :: a

data Alias

A variable that is aliased. Can be still in-scope, or have gone out of scope and be free. In the latter case, it behaves more like an equivalence class. See uniqueness-error18.fut for an example of why this is necessary.

Constructors

AliasBound

Fields

aliasVar :: VName

AliasFree

Fields

aliasVar :: VName

type Aliasing = Set Alias

Aliasing for a type, which is a set of the variables that are aliased.

data QualName vn

A name qualified with a breadcrumb of module accesses.

Constructors

QualName

Fields

qualQuals :: ![vn]

qualLeaf :: !vn