Language.Futhark.TypeChecker.Monad

Description

Main monad in which the type checker runs, as well as ancillary data definitions.

Documentation

data TypeM a

The type checker runs in this monad.

runTypeM :: Env -> ImportTable -> ImportName -> VNameSource -> TypeM a -> (Warnings, Either TypeError (a, VNameSource))

Run a TypeM computation.

askEnv :: TypeM Env

Retrieve the current Env.

askImportName :: TypeM ImportName

The name of the current file/import.

atTopLevel :: TypeM Bool

Are we type-checking at the top level, or are we inside a nested module?

enteringModule :: TypeM a -> TypeM a

We are now going to type-check the body of a module.

bindSpaced :: MonadTypeChecker m => [(Namespace, Name)] -> m a -> m a

Map source-level names do fresh unique internal names, and evaluate a type checker context with the mapping active.

qualifyTypeVars :: Env -> [VName] -> [VName] -> TypeBase (DimDecl VName) as -> TypeBase (DimDecl VName) as

Try to prepend qualifiers to the type names such that they represent how to access the type in some scope.

lookupMTy :: SrcLoc -> QualName Name -> TypeM (QualName VName, MTy)

Look up a module type.

lookupImport :: SrcLoc -> FilePath -> TypeM (FilePath, Env)

Look up an import.

localEnv :: Env -> TypeM a -> TypeM a

Evaluate a TypeM computation within an extended (not replaced) environment.

data TypeError

Information about an error during type checking.

Constructors

TypeError SrcLoc Notes Doc

withIndexLink :: Doc -> Doc -> Doc

Attach a reference to documentation explaining the error in more detail.

unappliedFunctor :: MonadTypeChecker m => SrcLoc -> m a

An unexpected functor appeared!

unknownVariable :: MonadTypeChecker m => Namespace -> QualName Name -> SrcLoc -> m a

An unknown variable was referenced.

unknownType :: MonadTypeChecker m => SrcLoc -> QualName Name -> m a

An unknown type was referenced.

underscoreUse :: MonadTypeChecker m => SrcLoc -> QualName Name -> m a

A name prefixed with an underscore was used.

data Notes

A collection of Notes.

aNote :: Pretty a => a -> Notes

A single note.

class Monad m => MonadTypeChecker m where

Monads that support type checking. The reason we have this internal interface is because we use distinct monads for checking expressions and declarations.

Minimal complete definition

warn, newName, newID, newTypeName, bindNameMap, bindVal, checkQualName, lookupType, lookupMod, lookupVar, typeError

Methods

warn :: Located loc => loc -> Doc -> m ()

newName :: VName -> m VName

newID :: Name -> m VName

newTypeName :: Name -> m VName

bindNameMap :: NameMap -> m a -> m a

bindVal :: VName -> BoundV -> m a -> m a

checkQualName :: Namespace -> QualName Name -> SrcLoc -> m (QualName VName)

lookupType :: SrcLoc -> QualName Name -> m (QualName VName, [TypeParam], StructRetType, Liftedness)

lookupMod :: SrcLoc -> QualName Name -> m (QualName VName, Mod)

lookupVar :: SrcLoc -> QualName Name -> m (QualName VName, PatType)

checkNamedDim :: SrcLoc -> QualName Name -> m (QualName VName)

typeError :: Located loc => loc -> Notes -> Doc -> m a

checkName :: MonadTypeChecker m => Namespace -> Name -> SrcLoc -> m VName

Elaborate the given name in the given namespace at the given location, producing the corresponding unique VName.

checkAttr :: MonadTypeChecker m => AttrInfo Name -> m (AttrInfo VName)

Type-check an attribute.

badOnLeft :: Either TypeError a -> TypeM a

Turn a Left TypeError into an actual error.

module Language.Futhark.Warnings

data Env

Modules produces environment with this representation.

Constructors

Env

Fields

envVtable :: Map VName BoundV

envTypeTable :: Map VName TypeBinding

envSigTable :: Map VName MTy

envModTable :: Map VName Mod

envNameMap :: NameMap

type TySet = Map (QualName VName) Liftedness

A mapping of abstract types to their liftedness.

data FunSig

A parametric functor consists of a set of abstract types, the environment of its parameter, and the resulting module type.

Constructors

FunSig

Fields

funSigAbs :: TySet

funSigMod :: Mod

funSigMty :: MTy

type ImportTable = Map String Env

A mapping from import strings to Envs. This is used to resolve import declarations.

type NameMap = Map (Namespace, Name) (QualName VName)

A mapping from names (which always exist in some namespace) to a unique (tagged) name.

data BoundV

Type parameters, list of parameter types (optinally named), and return type. The type parameters are in scope in both parameter types and the return type. Non-functional values have only a return type.

Constructors

BoundV [TypeParam] StructType

data Mod

Representation of a module, which is either a plain environment, or a parametric module ("functor" in SML).

Constructors

ModEnv Env

ModFun FunSig

data TypeBinding

A binding from a name to its definition as a type. We allow a return type here to support type abbreviations that hide some inner sizes (these must necessarily be Lifted or SizeLifted).

Constructors

TypeAbbr Liftedness [TypeParam] StructRetType

data MTy

Representation of a module type.

Constructors

MTy

Fields

mtyAbs :: TySet

Abstract types in the module type.

mtyMod :: Mod

anySignedType :: [PrimType]

All signed integer types.

anyUnsignedType :: [PrimType]

All unsigned integer types.

anyIntType :: [PrimType]

All integer types.

anyFloatType :: [PrimType]

All floating-point types.

anyNumberType :: [PrimType]

All number types.

anyPrimType :: [PrimType]

All primitive types.

data Namespace

The space inhabited by a name.

Constructors

Term
Functions and values.

Type

Signature

intrinsicsNameMap :: NameMap

The NameMap corresponding to the intrinsics module.

topLevelNameMap :: NameMap

The names that are available in the initial environment.

mkTypeVarName :: Name -> Int -> Name

Construct the name of a new type variable given a base description and a tag number (note that this is distinct from actually constructing a VName; the tag here is intended for human consumption but the machine does not care).