Language.Futhark.TypeChecker.Terms.DoLoop

Description

Type inference of loop. This is complicated because of the uniqueness and size inference, so the implementation is separate from the main type checker.

Documentation

type UncheckedLoop = (UncheckedPat, UncheckedExp, LoopFormBase NoInfo Name, UncheckedExp)

An un-checked loop.

type CheckedLoop = ([VName], Pat, Exp, LoopFormBase Info VName, Exp)

A loop that has been type-checked.

checkDoLoop :: (UncheckedExp -> TermTypeM Exp) -> UncheckedLoop -> SrcLoc -> TermTypeM (CheckedLoop, AppRes)

Type-check a loop expression, passing in a function for type-checking subexpressions.