Language.Futhark.Warnings

Description

A very simple representation of collections of warnings. Warnings have a position (so they can be ordered), and their Show-instance produces a human-readable string.

Documentation

data Warnings

The warnings produced by the compiler. The Show instance produces a human-readable description.

anyWarnings :: Warnings -> Bool

True if there are any warnings in the set.

singleWarning :: SrcLoc -> Doc -> Warnings

A single warning at the given location.

singleWarning' :: SrcLoc -> [SrcLoc] -> Doc -> Warnings

A single warning at the given location, but also with a stack trace (sort of) to the location.