Futhark.IR.SOACS.SOAC

Description

Definition of Second-Order Array Combinators (SOACs), which are the main form of parallelism in the early stages of the compiler.

Documentation

data SOAC rep

A second-order array combinator (SOAC).

Constructors

Stream SubExp [VName] (StreamForm rep) [SubExp] (Lambda rep)

Scatter SubExp [VName] (Lambda rep) [(Shape, Int, VName)]
Scatter length lambda inputs outputs

Scatter maps values from a set of input arrays to indices and values of a set of output arrays. It is able to write multiple values to multiple outputs each of which may have multiple dimensions.

inputs is a list of input arrays, all having size length, elements of which are applied to the lambda function. For instance, if there are two arrays, lambda will get two values as input, one from each array.

outputs specifies the result of the lambda and which arrays to write to. Each element of the list consists of a VName specifying which array to scatter to, a Shape describing the shape of that array, and an Int describing how many elements should be written to that array for each invocation of the lambda.

lambda is a function that takes inputs from inputs and returns values according to the output-specification in outputs. It returns values in the following manner:

index_0, index_1, ..., index_n, value_0, value_1, ..., value_m

For each output in outputs, lambda returns i * j index values and j output values, where i is the number of dimensions (rank) of the given output, and j is the number of output values written to the given output.

For example, given the following output specification:

([x1, y1, z1

, 2, arr1), ([x2, y2], 1, arr2)]

lambda will produce 6 (3 * 2) index values and 2 output values for arr1, and 2 (2 * 1) index values and 1 output value for arr2. Additionally, the results are grouped, so the first 6 index values will correspond to the first two output values, and so on. For this example, lambda should return a total of 11 values, 8 index values and 3 output values.

Hist SubExp [VName] [HistOp rep] (Lambda rep)
Hist length dest-arrays-and-ops fun arrays

The first SubExp is the length of the input arrays. The first list describes the operations to perform. The Lambda is the bucket function. Finally comes the input images.

JVP (Lambda rep) [SubExp] [SubExp]

VJP (Lambda rep) [SubExp] [SubExp]

Screma SubExp [VName] (ScremaForm rep)
A combination of scan, reduction, and map. The first SubExp is the size of the input arrays.

data StreamOrd

Is the stream chunk required to correspond to a contiguous subsequence of the original input (InOrder) or not? Disorder streams can be more efficient, but not all algorithms work with this.

Constructors

InOrder

Disorder

data StreamForm rep

What kind of stream is this?

Constructors

Parallel StreamOrd Commutativity (Lambda rep)

Sequential

data ScremaForm rep

The essential parts of a Screma factored out (everything except the input arrays).

Constructors

ScremaForm [Scan rep] [Reduce rep] (Lambda rep)

data HistOp rep

Information about computing a single histogram.

Constructors

HistOp

Fields

histShape :: Shape

histRaceFactor :: SubExp

Race factor RF means that only 1/RF bins are used.

histDest :: [VName]

histNeutral :: [SubExp]

histOp :: Lambda rep

data Scan rep

How to compute a single scan result.

Constructors

Scan

Fields

scanLambda :: Lambda rep

scanNeutral :: [SubExp]

scanResults :: [Scan rep] -> Int

How many reduction results are produced by these Scans?

singleScan :: Buildable rep => [Scan rep] -> Scan rep

Combine multiple scan operators to a single operator.

data Reduce rep

How to compute a single reduction result.

Constructors

Reduce

Fields

redComm :: Commutativity

redLambda :: Lambda rep

redNeutral :: [SubExp]

redResults :: [Reduce rep] -> Int

How many reduction results are produced by these Reduces?

singleReduce :: Buildable rep => [Reduce rep] -> Reduce rep

Combine multiple reduction operators to a single operator.

Utility

scremaType :: SubExp -> ScremaForm rep -> [Type]

The types produced by a single Screma, given the size of the input array.

soacType :: Typed (LParamInfo rep) => SOAC rep -> [Type]

The type of a SOAC.

typeCheckSOAC :: Checkable rep => SOAC (Aliases rep) -> TypeM rep ()

Type-check a SOAC.

mkIdentityLambda :: (Buildable rep, MonadFreshNames m) => [Type] -> m (Lambda rep)

Construct a lambda that takes parameters of the given types and simply returns them unchanged.

isIdentityLambda :: Lambda rep -> Bool

Is the given lambda an identity lambda?

nilFn :: Buildable rep => Lambda rep

A lambda with no parameters that returns no values.

scanomapSOAC :: [Scan rep] -> Lambda rep -> ScremaForm rep

Construct a Screma with possibly multiple scans, and the given map function.

redomapSOAC :: [Reduce rep] -> Lambda rep -> ScremaForm rep

Construct a Screma with possibly multiple reductions, and the given map function.

scanSOAC :: (Buildable rep, MonadFreshNames m) => [Scan rep] -> m (ScremaForm rep)

Construct a Screma with possibly multiple scans, and identity map function.

reduceSOAC :: (Buildable rep, MonadFreshNames m) => [Reduce rep] -> m (ScremaForm rep)

Construct a Screma with possibly multiple reductions, and identity map function.

mapSOAC :: Lambda rep -> ScremaForm rep

Construct a Screma corresponding to a map.

isScanomapSOAC :: ScremaForm rep -> Maybe ([Scan rep], Lambda rep)

Does this Screma correspond to a scan-map composition?

isRedomapSOAC :: ScremaForm rep -> Maybe ([Reduce rep], Lambda rep)

Does this Screma correspond to a reduce-map composition?

isScanSOAC :: ScremaForm rep -> Maybe [Scan rep]

Does this Screma correspond to pure scan?

isReduceSOAC :: ScremaForm rep -> Maybe [Reduce rep]

Does this Screma correspond to a pure reduce?

isMapSOAC :: ScremaForm rep -> Maybe (Lambda rep)

Does this Screma correspond to a simple map, without any reduction or scan results?

scremaLambda :: ScremaForm rep -> Lambda rep

Return the "main" lambda of the Screma. For a map, this is equivalent to isMapSOAC. Note that the meaning of the return value of this lambda depends crucially on exactly which Screma this is. The parameters will correspond exactly to elements of the input arrays, however.

ppScrema :: (PrettyRep rep, Pretty inp) => SubExp -> [inp] -> ScremaForm rep -> Doc

Prettyprint the given Screma.

ppHist :: (PrettyRep rep, Pretty inp) => SubExp -> [inp] -> [HistOp rep] -> Lambda rep -> Doc

Prettyprint the given histogram operation.

groupScatterResults :: [(Shape, Int, array)] -> [a] -> [(Shape, array, [([a], a)])]

groupScatterResults specification results

Groups the index values and result values of results according to the specification.

This function is used for extracting and grouping the results of a scatter. In the SOAC representation, the lambda inside a Scatter returns all indices and values as one big list. This function groups each value with its corresponding indices (as determined by the Shape of the output array).

The elements of the resulting list correspond to the shape and name of the output parameters, in addition to a list of values written to that output parameter, along with the array indices marking where to write them to.

See Scatter for more information.

groupScatterResults' :: [(Shape, Int, array)] -> [a] -> [([a], a)]

groupScatterResults' specification results

Groups the index values and result values of results according to the output specification. This is the simpler version of groupScatterResults, which doesn't return any information about shapes or output arrays.

See groupScatterResults for more information,

splitScatterResults :: [(Shape, Int, array)] -> [a] -> ([a], [a])

splitScatterResults specification results

Splits the results array into indices and values according to the output specification.

See groupScatterResults for more information.

Generic traversal

data SOACMapper frep trep m

Like Mapper, but just for SOACs.

Constructors

SOACMapper

Fields

mapOnSOACSubExp :: SubExp -> m SubExp

mapOnSOACLambda :: Lambda frep -> m (Lambda trep)

mapOnSOACVName :: VName -> m VName

identitySOACMapper :: Monad m => SOACMapper rep rep m

A mapper that simply returns the SOAC verbatim.

mapSOACM :: (Applicative m, Monad m) => SOACMapper frep trep m -> SOAC frep -> m (SOAC trep)

Map a monadic action across the immediate children of a SOAC. The mapping does not descend recursively into subexpressions and is done left-to-right.

traverseSOACStms :: Monad m => OpStmsTraverser m (SOAC rep) rep

A helper for defining TraverseOpStms.